% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliVis.R
\name{add_plot_features}
\alias{add_plot_features}
\title{Adds features to dataframe for plotting}
\usage{
add_plot_features(...)
}
\arguments{
\item{trellData}{An object of class "trellData" generated from \code{\link{format_data}}.}

\item{nestedData}{Nested trellData dataframe}

\item{p_val}{Specifies p-value for setting graph border colors}

\item{panel_variable}{Specifies what to divide trelliscope panels by, must be a column in trellData. Defaults to cnames$edata_cname of trellData.}

\item{comps_panel_y_axis}{Specifies what column should be on the y-axis, must be a column in trellData and numeric. Defaults setting plots fold change for combined and anova testing and counts for g-test.}

\item{value_panel_y_axis}{Specifies what column should be on the y-axis, must be a column in trellData and numeric. Defaults setting plots fold change for combined and anova testing and counts for g-test.}

\item{colors}{Specifies colors to use for border.}
}
\description{
Adds border color, hover, and label text as applicable to nested trellData dataframe for plotting
}
\author{
Rachel Richardson
}
