% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_log.R
\name{analysis_log}
\alias{analysis_log}
\title{Creates R markdown document report}
\usage{
analysis_log(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData',
'lipidData', 'nmrData', or 'seqData', created by
\code{\link{as.pepData}}, \code{\link{as.proData}},
\code{\link{as.metabData}}, \code{\link{as.lipidData}},
\code{\link{as.nmrData}}, or \code{\link{as.seqData}}, respectively.}
}
\description{
Creates R markdown document report
}
\examples{
dontrun{
library(pmartRdata)
data(omicsData = metab_object)
result <- analysis_log(omicsData = metab_object)
}

}
