% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyFilt.R
\name{applyFilt}
\alias{applyFilt}
\alias{applyFilt.customFilt}
\alias{applyFilt.cvFilt}
\alias{applyFilt.imdanovaFilt}
\alias{applyFilt.moleculeFilt}
\alias{applyFilt.proteomicsFilt}
\alias{applyFilt.rmdFilt}
\title{Apply a S3 filter  object to an MSomics S3 object}
\usage{
applyFilt(filter_object, omicsData, ...)

\method{applyFilt}{moleculeFilt}(filter_object, omicsData, min_num = 2)

\method{applyFilt}{cvFilt}(filter_object, omicsData, cv_threshold = 150)

\method{applyFilt}{rmdFilt}(filter_object, omicsData,
  pvalue_threshold = 0.001)

\method{applyFilt}{proteomicsFilt}(filter_object, omicsData,
  min_num_peps = NULL, degen_peps = FALSE)

\method{applyFilt}{imdanovaFilt}(filter_object, omicsData,
  min_nonmiss_anova = NULL, min_nonmiss_gtest = NULL)

\method{applyFilt}{customFilt}(filter_object, omicsData)
}
\arguments{
\item{filter_object}{an object of the class 'cvFilt', 'proteomicsFilt', 'rmdFilt', 'moleculeFilt', 'imdanovaFilt', or 'customFilt' created by \code{cv_filter}, \code{proteomics_filter}, \code{rmd_filter}, \code{molecule_filter}, or \code{imdanova_filter}, respectively.}

\item{omicsData}{an object of the class \code{pepData}, \code{proData}, \code{lipidData}, or \code{metabData} usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.lipidData}}, or  \code{\link{as.metabData}}, respectively.}

\item{...}{further arguments}
}
\value{
An object of the class \code{pepData}, \code{proData}, \code{lipidData}, or \code{metabData},  with specified cname_ids, edata_cnames, and emeta_cnames filtered out of the appropriate datasets.
}
\description{
This function takes a filter object of class 'cvFilt', 'rmdFilt', 'moleculeFilt', 'proteomicsFilt', 'imdanovaFilt', or 'customFilt' and applies the filter to a dataset of class \code{pepData}, \code{proData}, \code{lipidData}, or \code{metabData}.
}
\details{
Various further arguments can be specified depending on the class of the \code{filter_object} being applied.
For a \code{filter_object} of type 'moleculeFilt':
\tabular{ll}{
\code{min_num} \tab an integer value specifying the minimum number of times each biomolecule must be observed across all samples in order to retain the biomolecule. Default value is 2. \cr
}
For a \code{filter_object} of type 'cvFilt':
\tabular{ll}{
\code{cv_threshold} \tab an integer value specifying the maximum coefficient of variation (CV) threshold for the biomolecules. Biomolecules with CV greater than this threshold will be filtered. Default is 150. \cr
}
For a \code{filter_object} of type 'rmdFilt':
\tabular{ll}{
\code{pvalue_threshold} \tab numeric value between 0 and 1, specifying the p-value, below which samples will be removed from the dataset. Default is 0.001. \cr
}
For a \code{filter_object} of type 'proteomicsFilt':
\tabular{ll}{
\code{min_num_peps} \tab an optional integer value between 1 and the maximum number of peptides that map to a protein in omicsData. The value specifies the minimum number of peptides that must map to a protein. Any protein with less than \code{min_num_peps} mapping to it will be returned as a protein that should be filtered. Default value is NULL. \cr
\code{degen_peps} \tab logical indicator of whether to filter out degenerate peptides (TRUE) or not (FALSE). Default value is FALSE.\cr
}
For a \code{filter_object} of type 'imdanovaFilt':
\tabular{ll}{
\code{min_nonmiss_anova} \tab integer value specifying the minimum number of non-missing feature values allowed per group for \code{anova_filter}. Default value is 2. \cr
\code{min_nonmiss_gtest} \tab integer value specifying the minimum number of non-missing feature values allowed per group for \code{gtest_filter}. Default value is 3.\cr
}
There are no futher arguments for a \code{filter_object} of type 'customFilt'.
}
\examples{
dontrun{
library(pmartRdata)
data("pep_object")
to_filter <- molecule_filter(omicsData = pep_object)
pep_object2 <- apply(filter_object = to_filter, omicsData = pep_object, min_num = 2)
print(str(attributes(pep_object2)$filters))
to_filter2 <- imdanova_filter(omicsData = pep_object2)
pep_object3 <- apply(filter_object = to_filter2, omicsData = pep_object2, min_num_peps = 4)
print(str(attributes(pep_object3)$filters))
}

}
\author{
Lisa Bramer, Kelly Stratton
}
\seealso{
\code{\link{molecule_filter}}

\code{\link{imdanova_filter}}

\code{\link{rmd_filter}}

\code{\link{cv_filter}}

\code{\link{proteomics_filter}}

\code{\link{custom_filter}}
}

