% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.metabData.R
\name{as.metabData}
\alias{as.metabData}
\title{Convert Data to Appropriate pmartR Class}
\usage{
as.metabData(e_data, f_data, e_meta = NULL, ...)
}
\arguments{
\item{e_data}{a \eqn{p \times n + 1} data.frame of expression data, where \eqn{p} is the number of metabolites observed and \eqn{n} is the number of samples. Each row corresponds to data for each metabolite. One column specifying a unique identifier for each metabolite (row) must be present.}

\item{f_data}{a data.frame with \eqn{n} rows. Each row corresponds to a sample with one column giving the unique sample identifiers found in e_data column names and other columns providing qualitative and/or quantitative traits of each sample.}

\item{e_meta}{an optional data.frame with \eqn{p} rows. Each row corresponds to a metabolite with one column giving metabolite names (must be named the same as the column in \code{e_data}) and other columns giving meta information.}

\item{...}{further arguments}
}
\description{
Converts a list object or several data.frames of metabolomic-level data to an object of the class 'metabData'. Objects of the class 'metabData' are lists with two obligatory components \code{e_data} and \code{f_data}. An optional list component \code{e_meta} is used if analysis or visualization at other levels (e.g. metabolite identification) is also desired.
}
\details{
Objects of class 'metabData' contain some attributes that are referenced by downstream functions. These attributes can be changed from their default value by manual specification. A list of these attributes as well as their default values are as follows:
\tabular{ll}{
edata_cname  \tab Column name for metabolite information found in \code{e_data} and \code{e_meta} (if applicable). Default value is 'Metabolite'. \cr
 \tab \cr
emeta_cname  \tab Column name for metabolite information found in \code{e_meta} (if applicable). Default value is NULL. If \code{e_meta} is NULL, then \code{emeta_cname} will default to a NULL value. \cr
 \tab \cr
samp_cname \tab Column name for sample information found in \code{f_data}. Default value is 'sampleID'. \cr
\tab \cr
data_scale \tab Scale of the data provided in \code{e_data}. Acceptable values are 'log2', 'log10', 'log', and 'abundance', which indicate data is log base 2, base 10, natural log transformed, and raw abundance, respectively. \cr
\tab \cr
data_norm \tab A logical argument, specifying whether the data has been normalized or not. Default value is FALSE. \cr
\tab \cr
norm_method \tab Character string defining which normalization method was used. If \code{data_norm} is FALSE, this is set to NULL. Default value is NULL. \cr
\tab \cr
location_param \tab A vector of normalization location parameters for each sample. This is set to NULL if there are no location parameters from normalization, or \code{data_norm} is FALSE. Default value is NULL. \cr
\tab \cr
scale_param \tab A vector of normalization scale parameters for each sample. This is set to NULL if there are no scale parameters from normalization, or \code{data_norm} is FALSE. Default value is NULL.\cr
\tab \cr
data_types \tab Character string describing the type of data (e.g.'Positive ion'). Default value is NULL. \cr
}
Computed values included in the \code{data_info} attribute are as follows:
\tabular{ll}{
num_edata \tab The number of unique \code{edata_cname} entries.\cr
\tab \cr
num_miss_obs \tab The number of missing observations.\cr
\tab \cr
num_emeta \tab The number of unique \code{emeta_cname} entries. \cr
\tab \cr
prop_missing \tab The proportion of \code{e_data} values that are NA. \cr
\tab \cr
num_samps \tab The number of samples that make up the columns of \code{e_data}.\cr
\tab \cr
meta_info \tab A logical argument, specifying where the \code{e_meta} is provided.\cr
\tab \cr
}
}
\examples{
library(MSomicsDATA)
data("metab_edata")
data("metab_fdata")
mymetabData <- as.metabData(e_data = metab_edata, f_data = metab_fdata, edata_cname = "Metabolite", samp_cname = "SampleID")


}
\author{
Lisa Bramer, Kelly Stratton
}
\seealso{
\code{\link{as.lipidData}}

\code{\link{as.pepData}}

\code{\link{as.proData}}
}

