% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.omicsPlotter.R
\name{as.omicsPlotter}
\alias{as.omicsPlotter}
\title{Convert Omics data and pairwise statistics to a plotting object}
\usage{
as.omicsPlotter(omicsData, omicsStats, ...)
}
\arguments{
\item{omicsData}{A pmartR object of class pepData, lipidData, metabData, or proData}

\item{omicsStats}{A statistical results object produced by running \code{\link{imd_anova}} on omicsData.}

\item{...}{further arguments}
}
\description{
Converts a ResObject and its respective OmicsData into an easily plottable object.
}
\details{
Objects of class 'omicsPlotter' inherit attributes from omicsStats and omicsData. These attributes can be changed from their default value by manual specification. A list of these attributes as well as their default values are as follows:
\tabular{ll}{
data_scale \tab Scale of the data provided in \code{e_data}. Acceptable values are 'log2', 'log10', 'log', and 'abundance', which indicate data is log base 2, base 10, natural log transformed, and raw abundance, respectively. Default values is 'abundance'. \cr
\tab \cr
is_normalized \tab A logical argument, specifying whether the data has been normalized or not. Default value is FALSE. \cr
\tab \cr
norm_info \tab Default value is an empty list, which will be populated with a single named element \code{is_normalized = is_normalized}. When a normalization is applied to the data, this becomes populated with a list containing the normalization function, normalization subset and subset parameters, the location and scale parameters used to normalize the data, and the location and scale parameters used to backtransform the data (if applicable). \cr
\tab \cr
data_types \tab Character string describing the type of data (e.g.'Positive ion'). Default value is NULL. \cr
\tab\cr
check.names \tab Logical defaults to TRUE. Indicates whether 'check.names' attribute of returned omicsData object is TRUE or FALSE. \cr
}
Computed values included in the \code{data_info} attribute are as follows:
\tabular{ll}{
num_edata \tab The number of unique \code{edata_cname} entries.\cr
\tab \cr
num_miss_obs \tab The number of missing observations.\cr
\tab \cr
num_emeta \tab The number of unique \code{emeta_cname} entries. \cr
\tab \cr
prop_missing \tab The proportion of \code{e_data} values that are NA. \cr
\tab \cr
num_samps \tab The number of samples that make up the columns of \code{e_data}.\cr
\tab \cr
meta_info \tab A logical argument, specifying whether \code{e_meta} is provided.\cr
\tab \cr
}
}
\examples{
dontrun{
library(pmartRdata)
data("pro_edata")
data("pro_fdata")
myproData <- as.proData(e_data = pro_edata, f_data = pro_fdata, edata_cname = "Reference", fdata_cname = "SampleID", is_normalized=TRUE)
}

}
\author{
Rachel Richardson
}
