% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{column_matches_exact}
\alias{column_matches_exact}
\title{Helper to find the names of columns of a data.frame that contain exactly all
the elements of an input column}
\usage{
column_matches_exact(df, col)
}
\arguments{
\item{df}{A data.frame whose columns we want to match to some query column.}

\item{col}{Vector of values which will be compared to a column in df.}
}
\value{
vector of column names of df that contain exactly all the elements of
the input column
}
\description{
Helper to find the names of columns of a data.frame that contain exactly all
the elements of an input column
}
\keyword{internal}
