% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_omicsdata.R
\name{combine_lipidData}
\alias{combine_lipidData}
\title{Combines two omicsData objects with identical sample information.}
\usage{
combine_lipidData(obj_1, obj_2, retain_groups = FALSE, retain_filters = FALSE)
}
\arguments{
\item{obj_1, obj_2}{Two omicsData objects of the same supported type,
currently "lipidData".  See details for more requirements.}

\item{retain_groups}{Whether to attempt to apply existing group information
to the new object (defaults to FALSE).}

\item{retain_filters}{Whether to retain filter information in the new object
(defaults to FALSE).}
}
\value{
An object of the same type as the two input objects, with their
combined data.
}
\description{
Combines two omicsData objects with identical sample information.
}
\details{
General requirements:
\itemize{
\item sample names:  These must be identical for both objects (column names of e_data, and sample identifiers in f_data).
\item data attributes:  Objects must be on the same scale and both be either normalized or unnormalized.
\item group designation:  Objects must have the same grouping structure if retain_groups = T.
}
}
