% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliVis.R
\name{custom_fn}
\alias{custom_fn}
\title{Test function for format_plot(), custom_plot argument}
\usage{
custom_fn(data)
}
\arguments{
\item{data}{data subset from a trellData object.}
}
\description{
Test function for format_plot(), custom_plot argument. Plots a ggplot where x = 1:10, y = 1:10.
}
\author{
Rachel Richardson
}
