% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliVis.R
\name{custom_fn_cog}
\alias{custom_fn_cog}
\title{Test function for data_cog(), custom_cog argument}
\usage{
custom_fn_cog(data)
}
\arguments{
\item{data}{data subset from a trellData object.}
}
\description{
Test function for data_cog(), custom_cog argument. Returns a tibble with one row and 3 columns.
}
\author{
Rachel Richardson
}
