% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_filter.R
\name{cv.calc}
\alias{cv.calc}
\title{Calculate the Coefficient of Variation (CV)}
\usage{
cv.calc(data)
}
\arguments{
\item{data}{a vector of values}
}
\value{
cv value for the vector of values with missing values ignored in the calculation
}
\description{
This function calculates the coefficient of variation for a vector of data
}
\author{
Lisa Bramer
}
