% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliVis.R
\name{data_cogs}
\alias{data_cogs}
\title{Plot pairwise comparisons and data values in trellData object}
\usage{
data_cogs(nested_plot, trellData, p_val = 0.05, try_URL = FALSE,
  custom_cog = NULL)
}
\arguments{
\item{nested_plot}{A nested table generated from trellData using formatplot()}

\item{trellData}{A nested table generated from trellData using formatplot()}

\item{p_val}{Numeric that specifies p-value for Boolean significance cognotic. Default is 0.05.}

\item{try_URL}{Will attempt to link to PubChem, LipidMaps, or Uniprot based on information in edata_cname of omicsData or specified mapping_col for peptide data. Default is FALSE.}

\item{custom_cog}{The name of a user generated function with cognostics. Function should act on a subset of data and output a dataframe (or tibble or equivelent) with one row (summary of rows).}
}
\description{
Plot pairwise comparisons and data values in trellData object. Customizable for plot types, y axis limits, paneling variable (what overall group is plotted on each graph arrangement), as well as desired variables for the y and x axis.
}
\author{
Rachel Richardson
}
