% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqData_wrappers.R
\name{diffexp_seq}
\alias{diffexp_seq}
\title{Wrapper for Differential Expression workflows}
\usage{
diffexp_seq(
  omicsData,
  method = "EdgeR",
  p_adjust = "BH",
  comparisons = NULL,
  p_cutoff = 0.05,
  ...
)
}
\arguments{
\item{omicsData}{an object of type 'seqData', created by \code{\link{as.seqData}}}

\item{method}{a character string of length one specifying which wrapper to use. 
Can be 'edgeR', 'DESeq2', or 'limma-voom'}

\item{p_adjust}{an object of type 'seqData', created by \code{\link{as.seqData}}}

\item{comparisons}{`data.frame` with columns for "Control" and "Test"
containing the different comparisons of interest. Comparisons will be made
between the Test and the corresponding Control  If left NULL, then all
pairwise comparisons are executed.}

\item{p_cutoff}{an object of type 'seqData', created by \code{\link{as.seqData}}}

\item{...}{additional arguments passed to methods functions. Note, formatting 
option changes will interfere with wrapping functionality.}
}
\value{
data.frame object
}
\description{
For generating statistics for 'seqData' objects
}
\details{
Runs default differential expression workflows. Refer to ?Deseq2_wrapper,
?Voom_wrapper, and ?EdgeR_wrapper for additional details.
}
\examples{
\dontrun{
}

}
