% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{dim_reduction}
\alias{dim_reduction}
\title{Reduce Dimension of Data for Exploratory Data Analysis}
\usage{
dim_reduction(omicsData, k = 2)
}
\arguments{
\item{omicsData}{an object of the class 'pepdata', 'prodata', 'metabData',
'lipidData', 'nmrData', or 'seqData', 
usually created by \code{\link{as.pepData}},
\code{\link{as.proData}}, \code{\link{as.metabData}},
\code{\link{as.lipidData}}, \code{\link{as.nmrData}}, or
\code{\link{as.seqData}}, respectively.}

\item{k}{integer number of principal components to return. Defaults to 2.}
}
\value{
a data.frame with first \code{k} principal component scores, sample
  identifiers, and group membership for each sample (if group designation was
  previously run on the data).
}
\description{
Calculate principal components using projection pursuit estimation, which
implements an expectation-maximization (EM) estimation algorithm when data is
missing. For seqData counts, a generalized version of principal components 
analysis for non-normally distributed data is calculated under the assumption 
of a negative binomial distribution with global dispersion.
}
\details{
Any biomolecules seen in only one sample or with a variance less than
  1E-6 across all samples are not included in the PCA calculations. This
  function leverages code from \code{\link[pcaMethods]{pca}} and 
  \code{\link[glmpca]{glmpca}} .
}
\examples{
\dontrun{
library(pmartRdata)
data(lipid_object)
lipid_object <- edata_transform(omicsData = lipid_object,
                                data_scale="log2")
lipid_object <- group_designation(omicsData = lipid_object,
                                  main_effects = "Condition")
pca_lipids <- dim_reduction(omicsData = lipid_object)
plot(pca_lipids)
summary(pca_lipids)
}

}
\references{
Redestig, H., Stacklies, W., Scholz, M., Selbig, J., & Walther,
  D. (2007). \emph{pcaMethods - a bioconductor package providing PCA methods
  for incomplete data}. Bioinformatics. 23(9): 1164-7.
  
  Implements a generalized version of principal components analysis (GLM-PCA) 
  for dimension reduction of non-normally distributed data such as counts or 
  binary matrices. Townes FW, Hicks SC, Aryee MJ, Irizarry RA (2019) 
  <doi:10.1186/s13059-019-1861-6>. Townes FW (2019) <arXiv:1907.02647>.
}
