% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.trelliData.R
\name{.is_edata}
\alias{.is_edata}
\title{Test if a file is an edata file}
\usage{
.is_edata(edata)
}
\arguments{
\item{edata}{Must be a dataframe. Required.}
}
\value{
A boolean where TRUE means the file is an acceptable edata file.
}
\description{
Test if a file is an edata file
}
\examples{
\dontshow{if (requireNamespace("pmartRdata", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{

.is_edata(pmartRdata::lipid_edata)
}
\dontshow{\}) # examplesIf}
}
