% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edata_transform.R
\name{edata_transform}
\alias{edata_transform}
\title{Apply a Transformation to the Data}
\usage{
edata_transform(omicsData, data_scale)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData', 'lipidData', or 'nmrData' usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, \code{\link{as.lipidData}}, \code{\link{as.nmrData}}, respectively.}

\item{data_scale}{a character string indicating the type of transformation to be applied to the data. Valid values are: 'log2', 'log', 'log10', or 'abundance'. A value of 'abundance' indicates the data has previously undergone one of the log transformations and should be transformed back to raw values with no transformation applied.}
}
\value{
data object of the same class as omicsData
}
\description{
This function applies a transformation to the e_data element of omicsData
}
\details{
This function is intended to be used before analysis of the data begins. Data are typically analyzed on a log scale.
}
\examples{
dontrun{
library(pmartRdata)
data(metab_object)
metab_object2 <- edata_transform(omicsData = metab_object, data_scale="log2")
attr(metab_object2, "data_info")$data_scale
}
}
\author{
Kelly Stratton, Natalie Heller
}
