% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqData_wrappers.R
\name{EdgeR_wrapper}
\alias{EdgeR_wrapper}
\title{Wrapper for EdgeR workflow}
\usage{
EdgeR_wrapper(omicsData, p_adjust = "BH", comparisons = NULL, p_cutoff = 0.05)
}
\arguments{
\item{omicsData}{an object of type 'seqData', created by \code{\link{as.seqData}}}

\item{p_adjust}{an object of type 'seqData', created by \code{\link{as.seqData}}}

\item{comparisons}{an object of type 'seqData', created by \code{\link{as.seqData}}}

\item{p_cutoff}{an object of type 'seqData', created by \code{\link{as.seqData}}}
}
\value{
data.frame object
}
\description{
For generating statistics for 'seqData' objects
}
\examples{
\dontrun{
}

}
