% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.multiData.R
\name{fmeta_matches}
\alias{fmeta_matches}
\title{Check that the f_meta file contains a column aligned to each omicsData objects}
\usage{
fmeta_matches(omicsData_objects, f_meta)
}
\arguments{
\item{omicsData_objects}{A list of omicsdata objects containing sample
information matching that in f_meta}

\item{f_meta}{passed from \code{as.multiData}}
}
\value{
A list, each element of which contains a character vector of column
name matches in f_meta for each omicsData object.
}
\description{
Check that the f_meta file contains a column aligned to each omicsData objects
}
\keyword{internal}
