% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliVis.R
\name{format_plot}
\alias{format_plot}
\title{Plot pairwise comparisons and data values in trellData object}
\usage{
format_plot(trellData, ...)
}
\arguments{
\item{trellData}{An object of class "trellData" generated from \code{\link{as.trellData}}.}

\item{...}{further arguments}

\item{p_val}{Specifies p-value for setting graph border colors}

\item{panel_variable}{Specifies what to divide trelliscope panels by, must be a column in trellData. Defaults to cnames$edata_cname of trellData.}

\item{plot_text}{For comparisons only: TRUE/FALSE for p-value text above data}

\item{interactive}{Should the plots be rendered as plotly objects?}

\item{plot_type}{types of plots}

\item{y_limits}{y-limits}
}
\description{
Plot pairwise comparisons and data values in trellData object. Customizable for plot types, y axis limits, paneling variable (what overall group is plotted on each graph arrangement), as well as desired variables for the y and x axis.
}
\author{
Rachel Richardson
}
