% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_test.R
\name{g_test}
\alias{g_test}
\title{Implement the G-test}
\usage{
g_test(omicsData, return_sizes = FALSE)
}
\arguments{
\item{omicsData}{A pmartR data object of any class}

\item{return_sizes}{Logical, should the group sizes be returned as well?  If true, a separate data.frame containing the group sizes is returned}
}
\value{
a data frame with the following columns: group means, global G-test statistic and corresponding p-value
}
\description{
Can test for a qualitiative difference between groups for continuous data (see imd_test) or a quantiative difference between groups
for count data.  For continuous data, this is often used to test if the number of missing values is the same across all groups (i.e.a
qualitative difference).  For count data, this is used to test if the number of counts is the same for all groups.
}
\examples{
library(mintJansson)
rRNA_data <- pmartR::group_designation(omicsData = rRNA_data, main_effects = c("site","treatment"))
group_df <- attr(rRNA_data, "group_DF")
#This is will return a lot of errors because the peptides with too few data haven't 
#been filtered out yet
g_res <- g_test(omicsData = rRNA_data, groupData = group_df)

}
\references{
Webb-Robertson, Bobbie-Jo M., et al. "Combined statistical analyses of peptide intensities and peptide occurrences improves identification of significant peptides from MS-based proteomics data." Journal of proteome research 9.11 (2010): 5748-5756.
}
\author{
Bryan Stanfill
}
