% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliVis.R
\name{generate_plot_message}
\alias{generate_plot_message}
\title{Generates a message stating the specified plotting y-limits}
\usage{
generate_plot_message(...)
}
\arguments{
\item{trellData}{An object of class "trellData" generated from \code{\link{format_data}}.}

\item{comps_y_limits}{For comparisons: Set to "fixed" or "free" for automated y-axis calculating. "fixed" - axis generated based on the maximum/minimum across all plots. "free" - axis axis generated based on the maximum/minimum of individual plot.}

\item{comps_y_range}{For comparisons: Specify a range for the plot y-axis. Will calculate the range based on one of y_max or y_min parameters or from the median of y-values where y_max and y_min are not defined.}

\item{comps_y_max}{For comparisons: Sets the maximum y-value for the y-axis.}

\item{comps_y_min}{For comparisons: Sets the minimum y-value for the y-axis.}

\item{value_y_limits}{For values: Set to "fixed" or "free" for automated y-axis calculating. "fixed" - axis generated based on the maximum/minimum across all plots. "free" - axis axis generated based on the maximum/minimum of individual plot.}

\item{value_y_range}{For values: Specify a range for the plot y-axis. Will calculate the range based on one of y_max or y_min parameters or from the median of y-values where y_max and y_min are not defined.}

\item{value_y_max}{For values: Sets the maximum y-value for the y-axis.}

\item{value_y_min}{For values: Sets the minimum y-value for the y-axis.}
}
\description{
Generates a message stating the specified plotting y-limits
}
\author{
Rachel Richardson
}
