% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{get_data_info}
\alias{get_data_info}
\title{Fetch the data_info attribute}
\usage{
get_data_info(omicsData)
}
\arguments{
\item{omicsData}{An object of class pepData, proData, metabData, lipidData,
or nmrData.}
}
\value{
A list containing seven elements:
        \itemize{

          \item data_scale -- A Character string indicating the scale of the
          data in \code{e_data}.

          \item norm_info -- A list containing a single element indicating
          whether the data in \code{e_data} have been normalized.

          \item num_edata -- The number of unique entries present in the
          \code{edata_cname} column in \code{e_data}.

          \item num_miss_obs -- An integer. The number of missing
          observations in \code{e_data}.

          \item num_zero_obs -- An integer. The number of zeros
          in \code{e_data} for seqData objects.

          \item prop_missing -- A number between 0 and 1. The proportion of
          missing observations in \code{e_data}.

          \item num_samps -- An integer indicating the number of samples or
          columns (excluding the identifier column \code{edata_cname}) in
          \code{e_data}.

          \item data_types -- A character string describing the type of data
          in \code{e_data}.

        }
}
\description{
Retrieves the values in the data_info attribute from an omicsData object.
}
