% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{get_data_scale}
\alias{get_data_scale}
\title{Get data scale}
\usage{
get_data_scale(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData', or 'lipidData', usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, or \code{\link{as.lipidData}}, respectively.}
}
\value{
a character string describing data scale
}
\description{
This function returns data scale attribute
}
\examples{
dontrun{
get_data_scale(omicsData)
}
}
