% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{get_data_scale_orig}
\alias{get_data_scale_orig}
\title{Fetch the original data scale}
\usage{
get_data_scale_orig(omicsObject)
}
\arguments{
\item{omicsObject}{an object of class 'pepData', 'proData', 'metabData',
'lipidData', or 'nmrData'.}
}
\value{
A character string.
}
\description{
Retrieves the character string indicating the scale the data was originally
on when read into R.
}
