% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{get_fdata_cname}
\alias{get_fdata_cname}
\title{Get f_data cname}
\usage{
get_fdata_cname(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData', or 'lipidData', usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, or \code{\link{as.lipidData}}, respectively.}
}
\value{
a character string describing f_data cname
}
\description{
This function returns f_data cname
}
\examples{
dontrun{
get_fdata_cname(omicsData)
}
}
