% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{get_filter_type}
\alias{get_filter_type}
\title{Extracts the types of filters that have been applied. This function will be
used at the beginning of the applyFilt function to give a warning if the
same type of filter has already been applied.}
\usage{
get_filter_type(omicsData)
}
\arguments{
\item{omicsData}{An object of class pepData, proData, metabData, lipidData,
seqData, or nmrData.}
}
\value{
vector of filters used on omicsData
}
\description{
Retrieves the values in the filters attribute from an omicsData object.
}
