% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{get_filters}
\alias{get_filters}
\title{Get Filters}
\usage{
get_filters(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData', or 'lipidData', usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, or \code{\link{as.lipidData}}, respectively.}
}
\value{
a vector of filter names
}
\description{
This function returns filters attribute
}
\examples{
dontrun{
get_filters(omicsData)
}
}
