% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqData_wrappers.R
\name{get_group_formula}
\alias{get_group_formula}
\title{Get formula for group design}
\usage{
get_group_formula(omicsData)
}
\arguments{
\item{omicsData}{an object of type 'seqData', created by \code{\link{as.seqData}}}
}
\value{
A list with two elements:
\itemize{
 \item grouping_info: A data.frame with the grouping information used in
  the statistical analysis
\item formula_string: A character string with the formula used in the
 statistical analysis
}
}
\description{
For generating group design formulas and correctly ordered group data for
seqData statistical functions
}
