% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{get_group_table}
\alias{get_group_table}
\title{Get group table}
\usage{
get_group_table(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData', or 'lipidData', usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, or \code{\link{as.lipidData}}, respectively.}
}
\value{
a table containing number of samples per group
}
\description{
This function returns a table with number of samples per group
}
\examples{
dontrun{
get_group_table(omicsData)

}

}
