% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{get_isobaric_info}
\alias{get_isobaric_info}
\title{Fetch the isobaric_info attribute}
\usage{
get_isobaric_info(omicsData)
}
\arguments{
\item{omicsData}{An object of class pepData, proData, metabData, lipidData,
or nmrData.}
}
\value{
A list containing the following six elements:
        \itemize{
        
          \item exp_cname -- 
          
          \item channel_cname -- 
          
          \item refpool_channel -- 
          
          \item refpool_cname -- 
          
          \item refpool_notation -- 
          
          \item norm_info -- A list containing a single logical element that
          indicates whether the data have been normalized to a reference
          pool.
        
        }
}
\description{
Retrieves the values in the isobaric_info attribute from an omicsData object.
}
