% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{get_isobaric_norm}
\alias{get_isobaric_norm}
\title{Get isobaric norm}
\usage{
get_isobaric_norm(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'isobaricpepData' or 'proData', usually created by \code{\link{as.pepData}}, \code{\link{as.isobaricpepData}}.}
}
\value{
a character string describing isobaric norm attribute
}
\description{
This function returns the isobaric norm attribute
}
\examples{
dontrun{
get_isobaric_norm(omicsData)
}
}
