% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{get_meta_info}
\alias{get_meta_info}
\title{Fetch the meta_info attribute}
\usage{
get_meta_info(omicsData)
}
\arguments{
\item{omicsData}{An object of class pepData, proData, metabData, lipidData,
or nmrData.}
}
\value{
A list containing two elements:
        \itemize{
        
          \item meta_data -- Logical. Indicates if the \code{e_meta} data
          frame was provided. 
        
          \item num_emeta -- The number of unique entries present in the
          \code{emeta_cname} column in \code{e_meta}.
        
        }
}
\description{
Retrieves the values in the meta_info attribute from an omicsData object.
}
