% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{get_nmr_info}
\alias{get_nmr_info}
\title{Fetch the nmr_info attribute}
\usage{
get_nmr_info(omicsData)
}
\arguments{
\item{omicsData}{An object of class pepData, proData, metabData, lipidData,
or nmrData.}
}
\value{
A list containing the following three elements:
        \itemize{

          \item metabolite_name --

          \item sample_property_cname --

          \item norm_info -- A list containing two logical elements that
          indicate i) whether the data have been normalized to a spiked in
          metabolite or to a property taking sample-specific values and ii)
          whether the data have been back transformed so the values are on a
          similar scale to the raw values before normalization.

        }
}
\description{
Retrieves the values in the nmr_info attribute from an omicsData object.
}
