% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imd_anova.R
\name{group_comparison_imd}
\alias{group_comparison_imd}
\title{Group comparisons for the g-test}
\usage{
group_comparison_imd(groupData, comparisons, observed, absent)
}
\arguments{
\item{groupData}{data frame that assigns sample names to groups}

\item{comparisons}{dataframe that defiens the comparsions of interest}

\item{observed}{matrix of number of observed counts}

\item{absent}{matrix of number of observed counts}
}
\value{
A data.frame containing the p-values from the group comparisons.
}
\description{
Takes the results of imd_test() and returns group comparison p-values
}
\author{
Bryan Stanfill
}
