% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statRes_class.R
\name{gtest_heatmap}
\alias{gtest_heatmap}
\title{Plot a heatmap for the g-test results of imd-anova}
\usage{
gtest_heatmap(
  volcano,
  pval_thresh = 0.05,
  show_sig = TRUE,
  sig_text = TRUE,
  interactive = FALSE,
  color_low = NULL,
  color_high = NULL,
  plotly_layout = NULL
)
}
\arguments{
\item{volcano}{`data.frame` produced by \link{pmartR::make_volcano_plot_df}}

\item{show_sig}{Boolean whether to show the visual indicator that a certain 
bin combination is significant by the g-test}

\item{sig_text}{Boolean whether to show the text indicating the number of 
biomolecules falling into each bin. (in addition to the legend)}

\item{interactive}{passed from \code{\link[pmartR:plot.statRes]{pmartR::plot.statRes()}}.  If T, 
will build a plotly version of the plot.}

\item{color_low}{A character string specifying the color of the gradient for
low count values.}

\item{color_high}{A character string specifying the color of the gradient for
high count values.}

\item{plotly_layout}{A list of arguments, not including the plot, to be 
passed to plotly::layout if interactive = T.}
}
\value{
`ggplot or plotly` object, depending on if \code{interactive} is set 
to TRUE or FALSE respectively. A g-test heatmap
}
\description{
Plots a heatmap showing bins for combinations of # of biomolecules present
across groups.  Bins are colored by the number of biomolecules falling into
each bin, and have indicators for significance by g-test.  Internal to 
\link{pmartR::plot.statRes}
}
\keyword{internal}
