% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliVis.R
\name{list_y_limits}
\alias{list_y_limits}
\title{Digests user input for y_limits}
\usage{
list_y_limits(plot_type, y_limits)
}
\arguments{
\item{plot_type}{plots to graph in trelliVis}

\item{y_limits}{User input y-axis limits for plots}
}
\description{
Takes lists/strings/numeric input and digests as a list of y_limits for each plot type.
}
\seealso{
\link[pmartR]{format_plot}
}
\author{
Rachel Richardson
}
