% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingval_heatmapplot.R
\name{missingval_heatmapplot}
\alias{missingval_heatmapplot}
\title{missingval_heatmapplot}
\usage{
missingval_heatmapplot(omicsData)
}
\description{
takes in omicsData and returns a heatmap of omicsData$e_data
}
