% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingval_result.R
\name{missingval_result}
\alias{missingval_result}
\title{missingval_result}
\usage{
missingval_result(omicsData)
}
\description{
input is an omicsData object, outputs a list of two data frames, one contains NA by sample, the second contains NA by molecule
}
