% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{missingval_volcanoplot}
\alias{missingval_volcanoplot}
\title{missingval_volcanoplot}
\usage{
missingval_volcanoplot(statRes, comparison, x_lab = NULL, ...)
}
\arguments{
\item{comparison}{can either be a character string name of the comparison to plot, or an integer index refering to the comparisons attribute vector}

\item{vlines}{The x coordinate (integer in absolute value) where to draw vertical lines, defaults to NULL}

\item{pvalue_threshold}{numeric value, draws horizontal line at value, defaults to NULL
 
 
\tabular{ll}{
\code{x_lab} \tab character string to be used for x-axis label. Defaults to NULL \cr
\code{y_lab} \tab character string to be used for y-axis label. Defaults to NULL \cr
\code{title_plot} \tab character string to be used for the plot title. Defaults to NULL. \cr
\code{title_size} \tab integer value specifying the font size for the plot title. Default is 14. \cr
\code{x_lab_size} \tab integer value indicating the font size for the x-axis. Defaults to 11. \cr
\code{y_lab_size} \tab integer value indicating the font size for the y-axis. Defaults to 11. \cr
\code{bw_theme} \tab logical indicator of whether to use the "theme_bw". Defaults to FALSE, in which case the ggplot2 default theme is used. \cr
}}
}
\description{
fold change vs t-Test p-value
}
