% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyFilt.R
\name{MSomics_filter_worker}
\alias{MSomics_filter_worker}
\title{Remove items that need to be filtered out}
\usage{
MSomics_filter_worker(filter_object, omicsData)
}
\arguments{
\item{filter_object}{a list created by the functions above}

\item{omicsData}{an object of the class \code{pepData}, \code{proData}, \code{lipidData}, \code{metabData}, \code{cDNAdata}, or \code{gDNAdata} usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.lipidData}}, \code{\link{as.metabData}}, \code{\link{as.cDNAdata}}, \code{\link{as.gDNAdata}}, respectively.}
}
\value{
list
}
\description{
This function removes
}
\author{
Kelly Stratton, Lisa Bramer
}

