% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmissing_per_group.R
\name{nonmissing_per_group}
\alias{nonmissing_per_group}
\title{Computes the Number of Non-Missing Data Points by Group}
\usage{
nonmissing_per_group(omicsData)
}
\arguments{
\item{omicsData}{an optional object of one of the classes "pepData",
"proData", "metabData",  "lipidData", or "nmrData", usually created by
\code{\link{as.pepData}}, \code{\link{as.proData}},
\code{\link{as.metabData}}, \code{\link{as.lipidData}}, or
\code{\link{as.nmrData}}, respectively. Either omicsData or all of
e_data, groupDF, cname_id, and samp_id must be provided.}
}
\value{
a list of length two. The first element giving the total number of
        possible samples for each group. The second element giving a
        data.frame with the first column giving the peptide and the second
        through kth columns giving the number of non-missing observations for
        each of the \code{k} groups.
}
\description{
This function computes the number of non-missing observations for samples,
based on a group designation, for every biomolecule in the dataset
}
\author{
Lisa Bramer, Kelly Stratton
}
