% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spans.R
\name{normalize_global_basic}
\alias{normalize_global_basic}
\title{Normalize e_data within SPANS}
\usage{
normalize_global_basic(edata, norm_fn)
}
\arguments{
\item{norm_fn}{character string indicating the normalization function to use
for normalization. See details for the current offerings.}

\item{e_data}{a \eqn{p \times n + 1} data.frame, where \eqn{p} is the number
of peptides, lipids, or metabolites and \eqn{n} is the number of samples.
Each row corresponds to data for a peptide, protein, lipid, or metabolite,
with a column giving the identifer name.}
}
\value{
A list containing the location and scale parameters for normalizing
  the data.
}
\description{
This function is intended to be used in SPANS only. It is a VERY trimmed down
version of normalize_global. It is trimmed down because within SPANS we only
need the norm_params element from the output of the normalize_global
function. All of the other options and output can be ignored.
}
