% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_isobaric.R
\name{normalize_isobaric}
\alias{normalize_isobaric}
\title{Normalize an object of class isobaricpepData}
\usage{
normalize_isobaric(omicsData, apply_norm = FALSE)
}
\arguments{
\item{omicsData}{an object of the class 'isobaricpepData'}

\item{apply_norm}{logical, indicates whether normalization should be applied to omicsData$e_data}
}
\description{
Description of what this function does
}
\examples{
 
dontrun{ 
library(pmartRdata)
library(plyr)
data(isobaric_object)

isobaric_object = edata_transform(isobaric_object, "log2")
result = normalize_isobaric(isobaric_object, apply_norm = TRUE)
}

}
