% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_lowess.R
\name{normalize_lowess}
\alias{normalize_lowess}
\title{Lowess Normalization}
\usage{
normalize_lowess(omicsData, min_prop = NULL)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData', or 'lipidData', created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, or \code{\link{as.lipidData}}, respectively. The function \code{\link{group_designation}} must have been run on omicsData to use several of the subset functions (i.e. rip and ppp_rip).}

\item{min_prop}{numeric threshold between 0 and 1 giving the minimum value for the proportion of features subset (rows of \code{e_data})}
}
\value{
The normalized data is returned in an object of the appropriate S3 class (e.g. pepData), on the same scale as omicsData (e.g. if omicsData contains log2 transformed data, the normalization will be performed on the non-log2 scale and then re-scaled after normalization to be returned on the log2 scale).
}
\description{
Perform Lowess normalization
}
\details{
Intensity-dependent normalization. We use the robust scatter plot smoother ‘lowess’, implemented in the statistical software package R (17), to perform a local A-dependent normalization log2R/G → log2R/G – c(A) = log2R/[k(A)G] where c(A) is the lowess fit to the MA-plot. The lowess scatter plot smoother performs robust locally linear fits. In particular, it will not be affected by a small percentage of differentially expressed genes, which will appear as outliers in the MA-plot. The user-defined parameter f is the fraction of the data used for smoothing at each point; the larger the f value, the smoother the fit. We typically use f  = 40%. It is applied to data on the abundance scale (e.g. not a log scale). It is often used for microarry data. (https://www.ncbi.nlm.nih.gov/pmc/articles/PMC100354/)
}
\examples{
dontrun{
library(pmartRdata)
data(lipid_object)
norm_data <- normalize_lowess(omicsData = lipid_object, subset_fn = "all", norm_fn = "median", apply_norm = TRUE, backtransform = TRUE)
}

}
\references{

}
\author{
Kelly Stratton
}
