% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_quantile.R
\name{normalize_quantile}
\alias{normalize_quantile}
\title{Quantile Normalization}
\usage{
normalize_quantile(omicsData, min_prop = NULL)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData', or 'lipidData', created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, or \code{\link{as.lipidData}}, respectively. The function \code{\link{group_designation}} must have been run on omicsData to use several of the subset functions (i.e. rip and ppp_rip).}

\item{min_prop}{numeric threshold between 0 and 1 giving the minimum value for the proportion of features subset (rows of \code{e_data})}
}
\value{
The normalized data is returned in an object of the appropriate S3 class (e.g. pepData), on the same scale as omicsData (e.g. if omicsData contains log2 transformed data, the normalization will be performed on the non-log2 scale and then re-scaled after normalization to be returned on the log2 scale).
}
\description{
Perform quantile normalization
}
\details{
Quantile normalization is an algorithm for normalizing a set of data vecteors by giving them the same distribution. It is applied to data on the abundance scale (e.g. not a log scale). It is often used for microarry data.
}
\examples{
dontrun{
library(pmartRdata)
data(lipid_object)
norm_data <- normalize_quantile(omicsData = lipid_object, subset_fn = "all", norm_fn = "median", apply_norm = TRUE, backtransform = TRUE)
}

}
\references{

}
\author{
Kelly Stratton
}
