% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normRes_tests.R
\name{normRes_tests}
\alias{normRes_tests}
\title{Test the location and scale parameters from a normalization procedure}
\usage{
normRes_tests(norm_obj, test_fn = "kw")
}
\arguments{
\item{norm_obj}{An object of class 'pepData', 'proData', 'lipidData',
'metabData', 'isobaricpepData' that has had \code{normalize_global()} run
on it, or a 'normRes' object}

\item{test_fn}{The statistical test to use, current tests are "anova" and
"kw" for a Kruskal-Wallis test.}
}
\value{
A list with 2 entries containing the p_value of the test performed on
  the location and scale (if it exists) parameters.
}
\description{
Computes p-values from a test of dependence between normalization parameters
and group assignment of a normalized omicsData or normRes object.
}
