% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value_adjustment.R
\name{p_adjustment_anova}
\alias{p_adjustment_anova}
\title{Adjust p-values for multiple comparisons}
\usage{
p_adjustment_anova(p_values = NULL, diff_mean = NULL, t_stats = NULL,
  sizes = NULL, pval_adjust = "None")
}
\arguments{
\item{p_values}{A matrix (or \code{data.frame}) of p-values to be adjusted.}

\item{diff_mean}{A matrix (or \code{data.frame}) of groups means that are to be compared}

\item{t_stats}{A matrix (or \code{data.frame}) of t-test statistics resulting from from standard procedures}

\item{sizes}{A matrix (or \code{data.frame}) of group sizes}

\item{pval_adjust}{character vector specifying the type of multiple comparisons adjustment to implement. A NULL value corresponds to no adjustment. Valid options include: holm, bonferroni, dunnett, tukey or none.}
}
\value{
a data frame with the following columns: group means, global G-test statistic and corresponding p-value
}
\description{
Depending upon the \code{pval_adjust} method selected, the supplied p_values are compared against an adjusted \code{pval_thresh} value or the provided
means are used to compute new statistics, p-values are computed and compared against the provided \code{pval_thresh}.  A \code{data.frame} that indicates which
of the tests are significant, 1 if significant or 0 if insignificant.  If \code{means} is also provided and the p-value is signficant then the direction
of the change is indicated by the sign on 1, i.e., means<0 and p_value<pval_thresh will return -1, similarly for means>0.
}
\examples{
library(MSomicsDATA)
library(MSomicsQC)
mypepData <- edata_transform(mintR_data = pep_pepData, data_scale = "log2")
mypepData <- group_designation(mintR_data = mypepData, main_effects = c("Condition"))
group_df <- attr(mypepData, "group_DF")
imdanova_Filt <- imdanova_filter(mintR_data = mypepData)
mypepData <- MSomics_filter(filter_object = imdanova_Filt, mintR_data = mypepData, min_nonmiss_anova=2)
anova_res <- anova_test(omicsData = mypepData, groupData = group_df)

adjusted_pvalues <- p_adjustment_anova(p_values = anova_res$Comparisons, diff_mean = anova_res$Fold)
}
\author{
Bryan Stanfill
}
