% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot.customFilt}
\alias{plot.customFilt}
\title{plot.customFilt}
\usage{
\method{plot}{customFilt}(filter_obj)
}
\arguments{
\item{filter_obj}{An object of class customFilt.}
}
\description{
Currently plotting a customFilt object is not supported
}
