% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot.dataRes}
\alias{plot.dataRes}
\title{Plots an object of class dataRes}
\usage{
\method{plot}{dataRes}(
  dataRes_obj,
  metric = NULL,
  density = FALSE,
  ncols = NULL,
  interactive = FALSE,
  x_lab = NULL,
  x_lab_sd = NULL,
  x_lab_median = NULL,
  y_lab = NULL,
  y_lab_sd = NULL,
  y_lab_median = NULL,
  x_lab_size = 11,
  y_lab_size = 11,
  x_lab_angle = NULL,
  title_lab = NULL,
  title_lab_sd = NULL,
  title_lab_median = NULL,
  title_lab_size = 14,
  legend_lab = NULL,
  legend_position = "right",
  point_size = 2,
  bin_width = 1,
  bw_theme = TRUE,
  palette = NULL
)
}
\arguments{
\item{dataRes_obj}{An object of class dataRes. A dataRes object is created by
the \code{edata_summary} function.}

\item{metric}{A character string indicating which metric to use in plot, one
of 'mean', 'median', 'sd, 'pct_obs', 'min', or 'max'}

\item{density}{logical default to FALSE, if set to true a density plot of the
specified metric is returned}

\item{ncols}{An integer specifying the number columns for the histogram
facet_wrap. This argument is used when \code{metric} is not null. The
default is NULL.}

\item{interactive}{Logical. If TRUE produces an interactive plot.}

\item{x_lab}{A character string specifying the x-axis label when the metric
argument is NULL. The default is NULL in which case the x-axis label will
be "count".}

\item{x_lab_sd}{A character string used for the x-axis label for the
mean/standard deviation plot when the \code{metric} argument is not NULL.}

\item{x_lab_median}{A character string used for the x-axis label for the
mean/median plot when the \code{metric} argument is not NULL.}

\item{y_lab}{A character string specifying the y-axis label. The default is
NULL in which case the y-axis label will be the metric selected for the
\code{metric} argument.}

\item{y_lab_sd}{A character string used for the y-axis label for the
mean/standard deviation plot when the \code{metric} argument is not NULL.}

\item{y_lab_median}{A character string used for the y-axis label for the
mean/median plot when the \code{metric} argument is not NULL.}

\item{x_lab_size}{An integer value indicating the font size for the x-axis.
The default is 11.}

\item{y_lab_size}{An integer value indicating the font size for the y-axis.
The default is 11.}

\item{x_lab_angle}{An integer value indicating the angle of x-axis labels.}

\item{title_lab}{A character string specifying the plot title when the
\code{metric} argument is NULL.}

\item{title_lab_sd}{A character string used for the plot title for the
mean/standard deviation plot when the \code{metric} argument is not NULL.}

\item{title_lab_median}{A character string used for the plot title for the
mean/median plot when the \code{metric} argument is not NULL.}

\item{title_lab_size}{An integer value indicating the font size of the plot
title. The default is 14.}

\item{legend_lab}{A character string specifying the legend title.}

\item{legend_position}{A character string specifying the position of the
legend. Can be one of "right", "left", "top", or "bottom". The default is
"right".}

\item{point_size}{An integer specifying the size of the points. The default
is 2.}

\item{bin_width}{An integer indicating the bin width in a histogram. The
default is 0.5.}

\item{bw_theme}{Logical. If TRUE uses the ggplot2 black and white theme.}

\item{palette}{A character string indicating the name of the RColorBrewer
palette to use. For a list of available options see the details section in
\code{\link[RColorBrewer]{RColorBrewer}}.}
}
\value{
plots ggplot2 object
}
\description{
For plotting an S3 object of type dataRes
}
\details{
This function can only create plots of dataRes objects whose 'by' ==
  'molecule' and 'groupvar' attribute is non NULL
}
\examples{
\dontrun{
library (pmartRdata)

data(lipid_object)

lipid_object = edata_transform(lipid_object, "log2")

result = edata_summary(omicsData = lipid_object,
                       by = "molecule",
                       groupvar = "Condition")

plot(result)
}

}
