% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_isobaricnormRes.R
\name{plot.isobaricnormRes}
\alias{plot.isobaricnormRes}
\title{Plots an object of class isobaricnormRes}
\usage{
\method{plot}{isobaricnormRes}(isobaricnormRes_object, x_lab = NULL, ...)
}
\arguments{
\item{x_lab}{character string to be used for x-axis label. Defaults to NULL}

\item{...}{further arguments 

\tabular{ll}{
\code{y_lab} \tab character string to be used for y-axis label. Defaults to NULL \cr
\code{title_plot} \tab character string to be used for the plot title. Defaults to NULL. \cr
\code{title_size} \tab integer value specifying the font size for the plot title. Default is 14. \cr
\code{x_lab_size} \tab integer value indicating the font size for the x-axis. Defaults to 11. \cr
\code{y_lab_size} \tab integer value indicating the font size for the y-axis. Defaults to 11. \cr
\code{bw_theme} \tab logical indicator of whether to use the "theme_bw". Defaults to FALSE, in which case the ggplot2 default theme is used. \cr
\code{order} \tab logical indicates whether to order data by exp_cname'}}

\item{an}{object of type isobaricnormRes, created by \code{\link{normalize_isobaric}}}
}
\value{
plots ggplot2 object
}
\description{
For plotting an S3 object of type 'isobaricnormRes':
}
\examples{
dontrun{
library(pmartRdata)
data(isobaric_object)
result = normalize_isobaric(isobaric_object, apply_norm = F)

plot(result)
}

}
