% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot.naRes}
\alias{plot.naRes}
\title{Plots an object of class naRes}
\usage{
\method{plot}{naRes}(
  naRes_obj,
  omicsData,
  plot_type = "bar",
  order_by = NULL,
  color_by = NULL,
  interactive = FALSE,
  x_lab_bar = NULL,
  x_lab_scatter = NULL,
  y_lab_bar = NULL,
  y_lab_scatter = NULL,
  x_lab_size = 11,
  y_lab_size = 11,
  x_lab_angle = 60,
  title_lab_bar = NULL,
  title_lab_scatter = NULL,
  title_lab_size = 14,
  legend_lab_bar = NULL,
  legend_lab_scatter = NULL,
  legend_position = "right",
  point_size = 2,
  text_size = 3,
  bar_width = 0.8,
  bw_theme = TRUE,
  palette = NULL,
  display_count = TRUE,
  coordinate_flip = FALSE,
  use_VizSampNames = FALSE
)
}
\arguments{
\item{naRes_obj}{A list of two data frames, one contains NA values by sample,
the second contains NA values by molecule}

\item{omicsData}{an object of class 'pepData', 'proData', 'metabData',
'lipidData', or 'nmrData', created by \code{\link{as.pepData}},
\code{\link{as.proData}}, \code{\link{as.metabData}},
\code{\link{as.lipidData}}, or \code{\link{as.nmrData}}, respectively.}

\item{plot_type}{A character string specifying which type of plot to produce.
The two options are 'bar' or 'scatter'.}

\item{order_by}{A character string specifying a main effect by which to order
the bar plot. This main effect must be found in the column names of f_data
in the omicsData object. If \code{order_by} is "Group", the bar plot
will be ordered by the group variable from the group_designation function.
If NULL (default), the bar plot will be displayed in the order they appear
in the data.}

\item{color_by}{A character string specifying a main effect by which to color
the bar plot. This main effect must be found in the column names of f_data
in the omicsData object. If \code{color_by} is "Group", the bar plot
will be colored by the group variable from the group_designation function.
If NULL (default), the bar plot will have one default color.}

\item{interactive}{Logical. If TRUE produces an interactive plot.}

\item{x_lab_bar}{A character string used for the x-axis label for the bar
plot.}

\item{x_lab_scatter}{A character string used for the x-axis label for the
scatter plot.}

\item{y_lab_bar}{A character string used for the y-axis label for the bar
plot.}

\item{y_lab_scatter}{A character string used for the y-axis label for the
scatter plot.}

\item{x_lab_size}{An integer value indicating the font size for the x-axis.
The default is 11.}

\item{y_lab_size}{An integer value indicating the font size for the y-axis.
The default is 11.}

\item{x_lab_angle}{An integer value indicating the angle of x-axis labels.}

\item{title_lab_bar}{A character string used for the plot title when
\code{plot_type} is 'bar'.}

\item{title_lab_scatter}{A character string used for the plot title when
\code{plot_type} is 'scatter'.}

\item{title_lab_size}{An integer value indicating the font size of the plot
title. The default is 14.}

\item{legend_lab_bar}{A character string specifying the legend title when
creating a bar plot.}

\item{legend_lab_scatter}{A character string specifying the legend title when
creating a scatter plot.}

\item{legend_position}{A character string specifying the position of the
legend. Can be one of "right", "left", "top", or "bottom". The default is
"right".}

\item{point_size}{An integer specifying the size of the points. The default
is 2.}

\item{text_size}{An integer specifying the size of the text (number of
missing values by sample) within the bar plot. The default is 3.}

\item{bar_width}{An integer indicating the width of the bars in the bar plot.
The default is 0.8.}

\item{bw_theme}{Logical. If TRUE uses the ggplot2 black and white theme.}

\item{palette}{A character string indicating the name of the RColorBrewer
palette to use. For a list of available options see the details section in
\code{\link[RColorBrewer]{RColorBrewer}}.}

\item{display_count}{Logical. Indicates whether the missing value counts by
sample will be displayed on the bar plot. The default is TRUE.}

\item{coordinate_flip}{Logical. Indicates whether the x and y axes will be
flipped. The default is FALSE.}

\item{use_VizSampNames}{Logical. Indicates whether to use custom sample
names. The default is FALSE.}
}
\value{
plots ggplot2 object
}
\description{
For plotting an S3 object of type 'naRes'
}
\details{
This function takes in an object of class naRes and creates either a
  bar or scatter plot of missing values. When plot_type = 'bar', a sample
  name by missing values count bar chart is returned. When plot_type =
  'scatter' a mean intensity vs number of missing values (per molecule)
  scatter plot is returned. Note: If the omicsData object has had
  \code{\link{group_designation}} applied to it, the points in the plot will
  be colored by group.
}
\examples{
\dontrun{
library(pmartRdata)

data("lipid_object")

result<- missingval_result(lipid_object)

plot(result, plog_type = "bar", x_lab_angle = 50)
}

}
