% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pmartR.R
\name{plot.cvFilt}
\alias{plot.cvFilt}
\title{plot.cvFilt}
\usage{
\method{plot}{cvFilt}(filter_object, cv_threshold = NULL, ...)
}
\arguments{
\item{cv_threshold}{draws a vertical line at the pooled cv cutoff, values to the left of this cutoff are dropped if the filter is applied.}

\item{...}{Additional arguments
\tabular{ll}{
\code{log_scale} \tab logical indicating whether to use a log2 transformed x-axis. Defaults to TRUE.\cr
\code{x_lab} \tab character string to be used for x-axis label. Defaults to NULL, in which case a default label is used. \cr
\code{y_lab} \tab character string to be used for y-axis label. Defaults to NULL, in which case a default label is used. \cr
\code{title_plot} \tab character string to be used for the plot title. Defaults to NULL, in which case a default title is used. \cr
\code{title_size} \tab integer value specifying the font size for the plot title. Default is 14. \cr
\code{x_lab_size} \tab integer value indicating the font size for the x-axis. Defaults to 11. \cr
\code{y_lab_size} \tab integer value indicating the font size for the y-axis. Defaults to 11. \cr
\code{bw_theme} \tab logical indicator of whether to use the "theme_bw". Defaults to FALSE, in which case the ggplot2 default theme is used. \cr
}}
}
\description{
For plotting an S3 object of type 'cvFilt':
}
\examples{
data(pep_object)
pep_object <- group_designation(omicsData = pep_object, main_effects = "Condition")
cvfilt <- cv_filter(pep_object)

plot(cvfilt, cv_threshold = 20)
plot(cvfilt, cv_threshold = 10, log_scale = FALSE)

}
