% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pmartR.R
\name{plot-pmartR}
\alias{plot-pmartR}
\title{Produce a plot of a pmartR S3 Object}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'lipidData', or 'metabData' usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.lipidData}}, or  \code{\link{as.metabData}}, respectively.}

\item{corRes_object}{an object of class corRes. A correlation matrix of all samples.}

\item{filter_object}{a filter object for the respective \code{omicsData} class.}

\item{...}{further arguments}
}
\value{
a ggplot summarizing the pmartR object
}
\description{
This function will provide a plot for a \code{omicsData} object, any of the filter objects in pmartR, or a \code{corRes} object.
}
\details{
Various further arguments can be specified depending on the class of the object being plotted.
}
\examples{
dontrun{
library(pmartRdata)
data("pep_object")
cor_matrix <- cor_result(pep_object)
plot(cor_matrix)
plot(cor_matrix, interactive = TRUE)

to_filter <- molecule_filter(pep_object)
plot(to_filter, min_num = 4)

data("pro_object")
plot(pro_object)
plot(pro_object, order_by = "Condition")
plot(pro_object, color_by = "Condition")
plot(pro_object, facet_by = "Condition")
plot(pro_object, facet_by = "Condition", facet_cols = 1)

pep_object2 <- group_designation(pep_object, main_effects = "Condition")
pep_object2 <- edata_transform(pep_object2, "log2")
pca_res <- dim_reduction(pep_object2, k = 2)
plot(pca_res)
}
}
