% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot.SPANSRes}
\alias{plot.SPANSRes}
\title{Plots an object of class SPANSRes}
\usage{
\method{plot}{SPANSRes}(
  SPANSRes_obj,
  interactive = FALSE,
  x_lab = NULL,
  y_lab = NULL,
  x_lab_size = 11,
  y_lab_size = 11,
  x_lab_angle = NULL,
  title_lab = NULL,
  title_lab_size = 14,
  legend_lab = NULL,
  legend_position = "right",
  color_low = NULL,
  color_high = NULL,
  bw_theme = TRUE
)
}
\arguments{
\item{SPANSRes_obj}{an object of the class 'SPANSRes', usually created by
\code{\link{spans_procedure()}}.}

\item{interactive}{Logical. If TRUE produces an interactive plot.}

\item{x_lab}{A character string specifying the x-axis label.}

\item{y_lab}{A character string specifying the y-axis label.}

\item{x_lab_size}{An integer value indicating the font size for the x-axis.
The default is 11.}

\item{y_lab_size}{An integer value indicating the font size for the y-axis.
The default is 11.}

\item{x_lab_angle}{An integer value indicating the angle of x-axis labels.
The default is NULL.}

\item{title_lab}{A character string specifying the plot title.}

\item{title_lab_size}{An integer value indicating the font size of the plot
title. The default is 14.}

\item{legend_lab}{A character string specifying the legend title.}

\item{legend_position}{A character string specifying the position of the
legend. Can be one of "right", "left", "top", "bottom", or "none". The
default is "none".}

\item{color_low}{A character string specifying the color of the gradient for
low values.}

\item{color_high}{A character string specifying the color of the gradient for
high values.}

\item{bw_theme}{Logical. If TRUE uses the ggplot2 black and white theme.}
}
\value{
plots a plotly object
}
\description{
For plotting an S3 object of type 'SPANSRes'
}
