% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot.totalCountFilt}
\alias{plot.totalCountFilt}
\title{plot.totalCountFilt}
\usage{
\method{plot}{totalCountFilt}(
  filter_obj,
  min_count = NULL,
  interactive = FALSE,
  x_lab = NULL,
  y_lab = NULL,
  x_lab_size = 11,
  y_lab_size = 11,
  x_lab_angle = 0,
  title_lab = NULL,
  title_lab_size = 14,
  legend_lab = "",
  legend_position = "right",
  text_size = 3,
  bar_width = 0.8,
  bw_theme = TRUE,
  log_total_counts = TRUE,
  palette = NULL
)
}
\arguments{
\item{filter_obj}{An object of class totalCountFilt that contains the molecule
identifier and the number of total counts for which the molecule was measured
(not NA).}

\item{min_count}{An integer specifying the minimum number of samples in which a
biomolecule must appear. If a value is specified, a horizontal line will be
drawn when \code{cumulative=TRUE}, and bars will be colored appropriately
if \code{cumulative=FALSE}.  Defaults to NULL.}

\item{interactive}{Logical. If TRUE produces an interactive plot.}

\item{x_lab}{A character string specifying the x-axis label.}

\item{y_lab}{A character string specifying the y-axis label. The default is
NULL in which case the y-axis label will be the metric selected for the
\code{metric} argument.}

\item{x_lab_size}{An integer value indicating the font size for the x-axis.
The default is 11.}

\item{y_lab_size}{An integer value indicating the font size for the y-axis.
The default is 11.}

\item{x_lab_angle}{An integer value indicating the angle of x-axis labels.
The default is 0.}

\item{title_lab}{A character string specifying the plot title.}

\item{title_lab_size}{An integer value indicating the font size of the plot
title. The default is 14.}

\item{legend_lab}{A character string specifying the legend title.}

\item{legend_position}{A character string specifying the position of the
legend. Can be one of "right", "left", "top", "bottom", or "none". The
default is "none".}

\item{text_size}{An integer specifying the size of the text (number of
biomolecules by sample) within the bar plot. The default is 3.}

\item{bar_width}{An integer indicating the width of the bars in the bar plot.
The default is 0.8.}

\item{bw_theme}{Logical. If TRUE uses the ggplot2 black and white theme.}

\item{log_total_counts}{Logical. Indicates if the X-axis should be on 
the log scale. The default is TRUE.}

\item{palette}{A character string indicating the name of the RColorBrewer
palette to use. For a list of available options see the details section in
\code{\link[RColorBrewer]{RColorBrewer}}.}

\item{cumulative}{logical indicating whether the number of biomolecules
observed in \emph{at least} (TRUE) x number of samples or \emph{exactly}
(FALSE) x number of samples should be plotted.  Defaults to TRUE.}
}
\description{
For plotting an S3 object of type 'totalCountFilt':
}
\examples{
\dontrun{
data(seq_object)
seqfilt <- total_count_filter(pep_object)
plot(seqfilt, min_count = 5)
}

}
