% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statRes_class.R
\name{plot.statRes}
\alias{plot.statRes}
\title{Plotting function for `statRes` objects}
\usage{
\method{plot}{statRes}(x, plot_type = NULL, fc_threshold = NULL, ...)
}
\arguments{
\item{x}{`statRes` object to be plotted, usually the result of `imd_anova`}

\item{plot_type}{defines which plots to be produced, options are "bar", "heatmap", "volcano"; if left `NULL` then all options are produced}

\item{fc_threshold}{optional threshold value for fold change estimates that's added to the volcano plot}
}
\description{
Produces plots that summarize the results contained in a `statRes` object.
}
\examples{

library(MSomicsSTAT)
library(MSomicsDATA)
library(MSomicsQC)
#Trasnform the data
attr(pro_proData,"cnames")$fdata_cname <- "SampleID"

#Group the data by condition
myproData <- group_designation(omicsData = pro_proData, main_effects = c("Condition"))

#Apply the IMD ANOVA filter
imdanova_Filt <- imdanova_filter(omicsData = myproData)
myproData <- MSomics_filter(filter_object = imdanova_Filt, omicsData = myproData, min_nonmiss_anova=2)

#Implement the IMD ANOVA method and compuate all pairwise comparisons (i.e. leave the `comparisons` argument NULL)
anova_res <- imd_anova(omicsData = myproData, test_method = 'anova')
plot(anova_res)

imd_res <- imd_anova(omicsData = myproData, test_method = 'gtest')
plot(imd_res)

imd_anova_res <- imd_anova(omicsData = myproData, test_method = 'comb', pval_adjust='bon')
plot(imd_anova_res)
}
