% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqData_wrappers.R
\name{plotVoom}
\alias{plotVoom}
\title{Plot for voom wrapper}
\usage{
plotVoom(limma_voom)
}
\arguments{
\item{limma_voom}{attribute passed in statRes object from seqData where LimmaVoom was run}
}
\value{
plot result
}
\description{
For generating statistics for 'seqData' objects
}
\examples{
\dontrun{
}

}
