% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyFilt.R
\name{pmartR_filter_worker}
\alias{pmartR_filter_worker}
\title{Remove items that need to be filtered out}
\usage{
pmartR_filter_worker(filter_object, omicsData)
}
\arguments{
\item{filter_object}{a list created by the functions above}

\item{omicsData}{an object of the class \code{pepData}, \code{proData}, \code{lipidData}, or \code{metabData}, usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.lipidData}}, or \code{\link{as.metabData}}, respectively.}
}
\value{
list
}
\description{
This function removes
}
\author{
Kelly Stratton, Lisa Bramer
}
