% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pquant.R
\name{pquant}
\alias{pquant}
\title{Protein Quantitation using Mean or Median Peptide Abundances}
\usage{
pquant(pepData, combine_fn)
}
\arguments{
\item{pepData}{omicsData object of class 'pepData'}

\item{combine_fn}{A character string that can either be 'mean' or 'median'.}
}
\value{
An omicsData object of class 'proData'
}
\description{
This function takes in a pepData object and returns a proData object
}
