% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pquant_mean.R
\name{pquant_mean}
\alias{pquant_mean}
\title{Protein Quantitation using Mean Peptide Abundances}
\usage{
pquant_mean(pepData)
}
\arguments{
\item{pepData}{omicsData object of class 'pepData'}
}
\value{
an omicsData object of class 'proData'
}
\description{
This function takes in a pepData object and returns a proData object
}
\examples{
dontrun{
library(pmartRdata)
data(pep_object)
result = pquant_mean(pep_object) 
}

}
