% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pquant_median.R
\name{pquant_median}
\alias{pquant_median}
\title{Protein Quantitation using Median Peptide Abundances}
\usage{
pquant_median(pepData)
}
\arguments{
\item{pepData}{omicsData object of class 'pepData'}
}
\value{
an omicsData object of class 'proData'
}
\description{
This function takes in a pepData object and returns a proData object
}
\examples{
dontrun{
library(pmartRdata)
data(pep_object)
result = pquant_median(pep_object) 
}

}
