% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_filter_objects.R
\name{print.customFilt}
\alias{print.customFilt}
\title{print.customFilt}
\usage{
\method{print}{customFilt}(x, ...)
}
\arguments{
\item{x}{An object of type 'customFilt'}

\item{...}{further arguments passed to or from other methods}
}
\value{
No return value, prints details about x
}
\description{
For printing an S3 object of type 'customFilt'
}
