% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_data_objects.R
\name{print.dataRes}
\alias{print.dataRes}
\title{print.dataRes}
\usage{
\method{print}{dataRes}(x, ...)
}
\arguments{
\item{x}{An object of class 'dataRes'}

\item{...}{further arguments passed to or from other methods}
}
\value{
No return value, prints details about x.
}
\description{
For printing an S3 object of class 'dataRes'
}
