% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_data_objects.R
\name{print.metabData}
\alias{print.metabData}
\title{print.metabData}
\usage{
\method{print}{metabData}(x, ...)
}
\arguments{
\item{x}{An object of type 'metabData'}

\item{...}{further arguments passed to or from other methods}
}
\description{
For printing an S3 object of type 'metabData'
}
