% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_data_objects.R
\name{print.pepData}
\alias{print.pepData}
\title{print.pepData}
\usage{
\method{print}{pepData}(x, ...)
}
\arguments{
\item{x}{An object of type 'pepData'}

\item{...}{further arguments passed to or from other methods}
}
\value{
No return value, prints details about x
}
\description{
For printing an S3 object of type 'pepData'
}
