% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_data_objects.R
\name{print.proData}
\alias{print.proData}
\title{print.proData}
\usage{
\method{print}{proData}(x, ...)
}
\arguments{
\item{x}{An object of type 'proData'}

\item{...}{further arguments passed to or from other methods}
}
\value{
No return value, prints details about x
}
\description{
For printing an S3 object of type 'proData'
}
