% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_filter_objects.R
\name{print.proteomicsFilt}
\alias{print.proteomicsFilt}
\title{print.proteomicsFilt}
\usage{
\method{print}{proteomicsFilt}(x, ...)
}
\arguments{
\item{x}{An object of type 'proteomicsFilt'}

\item{...}{further arguments passed to or from other methods}
}
\description{
For printing an S3 object of type 'proteomicsFilt'
}
