% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_filter_objects.R
\name{print.RNAFilt}
\alias{print.RNAFilt}
\title{print.RNAFilt}
\usage{
\method{print}{RNAFilt}(x, ...)
}
\arguments{
\item{x}{An object of type 'RNAFilt'}

\item{...}{further arguments passed to or from other methods}
}
\description{
For printing an S3 object of type 'RNAFilt'
}
