% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_filter_objects.R
\name{print.totalCountFilt}
\alias{print.totalCountFilt}
\title{print.totalCountFilt}
\usage{
\method{print}{totalCountFilt}(filter_object)
}
\description{
For printing an S3 object of type 'totalCountFilt'
}
