% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary.R
\name{print.rmdFilterSummary}
\alias{print.rmdFilterSummary}
\title{RMD Filter Print Method}
\usage{
\method{print}{rmdFilterSummary}(x, ...)
}
\arguments{
\item{x}{the RMD filter summary to print}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print method for summary of RMD filter
}
