% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary.R
\name{print.RNAFiltSummary}
\alias{print.RNAFiltSummary}
\title{RNA Filter Print Method}
\usage{
\method{print}{RNAFiltSummary}(x, ...)
}
\arguments{
\item{x}{the RNAFilt summary to print}

\item{...}{further arguments passed to or from other methods}
}
\value{
No return value, prints details about x
}
\description{
Print method for summary of RNAFilt
}
