% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary.R
\name{print.totalCountFiltSummary}
\alias{print.totalCountFiltSummary}
\title{Print method for summary of Total Count filter}
\usage{
\method{print}{totalCountFiltSummary}(object)
}
\description{
Print method for summary of Total Count filter
}
