% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_quant.R
\name{prot_quant}
\alias{prot_quant}
\title{prot_quant wrapper function}
\usage{
prot_quant(pepData, method, proteoformRes = NULL)
}
\description{
Takes in a pepData object, the argument 'method' determines the quantitation method to be applied to pepData.
The argument proteoformRes defaults to NULL, is a list of data.frames which are results of applying bpquant to
the original pepData object for each protein. If proteoformRes is provided then,
}

