% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliVis.R
\name{recursive_format}
\alias{recursive_format}
\title{Recursive call of as.Trelldata and associated checks for list inputs}
\usage{
recursive_format(...)
}
\arguments{
\item{yvalues}{y-values for plotting}

\item{testtype}{consideration for different statistical tests}
}
\description{
Checks for validity of list inputs and handles different list combinations as input.
}
\author{
Rachel Richardson
}
