% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dopar.R, R/pipe.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%dopar\%}
\alias{\%>\%}
\title{Objects exported from other packages}
\value{
No return, used solely for roxygen2 documentation

No return, used solely for roxygen2 documentation
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{foreach}{\code{\link[foreach:foreach]{\%dopar\%}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

