% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.omicsData.R
\name{replace_nas}
\alias{replace_nas}
\title{Replace NA with 0}
\usage{
replace_nas(edata, edata_cname)
}
\arguments{
\item{edata}{A \eqn{p \times n + 1} data frame of expression data, where
\eqn{p} is the number of xxx observed and \eqn{n} is the
number of samples.}

\item{edata_cname}{A character string specifying the name of the ID column in
the e_data data frame.}
}
\value{
An updated e_data data frame where all instances of NA have been
        replaced with 0.
}
\description{
This function finds all instances of NA in e_data and replaces them with 0.
}
\details{
This function is used in the as.seqData functions to
         replace any NA values with 0s.
}
