% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.omicsData.R
\name{replace_zeros}
\alias{replace_zeros}
\title{Replace 0 with NA}
\usage{
replace_zeros(e_data, edata_cname)
}
\arguments{
\item{e_data}{A \eqn{p \times n + 1} data frame of expression data, where
\eqn{p} is the number of xxx observed and \eqn{n} is the
number of samples.}

\item{edata_cname}{A character string specifying the name of the ID column in
the e_data data frame.}
}
\value{
An updated e_data data frame where all instances of 0 have been
        replaced with NA.
}
\description{
This function finds all instances of 0 in e_data and replaces them with NA.
}
\details{
This function is used in the as.pepData, as.proData, as.lipidData,
         as.metabData, as.isobaricpepData, and as.nmrData functions to
         replace any 0 values with NAs.
}
