% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_dataRes.R
\name{report_dataRes}
\alias{report_dataRes}
\title{Creates a data frame displaying multiple metrics}
\usage{
report_dataRes(dataRes, minmax = FALSE, digits = 2)
}
\arguments{
\item{minmax}{logical, for specifying whether or not to include minimum and maximum data in the returned data frame. Defaults to FALSE.}

\item{digits}{integer indicating the number of decimal places to round}

\item{statRes}{an object of the class 'statRes', created by \code{\link{summarize}}.}
}
\value{
prints a data frame
}
\description{
This function takes in an object of class 'dataRes' and returns a data frame displaying a combination of metrics. The six summarizing metrics include, mean, standard deviation, median, percent observed, minimum and maximum.
}
\details{
If the 'by' attribute of the input 'statRes' object is set to 'sample', then its groupvar attribute must be set to NULL
}
\examples{
dontrun{
library(pmartRdata)
data(lipid_object)
lipid_object2 <- edata_transform(omicsData = lipid_object, data_scale = "log2")

dataRes_sample = summarize(omicsData = lipid_object2, groupvar = NULL, by = "sample")
report_dataRes(dataRes_sample)
}


}
