% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrollup.R
\name{rrollup}
\alias{rrollup}
\title{Applies rrollup function}
\usage{
rrollup(pepData, parallel = TRUE)
}
\arguments{
\item{pepData}{an omicsData object of class 'pepData'}

\item{parallel}{logical indicating whether or not to use "doParallel" loop in applying rrollup function. Defaults to TRUE.}
}
\value{
an omicsData object of class 'proData'
}
\description{
This function applies the rrollup function to a pepData object for each unique protein and returns a proData object.
}
\examples{
dontrun{
library(pmartRdata)
data(pep_object)
result = rrollup(pepData = pep_object) 
}

}
