% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_filter.R
\name{run_mad}
\alias{run_mad}
\title{Calculate the Median Absolute Deviance (MAD) of Sample Runs}
\usage{
run_mad(data_only)
}
\arguments{
\item{data_only}{a \eqn{p \times n} data.frame, where \eqn{p} is the number of peptides and \eqn{n} is the number of samples.}
}
\value{
data.frame with two elements: Sample, a character vector giving the sample names; and MAD, a numeric vector giving the MAD values
}
\description{
This function calculates the median absolute deviance across data for each sample run.
}
\details{
When calculating the MAD within a sample NA values are ignorned. If all peptide abundance values are missing within a sample, the MAD is replaced by the overall mean MAD values for the data.
}
\author{
Lisa Bramer
}
