% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_filter.R
\name{run_prop_missing}
\alias{run_prop_missing}
\title{Calculate the Fraction of Missing Data of Sample Runs}
\usage{
run_prop_missing(data_only)
}
\arguments{
\item{data_only}{a \eqn{p \times n} data.frame, where \eqn{p} is the number of peptides and \eqn{n} is the number of samples.}
}
\value{
data.frame with two elements: Sample, a character vector giving the sample names; and Prop_missing, a numeric vector giving the fraction of missing values per run
}
\description{
This function calculates the fraction of missing data for each sample run.
}
\author{
Lisa Bramer
}
