% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{set_check_names}
\alias{set_check_names}
\title{\bold{DEPRECATED}: Set check.names attribute of omicsData object}
\usage{
set_check_names(omicsData, set_to = TRUE)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData',
'lipidData', or 'nmrData', usually created by \code{\link{as.pepData}},
\code{\link{as.proData}}, \code{\link{as.metabData}},
\code{\link{as.lipidData}}, or \code{\link{as.nmrData}}, respectively.}

\item{set_to}{logical indicating what to set check.names attribute to.
Defaults to TRUE.}
}
\value{
omicsData object with updated check.names attribute
}
\description{
\emph{This function sets the check.names attribute of an omicsData object.}
\bold{This function has been deprecated in favor of handling checking names 
externally and will return an unmodified omicsData.}
}
