% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{set_filter}
\alias{set_filter}
\title{Sets/updates the filters attribute with a filter class object.}
\usage{
set_filter(type, threshold, filtered, method)
}
\arguments{
\item{type}{The type of filter, see \link{filter_objects}}

\item{threshold}{The threshold for filtering applied, if applicable.}

\item{filtered}{The samples/biomolecules that are filtered.}

\item{method}{A list of metadata about what method was used in applying the 
filter.}
}
\value{
A list containing all the elements in the filters attribute:
        (list all attributes)
}
\description{
This function will create a filter class object. The output will always have
the same elements but not all of them will be used for every filter type. This
object will be appended to the list in the filters attribute for an omicsData
object.
}
\keyword{internal}
