% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliVis.R
\name{set_increment}
\alias{set_increment}
\title{Sets y-axis increment for trellData labels plotting}
\usage{
set_increment(yvalues, ...)
}
\arguments{
\item{yvalues}{y-values for plotting}

\item{testtype}{consideration for different statistical tests}
}
\description{
Sets y-axis increment for trellData plotting. Used by plot_comp.
}
\author{
Rachel Richardson
}
