% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{set_isobaric_info}
\alias{set_isobaric_info}
\title{Sets/updates the values in the isobaric_info attribute.}
\usage{
set_isobaric_info(
  exp_cname,
  channel_cname,
  refpool_channel,
  refpool_cname,
  refpool_notation,
  norm_info,
  isobaric_norm
)
}
\arguments{
\item{exp_cname, channel_cname, refpool_channel, refpool_cname, refpool_notation, norm_info, isobaric_norm}{See \code{\link{normalize_isobaric}}}
}
\value{
A list containing all the elements in the isobaric_info attribute:
        (list all attributes)
}
\description{
Sets/updates the values in the isobaric_info attribute.
}
\keyword{internal}
