% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{set_meta_info}
\alias{set_meta_info}
\title{Sets/updates the values in the meta_info attribute}
\usage{
set_meta_info(e_meta, emeta_cname)
}
\arguments{
\item{e_meta, emeta_cname}{See the \code{as.omicsData} functions, e.g. 
\code{\link{as.lipidData}}}
}
\value{
A list containing all the elements in the meta_info attribute:
       (list all attributes)
}
\description{
Sets/updates the values in the meta_info attribute
}
\keyword{internal}
