% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{set_nmr_info}
\alias{set_nmr_info}
\title{Sets/updates the values in the nmr_info attribute.}
\usage{
set_nmr_info(
  metabolite_name,
  sample_property_cname,
  norm_info,
  nmr_norm,
  backtransform
)
}
\arguments{
\item{metabolite_name}{}

\item{sample_property_cname}{}

\item{norm_info}{}

\item{nmr_norm}{}

\item{backtransform}{}
}
\value{
A list containing all the elements in the nmr_info attribute:
        (list all attributes)
}
\description{
Sets/updates the values in the nmr_info attribute.
}
