% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.omicsPlotter.R
\name{set_ylimits}
\alias{set_ylimits}
\title{Sets y-axis limits for Omicsplotter plotting}
\usage{
set_ylimits(yvalues, increment)
}
\arguments{
\item{yvalues}{y-values for plotting}

\item{increment}{An increment set based on the maximum/minimum y-values}
}
\description{
Sets y-axis limits for Omicsplotter plotting. Used by plot_comp.
}
\author{
Rachel Richardson
}
