% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{statres_barplot}
\alias{statres_barplot}
\title{Fold change barplots for statres objects}
\usage{
statres_barplot(x, stacked = FALSE, fc_colors = c("red", "black", "green"))
}
\arguments{
\item{x, stacked, fc_colors}{passed from \code{\link[pmartR:plot.statRes]{pmartR::plot.statRes()}}}
}
\value{
`ggplot` object - barplot.
}
\description{
Plots a bar-chart with bar heights indicating the number of significant
biomolecules, grouped by test type and fold change direction.
}
\keyword{internal}
