% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statRes_class.R
\name{statRes_output}
\alias{statRes_output}
\title{Function to take raw output of `imd_anova` and create output for `statRes` object}
\usage{
statRes_output(
  imd_anova_out,
  omicsData,
  comparisons,
  test_method,
  pval_adjust_a,
  pval_adjust_g,
  pval_thresh
)
}
\arguments{
\item{imd_anova_out}{A \code{ata.frame} containing the results of the \code{imd_anova} call.}

\item{omicsData}{A pmartR data object of any class, which has a `group_df` attribute that is usually created by the `group_designation()` function}

\item{comparisons}{The comparisons made.}

\item{test_method}{The test method used ("anova", "gtest", or "combined").}

\item{pval_adjust_a}{The p-value adjustment method for ANOVA.}

\item{pval_adjust_g}{The p-value adjustment method for G-test.}

\item{pval_thresh}{The p-value threshold value.}
}
\value{
the final object of class statRes
}
\description{
Function to take raw output of `imd_anova` and create output for `statRes` object
}
