% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_isobaricnormRes.R
\name{summary-pmartR}
\alias{summary-pmartR}
\alias{summary.isobaricnormRes}
\title{Summarizes an object of class isobaricnormRes}
\usage{
\method{summary}{isobaricnormRes}(isobaricnormRes_object)
}
\arguments{
\item{an}{object of type isobaricnormRes, created by \code{\link{normalize_isobaric}}}
}
\value{
data.frame object
}
\description{
For creating a summary of an S3 object of type 'isobaricnormRes':
}
\examples{
dontrun{
library(pmartRdata)
data(isobaric_object)
result = normalize_isobaric(isobaric_object, apply_norm = F)

summary(result)
}

}
