% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_nmrnormRes.R
\name{summary-pmartR}
\alias{summary-pmartR}
\alias{summary.nmrnormRes}
\title{Summarizes an object of class nmrnormRes}
\usage{
\method{summary}{nmrnormRes}(nmrnormRes_object)
}
\arguments{
\item{nmrnormRes_object}{an object of type nmrnormRes, created by \code{\link{normalize_nmr}}}
}
\value{
data.frame object
}
\description{
For creating a summary of an S3 object of type 'nmrnormRes':
}
\examples{
dontrun{
library(pmartRdata)
data(nmr_object_identified)

nmr_object = edata_transform(nmr_object_identified, "log2")
nmr_norm = normalize_nmr(nmr_object, apply_norm = FALSE, metabolite_name = "unkm1.53")
summary(nmr_norm)

# alternate specification: #
data(nmr_object_identified)

nmr_object = edata_transform(nmr_object, "log2")
nmr_norm = normalize_nmr(nmr_object, apply_norm = FALSE, sample_property_cname = "Concentration")
summary(nmr_norm)

}

}
