% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary.R
\name{summary-pmartR-filter}
\alias{summary-pmartR-filter}
\alias{summary.moleculeFilt}
\alias{summary.proteomicsFilt}
\alias{summary.imdanovaFilt}
\alias{summary.rmdFilt}
\alias{summary.cvFilt}
\alias{summary.customFilt}
\title{Produce a basic summary of a pmartR filter S3 Object}
\usage{
\method{summary}{moleculeFilt}(filter_object, min_num = NULL)

\method{summary}{proteomicsFilt}(filter_object, min_num_peps = NULL,
  degen_peps = FALSE)

\method{summary}{imdanovaFilt}(filter_object, min_nonmiss_anova = NULL,
  min_nonmiss_gtest = NULL)

\method{summary}{rmdFilt}(filter_object, pvalue_threshold = NULL)

\method{summary}{cvFilt}(filter_object, cv_threshold = NULL)

\method{summary}{customFilt}(filter_object)
}
\arguments{
\item{filter_object}{S3 object of class 'moleculeFilt', 'proteomicsFilt', 'imdanovaFilt', 'rmdFilt', 'cvFilt', or 'customFilt', created by \code{\link{molecule_filter}}, \code{\link{proteomics_filter}}, \code{\link{imdanova_filter}}, \code{\link{rmd_filter}}, \code{\link{cv_filter}}, or \code{\link{custom_filter}} respectively.}

\item{min_num}{an integer value specifying the minimum number of times each feature must be observed across all samples. Default value is 2.}

\item{min_num_peps}{an optional integer value between 1 and the maximum number of peptides that map to a protein in the data. The value specifies the minimum number of peptides that must map to a protein. Any protein with less than \code{min_num_peps} mapping to it will be returned as a protein that should be filtered. Default value is NULL.}

\item{degen_peps}{logical indicator of whether to filter out degenerate peptides (TRUE) or not (FALSE). Default value is FALSE.}

\item{min_nonmiss_anova}{the minimum number of non-missing feature values allowed per group for \code{anova_filter}. Suggested value is 2.}

\item{min_nonmiss_gtest}{the minimum number of non-missing feature values allowed per group for \code{gtest_filter}. Suggested value is 3.}

\item{pvalue_threshold}{A threshold for the Robust Mahalanobis Distance (RMD) p-value. All samples below the threshold will be filtered out. Default value is NULL.}

\item{cv_threshold}{numeric value greater than 1 and less than the value given by filter_object$CV_pooled.  CV values above cv_threshold are filtered out. Default value is NULL.}
}
\value{
a summary table for the pmartR filter object. If assigned to a variable, the elements of the summary table are saved in a list format.
}
\description{
This function will provide basic summary statistics for filter objects from the pmartR package.
}
\examples{
dontrun{
library(pmartRdata)
data("pep_object")
myfilter <- molecule_filter(pep_object)
summary(myfilter)
summary(myfilter, min_num = 2)
}

}
\author{
Lisa Bramer, Kelly Stratton, Thomas Johansen
}
