% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_pmartR.R
\name{summary-pmartR}
\alias{summary-pmartR}
\alias{summary.pepData}
\alias{summary.proData}
\alias{summary.lipidData}
\alias{summary.metabData}
\alias{summary.corRes}
\alias{summary.dimRes}
\alias{summary.moleculeFilt}
\alias{summary.proteomicsFilt}
\alias{summary.imdanovaFilt}
\alias{summary.rmdFilt}
\alias{summary.cvFilt}
\alias{summary.customFilt}
\alias{summary.normRes}
\alias{summary.SPANSRes}
\title{Produce a basic summary of an pmartR S3 Object}
\usage{
\method{summary}{pepData}(omicsData)

\method{summary}{proData}(omicsData)

\method{summary}{lipidData}(omicsData)

\method{summary}{metabData}(omicsData)

\method{summary}{corRes}(corRes_object)

\method{summary}{dimRes}(dimRes_object)

\method{summary}{moleculeFilt}(filter_object, min_num = NULL)

\method{summary}{proteomicsFilt}(filter_object, min_num_peps = NULL,
  degen_peps = FALSE)

\method{summary}{imdanovaFilt}(filter_object, min_nonmiss_anova = NULL,
  min_nonmiss_gtest = NULL)

\method{summary}{rmdFilt}(filter_object, pvalue_threshold = NULL)

\method{summary}{cvFilt}(filter_object, cv_threshold = NULL)

\method{summary}{customFilt}(filter_object)

\method{summary}{normRes}(omicsNorm)

\method{summary}{SPANSRes}(SPANSRes_obj)
}
\arguments{
\item{omicsData}{an object of the class 'lipidData', 'metabData', 'pepData', or 'proData' usually created by \code{\link{as.lipidData}}, \code{\link{as.metabData}}, \code{\link{as.pepData}}, or \code{\link{as.proData}}, respectively.}

\item{corRes_object}{an object of class corRes. A correlation matrix of all samples.}

\item{dimRes_object}{an object of class dimRes. A list containing sample identifiers and the principle components scores.}

\item{filter_object}{S3 object of class 'moleculeFilt', 'proteomicsFilt', or 'imdanovaFilt' created by \code{\link{molecule_filter}}, \code{\link{proteomics_filter}}, or \code{\link{imdanova_filter}}, respectively.}

\item{min_num}{an integer value specifying the minimum number of times each feature must be observed across all samples. Default value is 2.}

\item{min_num_peps}{an optional integer value between 1 and the maximum number of peptides that map to a protein in the data. The value specifies the minimum number of peptides that must map to a protein. Any protein with less than \code{min_num_peps} mapping to it will be returned as a protein that should be filtered. Default value is NULL.}

\item{degen_peps}{logical indicator of whether to filter out degenerate peptides (TRUE) or not (FALSE). Default value is FALSE.}

\item{min_nonmiss_anova}{the minimum number of non-missing feature values allowed per group for \code{anova_filter}. Suggested value is 2.}

\item{min_nonmiss_gtest}{the minimum number of non-missing feature values allowed per group for \code{gtest_filter}. Suggested value is 3.}

\item{pvalue_threshold}{A threshold for the Robust Mahalanobis Distance (RMD) p-value. All samples below the threshold will be filtered out. Default value is NULL.}

\item{cv_threshold}{CV values above cv_threshold are filtered out. Default value is NULL.}

\item{SPANSres_obj}{an object of class SPANSRes created by calling spans_procedure() on a grouped pepData or proData object.}
}
\value{
a summary table for the MSomicsR object. If assigned to a variable, the elements of the summary table are saved in a list format.
}
\description{
This function will provide basic summary statistics for objects from the MSomicsR package.
}
\examples{
dontrun{
library(pmartRdata)
data("pep_object")
summary(pep_object)
to_filter <- molecule_filter(pep_object)
summary(to_filter)
summary(to_filter, min_num = 4)
}

}
\author{
Lisa Bramer, Kelly Stratton, Thomas Johansen
}
