% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_trelliData.R
\name{summary-trelliData}
\alias{summary-trelliData}
\alias{summary.trelliData}
\title{Summarizes potential plotting options for a trelliData object}
\usage{
\method{summary}{trelliData}(object, ...)
}
\arguments{
\item{trelliData}{An object from the as.trelliData.edata or as.trelliData
functions}
}
\description{
Summarizes potential plotting options for a trelliData object
}
\examples{
\dontrun{

library(dplyr)

# Use an edata example. Build with as.trelliData.edata.
summary(trelliData)
summary(trelliData \%>\% trelli_panel_by("Lipid"))
summary(trelliData \%>\% trelli_panel_by("Sample"))

# Use an omicsData example. Build with as.trelliData.
summary(trelliData2)

# Use a statRes example. Build with as.trelliData.
summary(trelliData3)
}

}
