% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imd_anova.R
\name{take_diff}
\alias{take_diff}
\title{Compute pairwise differences}
\usage{
take_diff(omicsData)
}
\arguments{
\item{omicsData}{Any one of the omicsData objects (pepData, metabData, ...).}
}
\value{
A data.frame containing the differences between paired samples.
}
\description{
Computes the differences for paired data according to the information in the
pairing column of f_data. This variable name is also an attribute of the
group_DF attribute.
}
\author{
Evan A Martin
}
