% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_norm_object.R
\name{test_normRes}
\alias{test_normRes}
\title{Test the location and scale parameters from a normalization procedure}
\usage{
test_normRes(norm_object, test_fn = "kw", group_vec = NULL)
}
\arguments{
\item{norm_object}{An object of class 'pepData', 'proData', 'lipidData', 'metabData', 'isobaricpepData' that has had \code{normalize_global()} run on it, or a 'normRes' object}

\item{test_fn}{The statistical test to use, current tests are "anova" and "kw" for a Kruskal-Wallis test.}

\item{group_vec}{character vector of the same length as the number of normalization parameters (number of samples)}
}
\value{
A list with 2 entries containing the p_value of the test performed on the location and scale(if they exist) parameters.
}
\description{
Computes p-values from a test of dependence between normalization parameters and group assignment of a normalized omicsData or normRes object.
}
