% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliPlots.R
\name{trelli_abundance_boxplot}
\alias{trelli_abundance_boxplot}
\title{Boxplot trelliscope building function for abundance data}
\usage{
trelli_abundance_boxplot(
  trelliData,
  cognostics = c("count", "mean abundance"),
  ggplot_params = NULL,
  interactive = FALSE,
  include_points = TRUE,
  path = .getDownloadsFolder(),
  name = "Trelliscope",
  test_mode = FALSE,
  test_example = 1,
  single_plot = FALSE,
  ...
)
}
\arguments{
\item{trelliData}{A trelliscope data object made by as.trelliData or
as.trelliData.edata, and grouped by trelli_panel_by. Required.}

\item{cognostics}{A vector of cognostic options for each plot. Valid entries
are "count", "mean abundance", "median abundance", and "cv abundance". 
If data are paneled by a biomolecule, the count will be "sample count".
If data are paneled by a sample or a biomolecule class, the count will be "biomolecule count". 
If statRes data is included, "anova p-value" and "fold change" data per comparisons
may be added. If grouping information is included, only "sample count" and 
"mean abundance" will be calclulated, along with "anova p-value" and "fold change"
if specified. "anova p-value" will not be included if paneling a trelliscope
display by a biomolecule class. Default is "sample count" and "mean abundance".}

\item{ggplot_params}{An optional vector of strings of ggplot parameters to
the backend ggplot function. For example, c("ylab('')", "ylim(c(2,20))").
Default is NULL.}

\item{interactive}{A logical argument indicating whether the plots should be
interactive or not. Interactive plots are ggplots piped to ggplotly (for
now). Default is FALSE.}

\item{include_points}{Add points as a geom_jitter. Default is TRUE.}

\item{path}{The base directory of the trelliscope application. Default is
Downloads.}

\item{name}{The name of the display. Default is Trelliscope.}

\item{test_mode}{A logical to return a smaller trelliscope to confirm plot
and design. Default is FALSE.}

\item{test_example}{A vector of plot indices to return for test_mode. Default
is 1.}

\item{single_plot}{A TRUE/FALSE to indicate whether 1 plot (not a
trelliscope) should be returned. Default is FALSE.}

\item{...}{Additional arguments to be passed on to the trelli builder}
}
\description{
Specify a boxplot design and cognostics for the abundance
  boxplot trelliscope. Each boxplot will have its own groups as specified by
  the first main effect in group_designation.
}
\examples{
\dontrun{

## Generate trelliData objects using the as.trelliData.edata example code.

# Build the abundance boxplot with an edata file where each panel is a biomolecule. 
trelli_panel_by(trelliData = trelliData1, panel = "Peptide") \%>\% 
   trelli_abundance_boxplot(test_mode = TRUE, test_example = 1:10)
   
# Build the abundance boxplot wher each panel is a sample.
# Include all applicable cognostics. Remove points. 
trelli_panel_by(trelliData = trelliData1, panel = "Sample") \%>\% 
   trelli_abundance_boxplot(test_mode = TRUE, test_example = 1:10, 
                            include_points = FALSE,
                            cognostics = c("count", 
                                           "mean abundance", 
                                           "median abundance", 
                                           "cv abundance")
                           )

# Build the abundance boxplot with an omicsData object.
# Let the panels be biomolecules. Here, grouping information is included.
trelli_panel_by(trelliData = trelliData2, panel = "Peptide") \%>\% 
   trelli_abundance_boxplot(test_mode = TRUE, test_example = 1:10)
   
# Build the abundance boxplot with an omicsData object. The panel is a biomolecule class,
# which is proteins in this case.
trelli_panel_by(trelliData = trelliData2, panel = "RazorProtein") \%>\% 
   trelli_abundance_boxplot(test_mode = TRUE, test_example = 1:10)
    
# Build the abundance boxplot with an omicsData and statRes object.
# Panel by a biomolecule, and add statistics data to the cognostics
trelli_panel_by(trelliData = trelliData4, panel = "Peptide") \%>\%
   trelli_abundance_boxplot(test_mode = TRUE, test_example = 1:10,
                            cognostics = c("mean abundance", "anova p-value", "fold change"))
 
# Other options include modifying the ggplot  
trelli_panel_by(trelliData = trelliData1, panel = "Peptide") \%>\% 
   trelli_abundance_boxplot(test_mode = TRUE, test_example = 1:10, 
     ggplot_params = c("ylab('')", "ylim(c(20,30))"))

# Or making the plot interactive 
trelli_panel_by(trelliData = trelliData4, panel = "RazorProtein") \%>\% 
    trelli_abundance_boxplot(interactive = TRUE, test_mode = TRUE, test_example = 1:10)

}

}
\author{
David Degnan, Lisa Bramer
}
