% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliPlots.R
\name{trelli_abundance_heatmap}
\alias{trelli_abundance_heatmap}
\title{Heatmap trelliscope building function for abundance data}
\usage{
trelli_abundance_heatmap(
  trelliData,
  cognostics = c("sample count", "mean abundance", "biomolecule count"),
  ggplot_params = NULL,
  interactive = FALSE,
  path = .getDownloadsFolder(),
  name = "Trelliscope",
  test_mode = FALSE,
  test_example = 1,
  single_plot = FALSE,
  ...
)
}
\arguments{
\item{trelliData}{A trelliscope data object made by as.trelliData, and
grouped by an emeta variable. Required.}

\item{cognostics}{A vector of cognostic options. Defaults are "sample count", 
"mean abundance" and "biomolecule count". "sample count" and "mean abundance"
are reported per group, and "biomolecule count" is the total number of biomolecules
in the biomolecule class (e_meta column).}

\item{ggplot_params}{An optional vector of strings of ggplot parameters to
the backend ggplot function. For example, c("ylab('')", "xlab('')").
Default is NULL.}

\item{interactive}{A logical argument indicating whether the plots should be
interactive or not. Interactive plots are ggplots piped to ggplotly (for
now). Default is FALSE.}

\item{path}{The base directory of the trelliscope application. Default is
Downloads.}

\item{name}{The name of the display. Default is Trelliscope}

\item{test_mode}{A logical to return a smaller trelliscope to confirm plot
and design. Default is FALSE.}

\item{test_example}{A vector of plot indices to return for test_mode. Default
is 1.}

\item{single_plot}{A TRUE/FALSE to indicate whether 1 plot (not a
trelliscope) should be returned. Default is FALSE.}
}
\description{
Specify a plot design and cognostics for the abundance heatmap
  trelliscope. Data must be grouped by an e_meta column. Main_effects order
  the y-variables. All statRes data is ignored.
}
\examples{
\dontrun{

## Build the abundance heatmap with an omicsData object with emeta variables. Generate trelliData in as.trelliData.
trelli_panel_by(trelliData = trelliData2, panel = "LipidFamily") \%>\% 
   trelli_abundance_heatmap(test_mode = TRUE, test_example = 1:3)
   
## Users can modify the plotting function with ggplot parameters and interactivity, 
## and can also select certain cognostics.     
trelli_panel_by(trelliData = trelliData4, panel = "LipidFamily") \%>\% 
   trelli_abundance_heatmap(test_mode = TRUE, test_example = 1:5, 
     ggplot_params = c("ylab('')", "xlab('')"), interactive = TRUE, cognostics = c("mean", "median"))  
   
}


}
\author{
David Degnan, Lisa Bramer
}
