% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliVis.R
\name{trelliVis}
\alias{trelliVis}
\title{Plot pairwise comparisons and data values in trellData object}
\usage{
trelliVis(...)
}
\arguments{
\item{omicsData}{A pmartR object of class pepData, lipidData, metabData, or proData. Can use list(pepData, proData) for associated data.}

\item{omicsStats}{A statistical results object produced by running \code{imd_anova} on omicsData. Can use list(pepStats, proStats) for associated data.}

\item{omicsFormat}{Output of as.trellData() function}

\item{p_val}{Numeric that specifies p-value for significance calculations. Default is 0.05.}

\item{panel_variable}{String: Name of column that plot panels are sorted by (e.g. each plotting arrangement has a unique identifier from panel variable). Default is emeta_cname if present, edata_cname where emeta_cname is not present.}

\item{try_URL}{Will attempt to link to PubChem, LipidMaps, or Uniprot based on information in edata_cname of omicsData or specified mapping_col for peptide data. Default is FALSE.}

\item{trelli_name}{String: name of display, or list of names where a list is provided for omicsData and omicsStats}

\item{trelli_path_out}{String: path to where trelliscope is stored. Default is "./TrelliDisplay"}

\item{interactive}{Should the plots be rendered as plotly objects?}

\item{plot_text}{Disable plot text}

\item{y_limits}{Y limits}

\item{plot_type}{plots for plotting}

\item{self_contained}{Should display be generated in document? Defaults to FALSE}

\item{custom_cog}{The name of a user generated function with cognostics. Function should act on a subset of data and output a dataframe (or tibble or equivelent) with one row (summary of rows).}
}
\description{
Plot pairwise comparisons and data values of omicsData and omicsStats objects. Customizable for plot types, y axis limits, paneling variable (what overall group is plotted on each graph arrangement), as well as desired variables for the y and x axis.
}
\author{
Rachel Richardson
}
