% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliVis.R
\name{validate_omics_input}
\alias{validate_omics_input}
\title{Validate inputs for omicsData and omicsStats in as.trellData processing.}
\usage{
validate_omics_input(omicsData = NULL, omicsStats = NULL)
}
\arguments{
\item{omicsData}{A pmartR object of class pepData, lipidData, metabData, or proData}

\item{omicsStats}{A statistical results object produced by running \code{imd_anova} on omicsData.}
}
\description{
Checks for validity of omicsData and omicsStats inputs.
}
\seealso{
\link[pmartR]{as.trellData}
}
\author{
Rachel Richardson
}
