% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edata_replace.R
\name{vector_replace}
\alias{vector_replace}
\title{Replace x with y for a single vector}
\usage{
vector_replace(one_vector, x, y)
}
\arguments{
\item{one_vector}{numeric vector}

\item{x}{value to be replaced}

\item{y}{replacement value}
}
\value{
numeric vector
}
\description{
Replace x with y for a single vector
}
\author{
Kelly Stratton
}

