% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_data_info.R
\name{verify_data_info}
\alias{verify_data_info}
\title{Verifies some basic attributes of an omicsData object.  Intended use is in testing scripts and internally within some functions.}
\usage{
verify_data_info(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData', or 'lipidData', created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, or \code{\link{as.lipidData}}, respectively.}
}
\value{
TRUE if the object is consistent for certain parts of the object. FALSE otherwise.
}
\description{
Verifies some basic attributes of an omicsData object.  Intended use is in testing scripts and internally within some functions.
}
\details{
This function will invisibly return TRUE or FALSE depending on whether or not the objects contents are consistent with its attributes.  If FALSE, there will also be a print output of which values are inconsistent.
}
\examples{
dontrun{
library(pmartRdata)
data(pep_edata)
bool_val <- verify_data_info(pep_edata)
}

}
