% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{wrap_message}
\alias{wrap_message}
\title{Custom message functions to pretty-print text with newlines so you can follow
character limit guidelines in source code.}
\usage{
wrap_message(..., prefix = " ", initial = "")
}
\description{
Custom message functions to pretty-print text with newlines so you can follow
character limit guidelines in source code.
}
