% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stat_results.R
\name{write_stat_results}
\alias{write_stat_results}
\title{Creates a list of three sheets, Normalized Data, DA Test Results, and Only DA biomolecules - for OMICS project}
\usage{
write_stat_results(omicsData, statResData, refCondition, filePath = NULL)
}
\arguments{
\item{omicsData}{an object of one of the classes "pepData", "proData", "metabData", or "lipidData", usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, or \code{\link{lipidData}}, respectively.}

\item{statResData}{an object of class statRes, created by \code{\link{imd_anova}}}

\item{refCondition}{character string, what the reference condition looks like (e.g. "MOCK", "Mock", "mock", etc.)}

\item{filePath}{default is NULL, if NULL then workbook will not be written out, if it is a character string specifying the file path, file name and .xlsx extension}
}
\value{
a list of three data frames Normalized_Data, DA_Test_Results, and DA_"data type"_Only
}
\description{
Creates a list of three sheets, Normalized Data, DA Test Results, and Only DA biomolecules - for OMICS project
}
\examples{
dontrun{
library(pmartR)
library(pmartRdata)

myproData <- group_designation(omicsData = pro_object, main_effects = c("Condition"))

imdanova_Filt <- imdanova_filter(omicsData = myproData)
myproData <- applyFilt(filter_object = imdanova_Filt, omicsData = myproData, min_nonmiss_anova=2)

imd_anova_res <- imd_anova(omicsData = myproData, test_method = 'comb', pval_adjust='bon')

result = write_stat_results(omicsData = myproData, statResData = imd_anova_res, refCondition = "Mock")
}

}
\author{
Natalie Heller
}
