% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stat_results_omics.R
\name{write_stat_results_omics}
\alias{write_stat_results_omics}
\title{Creates a list of three sheets, Normalized Data, DA Test Results, and Only DA biomolecules - for OMICS project}
\usage{
write_stat_results_omics(omicsData, statResData, refCondition, filePath = NULL)
}
\arguments{
\item{omicsData}{an object of one of the classes "pepData", "proData",
"metabData", or "lipidData", usually created by \code{\link{as.pepData}},
\code{\link{as.proData}}, \code{\link{as.metabData}}, or
\code{\link{lipidData}}, respectively.}

\item{statResData}{an object of class statRes, created by
\code{\link{imd_anova}}}

\item{refCondition}{character string, what the reference condition looks like
(e.g. "MOCK", "Mock", "mock", etc.)}

\item{filePath}{default is NULL, if NULL then workbook will not be written
out, if it is a character string specifying the file path, file name and
.xlsx extension}
}
\value{
a list of three data frames Normalized_Data, DA_Test_Results, and
  DA_"data type"_Only
}
\description{
Creates a list of three sheets, Normalized Data, DA Test Results, and Only DA biomolecules - for OMICS project
}
\author{
Natalie Heller
}
