% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_funcs.R
\name{zero_one_scale}
\alias{zero_one_scale}
\title{Zero to One scaling}
\usage{
zero_one_scale(
  e_data,
  edata_id,
  subset_fn,
  feature_subset,
  backtransform = FALSE,
  apply_norm = FALSE,
  check.names = NULL
)
}
\arguments{
\item{e_data}{e_data a \eqn{p \times n + 1} data.frame, where \eqn{p} is the
number of peptides, lipids, or metabolites and \eqn{n} is the number of
samples. Each row corresponds to data for a peptide, protein, lipid, or
metabolite, with one column giving the biomolecule identifier name.}

\item{edata_id}{character string indicating the name of the peptide, protein,
lipid, or metabolite identifier. Usually obtained by calling
\code{attr(omicsData, "cnames")$edata_cname}.}

\item{subset_fn}{character string indicating the subset function to use for
normalization.}

\item{feature_subset}{character vector containing the feature names in the
subset to be used for normalization}

\item{backtransform}{logical argument. Not utilized in this scaling, but 
included for code compatibility.}

\item{apply_norm}{logical argument. If TRUE, the normalization will be
applied to the data. Defaults to FALSE.}
}
\value{
List containing two elements: \code{norm_params} is list with two
  elements:
\tabular{ll}{
scale \tab Range of each sample used in scaling \cr
\tab \cr
location \tab NULL
\cr
}

\code{backtransform_params} is a list with two elements:
\tabular{ll}{
scale \tab NULL \cr
\tab \cr
location \tab NULL
\cr
}

If \code{apply_norm} is TRUE, the transformed data is returned as a third
list item.
}
\description{
Re-scales the data to be between 0 and 1
}
\details{
The sample-wise mean of the feature subset specified for
  normalization is subtracted from each feature in e_data to get the
  normalized data. The location estimates are the sample-wise means of the
  subset data. There are no scale estimates for mean centering, though the
  function returns a NULL list element as a placeholdfer for a scale
  estimate. If backtransform is TRUE, the global median of the subset data
  (across all samples) is added back to the normalized values. Medians are
  taken ignoring any NA values.
}
\author{
Lisa Bramer, Kelly Stratton, Rachel Richardson
}
