% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_funcs.R
\name{zscore_transform}
\alias{zscore_transform}
\title{Z-Score Transformation}
\usage{
zscore_transform(e_data, edata_id, feature_subset, backtransform = FALSE,
  apply_norm = FALSE, check.names = TRUE)
}
\arguments{
\item{e_data}{a \eqn{p \times n + 1} data.frame, where \eqn{p} is the number of peptides, lipids, or metabolites and \eqn{n} is the number of samples. Each row corresponds to data for a peptide, protein, lipid, or metabolite, with a column giving the identifer name.}

\item{edata_id}{character string indicating the name of the peptide, protein, lipid, or metabolite identifier. Usually obtained by calling \code{attr(omicsData, "cnames")$edata_cname}.}

\item{feature_subset}{character vector containing the feature names in the subset to be used for normalization}

\item{backtransform}{logical argument. If TRUE, the data will be back transformed after normalization so that the values are on a scale similar to their raw values. See details for more information. Defaults to FALSE.}

\item{apply_norm}{logical argument. If TRUE, the normalization will be applied to the data. Defaults to FALSE.}
}
\value{
List containing two elements: \code{norm_params} is list with two elements:
\tabular{ll}{
scale \tab numeric vector of length \code{n} standard deviations for each sample
\tab \cr
location \tab numeric vector of length \code{n} means for each sample
\cr
}

\code{backtransform_params} is a list with two elements:
\tabular{ll}{
scale \tab numeric value giving the pooled standard deviation across all samples
\tab \cr
location \tab numeric value giving global mean across all samples
\cr
}
If \code{backtransform} is set to TRUE then each list item under \code{backtransform_params} will be NULL.

If \code{apply_norm} is TRUE, the transformed data is returned as a third list item.
}
\description{
Normalizes the data via z-score transformation using the feature subset specified for normalization
}
\details{
Each feature is scaled by subtracting the mean of the feature subset specified for normalization and then dividing the result by the standard deviation (SD) of the feature subset specified for normalization to get the normalized data.
The location estimates are the subset means for each sample. The scale estimates are the subset SDs for each sample. If backtransform is TRUE, the normalized feature values are multiplied by a pooled standard deviation (estimated across all samples) and a global mean of the subset data (across all samples) is added back to the normalized values. Means are taken ignoring any NA values.
}
\examples{
\dontrun{
library(pmartRdata)
data(pep_edata)
peps_subset <- pep_edata$Mass_Tag_ID
peps_zscore <- zscore_transform(e_data = pep_edata, edata_id = "Mass_Tag_ID", feature_subset = peps_subset)
}

}
\author{
Lisa Bramer, Kelly Stratton, Bryan Stanfill
}
