% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliVis.R
\name{add_plot_features}
\alias{add_plot_features}
\title{Adds features to dataframe for plotting}
\usage{
add_plot_features(trellData, nestedData, p_val = NULL,
  panel_variable = NULL, y_axis)
}
\arguments{
\item{trellData}{An object of class "trellData" generated from \code{\link{as.trellData}}.}

\item{nestedData}{Nested trellData dataframe}

\item{p_val}{Specifies p-value for setting graph border colors}

\item{panel_variable}{Specifies what to divide trelliscope panels by, must be a column in trellData.}

\item{y_axis}{Specifies what column should be on the y-axis, must be a column in trellData and numeric.}
}
\description{
Adds border color, hover, and label text as applicable to nested trellData dataframe for plotting
}
\seealso{
\link[pmartR]{format_plot}
}
\author{
Rachel Richardson
}
