% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_log.R
\name{analysis_log}
\alias{analysis_log}
\title{Creates R markdown document report}
\usage{
analysis_log(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData', 'lipidData', or 'nmrData', usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, or \code{\link{as.lipidData}}, respectively.}
}
\description{
Creates R markdown document report
}
\examples{
\dontshow{if (requireNamespace("pmartRdata", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
library(pmartRdata)
data(metab_object)
result = analysis_log(metab_object)
}
\dontshow{\}) # examplesIf}
}
