% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_result.R
\name{cor_result}
\alias{cor_result}
\title{Correlation matrix of biomolecule data}
\usage{
cor_result(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData',
'lipidData', or 'nmrData' created by \code{\link{as.pepData}},
\code{\link{as.proData}}, \code{\link{as.metabData}},
\code{\link{as.lipidData}}, \code{\link{as.nmrData}}, respectively.}
}
\value{
An \eqn{n \times n} matrix giving the correlation between samples.
}
\description{
This function returns an object of class corRes (correlation Result)
}
\details{
The Pearson correlation between samples is calculated based on
  biomolecules that are observed in both samples.  the \eqn{n \times n}
  correlation matrix of normalized data. See \code{\link{cor}} for further
  details.
}
\examples{
\dontrun{
library(pmartR)

data(pep_object)

my_correlation <- cor_result(omicsData = pep_object)
}

}
\author{
Kelly Stratton, Lisa Bramer
}
