% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_filter.R
\name{custom_filter}
\alias{custom_filter}
\title{Custom Filter}
\usage{
custom_filter(omicsData, e_data_remove = NULL, f_data_remove = NULL,
  e_meta_remove = NULL, e_data_keep = NULL, f_data_keep = NULL,
  e_meta_keep = NULL)
}
\arguments{
\item{omicsData}{an object of class "pepData", "proData", "metabData", or "lipidData", created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, or \code{\link{as.lipidData}}, respectively.}

\item{e_data_remove}{character vector specifying the names of the e_data identifiers to remove from the data. This argument can only be specified with other 'remove' arguments.}

\item{f_data_remove}{character vector specifying the names of f_data identifiers to remove from the data. This argument can only be specified with other 'remove' arguments.}

\item{e_meta_remove}{character vector specifying the names of the e_meta identifiers to remove from the data. This argument can only be specified with other 'remove' arguments.}

\item{e_data_keep}{character vector specifying the names of the e_data identifiers to keep from the data. This argument can only be specified with other 'keep' arguments.}

\item{f_data_keep}{character vector specifying the names of f_data identifiers to keep from the data. This argument can only be specified with other 'keep' arguments.}

\item{e_meta_keep}{character vector specifying the names of the e_meta identifiers to keep from the data. This argument can only be specified with other 'keep' arguments.}
}
\value{
An S3 object of class 'customFilt', which is a list with 3 elements: e_data_remove, f_data_remove, and e_meta_remove.
}
\description{
This function creates a customFilt S3 object based on user-specified items to filter out of the dataset
}
\examples{
dontrun{
library(pmartRdata)
data("metab_object")
to_filter <- custom_filter(metab_object, e_data_remove = "fumaric acid", f_data_remove = "Infection1")
summary(to_filter)
to_filter2 <- custom_filter(metab_object, e_data_remove = "fumaric acid")
summary(to_filter2)
}

}
\author{
Kelly Stratton
}
