% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliVis.R
\name{custom_fn_cog}
\alias{custom_fn_cog}
\title{Test function for data_cog(), custom_cog argument}
\usage{
custom_fn_cog(data)
}
\arguments{
\item{data}{data subset from a trellData object.}
}
\description{
Test function for data_cog(), custom_cog argument. Returns a tibble with one row and 3 columns.
}
\examples{

dontrun{
library(pmartRdata)
library(pmartR)
library(ggplot2)
data("metab_object")
mymetabData <- pmartR::edata_transform(metab_object, "log10")
mymetabData <- pmartR::group_designation(metab_object, "Condition")
mymetabData <- pmartR::normalize_global(metab_object, "all", "median", apply_norm = TRUE, backtransform = TRUE)

custom_fn <- function(data){
ggplot2::ggplot() + ggplot2::geom_point(ggplot2::aes(x = 1:10, y = 1:10))
}

custom_fn_cog <- function(data){
tibble::tibble(test1 = 1, test2 = 2, test3 = 3)
}

format <- pmartR::as.trellData(mymetabData)
plot <- pmartR:::format_plot(format, custom_plot = "custom_fn")
pmartR:::data_cogs(plot, format, custom_cog = "custom_fn_cog")

}

}
\seealso{
\link[pmartR]{data_cogs}
}
\author{
Rachel Richardson
}
