% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_filter.R
\name{cv_filter}
\alias{cv_filter}
\title{Filter Based on Pooled Coefficient of Variation (CV) Values}
\usage{
cv_filter(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData', 'lipidData', or 'nmrData', created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, \code{\link{as.lipidData}}, \code{\link{as.nmrData}}, respectively. Note, if \code{\link{group_designation}} has not been run, the CV is calculated based on all samples for each biomolecule.}
}
\value{
An S3 object of class 'cvFilt' giving the pooled CV for each biomolecule and additional attributes used for plottinga data.frame with a column giving the peptide, protein, or gene name and a column giving the pooled CV value
}
\description{
A pooled CV is calculated for each biomolecule.
}
\details{
For each biomolecule, the CV of each group is calculated as the standard deviation divided by the mean, excluding missing values. A pooled CV estimate is then calculated based on the methods of Ahmed (1995).
}
\examples{
dontrun{
library(pmartRdata)
data("pep_object")
pep_object2 <- group_designation(omicsData = pep_object, main_effects = "Condition")
to_filter <- cv_filter(omicsData = pep_object2)
summary(to_filter, cv_threshold = 30)
}

}
\references{
Ahmed, S.E. (1995). \emph{A pooling methodology for coefficient of variation}. The Indian Journal of Statistics. 57: 57-75.
}
\author{
Lisa Bramer, Kelly Stratton
}
