% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.omicsPlotter.R
\name{data_cogs}
\alias{data_cogs}
\title{Plot pairwise comparisons and data values in trellData object}
\usage{
data_cogs(...)
}
\arguments{
\item{trellData}{An object of class "trellData" generated from \code{\link{format_data()}}.}

\item{nested_plot}{A nested table generated from trellData using formatplot()}

\item{omicsData}{A pmartR object of class pepData, lipidData, metabData, or proData}

\item{omicsStats}{A statistical results object produced by running \code{imd_anova} on omicsData.}

\item{p_val}{Numeric that specifies p-value for Boolean significance cognotic. Default is 0.05.}

\item{mapping_col}{String: For proData - name of column with peptide information. For pepData - name of column with protein information. Default is NULL.}

\item{panel_variable}{String: Name of column that plot panels are sorted by (e.g. each plotting arrangement has a unique identifier from panel variable). Default is emeta_cname if present, edata_cname where emeta_cname is not present.}

\item{try_URL}{Will attempt to link to PubChem, LipidMaps, or Uniprot based on information in edata_cname of omicsData or specified mapping_col for peptide data. Default is FALSE.}
}
\description{
Plot pairwise comparisons and data values in trellData object. Customizable for plot types, y axis limits, paneling variable (what overall group is plotted on each graph arrangement), as well as desired variables for the y and x axis.
}
\author{
Rachel Richardson
}
