% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqData_wrappers.R
\name{Deseq2_wrapper}
\alias{Deseq2_wrapper}
\title{Wrapper for DESeq2 workflow}
\usage{
Deseq2_wrapper(
  omicsData,
  test = "Wald",
  p_adjust = "BH",
  comparisons = NULL,
  p_cutoff = 0.05,
  ...
)
}
\arguments{
\item{omicsData}{an object of type 'seqData', created by \code{\link{as.seqData}}}

\item{test}{an object of type 'seqData', created by \code{\link{as.seqData}}}

\item{p_adjust}{an object of type 'seqData', created by \code{\link{as.seqData}}}

\item{comparisons}{an object of type 'seqData', created by \code{\link{as.seqData}}}

\item{...}{additional arguments passed to function}
}
\value{
statres data.frame object
}
\description{
For generating statistics for 'seqData' objects
}
\details{
Runs default DESeq workflow. Defaults to Wald test, no independent filtering, and 
running in parallel. Additional arguments can be passed for use in the function, 
refer to DESeq() and results() in DESeq2 package
}
\examples{
\dontrun{
}

}
