% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{dim_reduction}
\alias{dim_reduction}
\title{Reduce Dimension of Data for Exploratory Data Analysis}
\usage{
dim_reduction(omicsData, k = 2)
}
\arguments{
\item{omicsData}{an object of the class 'pepdata', 'prodata', 'metabData', or 'lipidData' usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, or \code{\link{as.lipidData}}, respectively.}

\item{k}{integer number of principal components to return. Defaults to 2.}
}
\value{
a data.frame with first \code{k} principal component scores, sample identifiers, and group membership for each sample (if group designation was previously run on the data).
}
\description{
Calculate principal components using projection pursuit estimation, which implements an expectation-maximization (EM) estimation algorithm when data is missing.
}
\details{
Any biomoleculs seen in only one sample or with a variance less than 1E-6 across all samples are not included in the PCA calculations. This function leverages code from \code{\link[pcaMethods]{pca}}.
}
\examples{
dontrun{
library(pmartRdata)
data(lipid_object)
lipid_object <- edata_transform(omicsData = lipid_object, data_scale="log2")
lipid_object <- group_designation(omicsData = lipid_object, main_effects = "Condition")
pca_lipids <- dim_reduction(omicsData = lipid_object)
plot(pca_lipids)
summary(pca_lipids)
}

}
