% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{dim_reduction}
\alias{dim_reduction}
\title{Reduce Dimension of Data for Exploratory Data Analysis}
\usage{
dim_reduction(omicsData, k)
}
\arguments{
\item{omicsData}{an object of the class 'pepdata', 'prodata', or 'genedata' usually created by \code{\link{as.pepdata}}, \code{\link{as.prodata}}, or \code{\link{as.genedata}}, respectively.}

\item{k}{number of principal components to return}
}
\value{
a data.frame with first \code{k} principal component scores, sample identifiers, and group membership for each sample (if group designation was previously run on the data).
}
\description{
Calculate principal components using projection pursuit estimation, which implements an expectation-maximization (EM) estimation algorithm when data is missing.
}
\details{
Any biomoleculs seen in only one sample or with a variance less than 1E-6 across all samples are not included in the PCA calculations. This function leverages code from \code{\link[pcaMethods]{pca}}.
}

