% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edata_replace.R
\name{edata_replace}
\alias{edata_replace}
\alias{edata_replace.lipidData}
\alias{edata_replace.metabData}
\alias{edata_replace.pepData}
\alias{edata_replace.proData}
\title{Replace Values Equal to x with y}
\usage{
edata_replace(omicsData, x, y, check.names = TRUE)

\method{edata_replace}{pepData}(omicsData, x, y, check.names = TRUE)

\method{edata_replace}{proData}(omicsData, x, y, check.names = TRUE)

\method{edata_replace}{metabData}(omicsData, x, y, check.names = TRUE)

\method{edata_replace}{lipidData}(omicsData, x, y, check.names = TRUE)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData', or 'lipidData', usually created by \code{pepData}, \code{proData}, \code{metabData}, or \code{lipidData}, respectively.}

\item{x}{value to be replaced, usually numeric or NA}

\item{y}{replacment value, usually numeric or NA}

\item{check.names}{logical. If TRUE then the names of the variables in the data frame are checked to ensure that they are syntactically valid variable names and are not duplicated. If necessary they are adjusted (by \link{\code{make.names}}) so that they are.}
}
\value{
data object of the same class as omicsData
}
\description{
This function finds all values of x in the e_data element of omicsData and replaces them with y
}
\details{
This function is often used to replace any 0 values in peptide, protein, metabolite, or lipid data with NA's.
}
\examples{
library(pmartRdata)
data(metab_metabData)
metab_metabData2 <- edata_replace(omicsData = metab_metabData, x=0, y=NA)

}
\author{
Kelly Stratton
}

