% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.multiData.R
\name{find_fmeta_cnames}
\alias{find_fmeta_cnames}
\title{Set column names in f_meta for each object.  May result in duplicates.}
\usage{
find_fmeta_cnames(res)
}
\arguments{
\item{res}{The output of \code{fmeta_matches}, A list, the i-th element of 
which contains a character vector of column name matches in f_meta for the 
i-th omicsData object.}
}
\value{
Character vector containing a column name in f_meta that the i-th 
object matches to
}
\description{
Set column names in f_meta for each object.  May result in duplicates.
}
\keyword{internal}
