% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comparisons.R, R/helper_fn.R
\name{get_comparisons}
\alias{get_comparisons}
\title{Returns data.frame with comparisons to be made}
\usage{
get_comparisons(compObj)

get_comparisons(compObj)
}
\arguments{
\item{compObj}{is an object with the comparison attribute; specifically objects of class 'statRes' and 'trellData' objects derived from 'statRes' objects in \code{\link{format_data}}}

\item{comp_type}{string taking one of the following values: control, pairwise, custom. Specifying "control" indicates that all other groups are to be compared to this single control group. Specifying "pairwise" indicates that all pairwise comparisons are to be made.}

\item{omicsData}{A pmartR data object of any class, which has a `group_df` attribute that is usually created by the `group_designation()` function}

\item{control_group}{string indicating the group to use for the control group. Only required when comp_type="control".}
}
\value{
data.frame with columns for Test and Control. Each row corresponds to a comparison of interest. 
#'

returns a data frame with comparisons and their indices
}
\description{
The method creates a data.frame containing the comparisons to be made using differential statistics.

This function returns comparisons from statRes or trellData object
}
\details{
This function takes in the omicsData and type of comparison, and returns a data.frame where each row corresponds to a comparison of interest.
}
\examples{
dontrun{
library(pmartR)
library(pmartRdata)

my_prodata = group_designation(omicsData = pro_object, main_effects = c("Condition"))

imdanova_Filt = imdanova_filter(omicsData = my_prodata)
my_prodata = applyFilt(filter_object = imdanova_Filt, omicsData = my_prodata, min_nonmiss_anova=2) 
imd_anova_res = imd_anova(omicsData = my_prodata, test_method = 'comb', pval_adjust='bon')

result = get_comparisons(imd_anova_res)
}
}
\author{
Kelly Stratton
}
