% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comparisons.R, R/helper_fn.R
\name{get_comparisons}
\alias{get_comparisons}
\alias{get_comparisons}
\title{Returns data.frame with comparisons to be made}
\usage{
get_comparisons(statRes)

get_comparisons(statRes)
}
\arguments{
\item{statRes}{is an object of class 'statRes'}

\item{comp_type}{string taking one of the following values: control, pairwise, custom. Specifying "control" indicates that all other groups are to be compared to this single control group. Specifying "pairwise" indicates that all pairwise comparisons are to be made.}

\item{omicsData}{A mintR data object of any class, which has a `group_df` attribute that is usually created by the `group_designation()` function}

\item{control_group}{string indicating the group to use for the control group. Only required when comp_type="control".}
}
\value{
data.frame with columns for Test and Control. Each row corresponds to a comparison of interest. 
#'

returns a data frame with comparisons and their indices
}
\description{
The method creates a data.frame containing the comparisons to be made using differential statistics.

This function returns comparisons from statRes object
}
\details{
This function takes in the omicsData and type of comparison, and returns a data.frame where each row corresponds to a comparison of interest.
}
\examples{
dontrun{
get_comparisons(statRes_object)
}
}
\author{
Kelly Stratton
}
