% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{get_data_norm}
\alias{get_data_norm}
\title{Fetch the normalization status of the data}
\usage{
get_data_norm(omicsObject)
}
\arguments{
\item{omicsObject}{an object of the class 'pepData', 'proData', 'metabData',
'lipidData', 'nmrData', 'statRes', or 'trelliData', usually created by
\code{\link{as.pepData}}, \code{\link{as.proData}},
\code{\link{as.metabData}}, \code{\link{as.lipidData}},
\code{\link{as.nmrData}}, \code{\link{imd_anova}}, or
\code{\link{as.trelliData}} respectively.}
}
\value{
A logical value indicating whether the data have been normalized.
}
\description{
This function returns the norm_info element of the data_info attribute
indicating whether the data have been normalized.
}
