% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{get_data_norm}
\alias{get_data_norm}
\title{Get data norm}
\usage{
get_data_norm(omicsObject)
}
\arguments{
\item{omicsObject}{an object of the class 'pepData', 'proData', 'metabData', 'lipidData', 'statRes', or 'trellData', usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, \code{\link{as.lipidData}}, \code{\link{imd_anova}}, or \code{\link{format_data}} respectively.}
}
\value{
a character string describing data norm
}
\description{
This function returns data norm attribute
}
\examples{
dontrun{
get_data_norm(omicsObject)
}
}
