% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{get_emeta_cname}
\alias{get_emeta_cname}
\title{Get e_meta cname}
\usage{
get_emeta_cname(omicsObject)
}
\arguments{
\item{omicsObject}{an object of the class 'pepData', 'proData', 'metabData', 'lipidData', 'nmrData', 'statRes', or 'trellData', usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, \code{\link{as.lipidData}}, \code{\link{as.nmrData}}, \code{\link{imd_anova}}, or \code{\link{format_data}} respectively.}
}
\value{
a character string describing e_meta cname
}
\description{
This function returns e_meta cname
}
\examples{
dontrun{
get_emeta_cname(omicsObject)
}
}
