% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{get_fdata_cname}
\alias{get_fdata_cname}
\title{Get f_data cname}
\usage{
get_fdata_cname(omicsObject)
}
\arguments{
\item{omicsObject}{an object of the class 'pepData', 'proData', 'metabData', 'lipidData', 'nmrData', 'statRes', or 'trellData', usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, \code{\link{as.lipidData}}, \code{\link{as.nmrData}}, \code{\link{imd_anova}}, or \code{\link{format_data}} respectively.}
}
\value{
a character string describing f_data cname
}
\description{
This function returns f_data cname
}
\examples{
dontrun{
get_fdata_cname(omicsObject)
}
}
