% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{get_group_info}
\alias{get_group_info}
\title{Get group info of omicsData object}
\usage{
get_group_info(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData', or 'lipidData', usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, or \code{\link{as.lipidData}}, respectively.}
}
\value{
a data.frame containing omicsData object group info
}
\description{
This function returns the "group_DF" attribute of an omicsData object
}
\examples{
dontrun{
get_group_info(omicsData)

}

}
