% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{get_group_table}
\alias{get_group_table}
\title{Get group table}
\usage{
get_group_table(omicsObject)
}
\arguments{
\item{omicsObject}{an object of the class 'pepData', 'proData', 'metabData',
'lipidData', 'nmrData', 'statRes', or 'trellData', usually created by
\code{\link{as.pepData}}, \code{\link{as.proData}},
\code{\link{as.metabData}}, \code{\link{as.lipidData}},
\code{\link{as.nmrData}}, \code{\link{imd_anova}}, or
\code{\link{format_data}} respectively.}
}
\value{
a table containing number of samples per group
}
\description{
This function returns a table with number of samples per group
}
