% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imd_anova.R
\name{get_pred_grid}
\alias{get_pred_grid}
\title{Build the prediction grid to compute least squares means.}
\usage{
get_pred_grid(xmatrix, groups, continuous_covar_inds = NULL)
}
\arguments{
\item{xmatrix}{The design matrix from which to construct the prediction grid}

\item{groups}{The group assignments corresponding to rows in the design matrix}

\item{continuous_covar_inds}{The column indices of xmatrix corresponding to continuous covariates.}
}
\value{
A matrix of the prediction grid
}
\description{
Build the prediction grid to compute least squares means.
}
\author{
Daniel Claborne
}
