% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{getchecknames}
\alias{getchecknames}
\title{Return check.names attribute of omicsData object}
\usage{
getchecknames(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData', 'lipidData', or 'nmrData' usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, \code{\link{as.lipidData}}, or \code{\link{as.nmrData}}, respectively.}
}
\value{
returns logical, either TRUE or FALSE
}
\description{
This function returns check.names attribute from omicsData object
}
\examples{
dontrun{
getchecknames(omicsData)
}
}
