% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_comparison.R
\name{group_comparison_anova}
\alias{group_comparison_anova}
\title{Group comparisons for the anova test}
\usage{
group_comparison_anova(groupData, comparisons, anova_results_full)
}
\arguments{
\item{groupData}{data frame that assigns sample names to groups}

\item{comparisons}{dataframe that defiens the comparsions of interest}

\item{anova_results}{results of the pmartR::anova_test() function}
}
\value{
A data.frame containing the p-values from the group comparisons.
}
\description{
Takes the results of anova_test() and returns group comparison p-values
}
\author{
Bryan Stanfill
}
