% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_designation.R
\name{group_designation}
\alias{group_designation}
\title{Creates Attribute of omicsData object containing Group Membership Based on
Specified Main Effects}
\usage{
group_designation(
  omicsData,
  main_effects = NULL,
  covariates = NULL,
  cov_type = NULL,
  pairs = NULL,
  time_course = NULL
)
}
\arguments{
\item{omicsData}{an object of the class 'lipidData', 'metabData', 'pepData',
'proData', 'isobaricpepData', or 'nmrData', usually created by
\code{\link{as.lipidData}}, \code{\link{as.metabData}},
\code{\link{as.pepData}}, \code{\link{as.proData}},
\code{\link{as.isobaricpepData}}, or \code{\link{as.nmrData}},
respectively.}

\item{main_effects}{a character vector with no more than two variable names
that should be used as main effects to determine group membership of
samples. The variable name must match a column name from \code{f_data}.}

\item{covariates}{a character vector of no more than two variable names that
should be used as covariates in downstream analyses. Covariates are
typically variables that a user wants to account for in the analysis but
quantifying/examining the effect of the variable is not of interest.}

\item{cov_type}{An optional character vector (must be the same length as
\code{covariates} if used) indicating the class or type of each covariate.
For example, "numeric", "character", or "factor". Partial matching ("num"
for "numeric") is NOT used and the entire class/type must be typed out. If
the class of a covariate does not match the input to \code{cov_type} the
covariate will be coerced to that type. For example, if the covariate is a
numeric vector of 0s and 1s (indicating two categories) and the input to
\code{cov_type} is a class other than numeric this vector will be coerced
to a character vector. The default value is NULL. In this case the class of
the covariates is neither checked nor altered.}

\item{pairs}{A character string indicating the column in \code{f_data} that
contains the pairing information. This string must match the column name
exactly.}

\item{time_course}{an optional character string specifying the variable name
of a time course variable, if applicable to the experimental data.
CURRENTLY NOT SUPPORTED}
}
\value{
An object of the same class as the input \code{omicsData} object -
  the provided object with the samples filtered out, if any NAs were produced
  in designating groups. An attribute 'group_DF', a data.frame with columns
  for sample id and group, is added to the object. If two main effects are
  provided the original main effect levels for each sample are returned as
  the third and fourth columns of the data.frame. Additionally, the
  covariates provided will be listed as attributes of this data.frame. Use of
  time_course is currently not supported, however the eventual functionality
  is such that if time_course is included, a column for 'TimeCourse' will be
  output as well.
}
\description{
The method assigns each sample to a group, for future analyses, based on the
variable(s) specified as main effects.
}
\details{
Groups are formed based on the levels of the main effect variables.
  One or two main effect variables are allowed. In the case of two main
  effect variables, groups are formed based on unique combinations of the
  levels of the two main effect variables. Any samples with level NA for a
  main effect variable will be removed from the data and will not be included
  in the final group designation results. Groups with a single sample are
  allowed, as is a single group.
}
\examples{
\dontrun{
library(pmartRdata)
data(lipid_object)
lipid_object2 <- group_designation(omicsData = lipid_object,
                                   main_effects = "Condition")
attr(lipid_object2, "group_DF")
}

}
\author{
Lisa Bramer, Kelly Stratton
}
