% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtest_filter.R
\name{gtest_filter}
\alias{gtest_filter}
\title{Identifies peptides to be filtered out in preparation for IMD-ANOVA.}
\usage{
gtest_filter(nonmiss_per_group, groupDF = NULL, omicsData = NULL,
  e_data = NULL, alpha = NULL, min_nonmiss_gtest = NULL,
  cname_id = NULL, samp_id = NULL)
}
\arguments{
\item{nonmiss_per_group}{list created by \code{\link{nonmissing_per_group}}. The first element giving the total number of possible samples for each group. The second element giving a data.frame with the first column giving the biomolecule and the second through kth columns giving the number of non-missing observations for each of the \code{k} groups.}

\item{groupDF}{data.frame created by \code{group_designation} with columns for sampleID and Group. If omicsData is supplied, this argument is optional; if e_data is supplied, this argument is required. If two main effects are provided the original main effect levels for each sample are returned as the third and fourth columns of the data.frame.}

\item{omicsData}{an optional object of one of the classes "pepData", "proData", "lipidData", or "metabData" usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.lipidData}}, or \code{\link{as.metabData}}, respectively.}

\item{e_data}{an optional \eqn{p} by \eqn{n} data.frame, where \eqn{p} is the number of peptides, proteins, lipids, or metabolites and \eqn{n} is the number of samples.}

\item{alpha}{numeric value specifing the p-value cut-off to be used when determining whether to filter out each peptide. Either \code{alpha} or \code{min.nonmiss.allowed}, but not both, must be specified; this specification determines which approach is used for filtering the peptides. See \link{Details} for more information.}

\item{min_nonmiss_gtest}{the minimum number of non-missing peptide values allowed in a minimum of one group. Default value is 3. Either \code{alpha} or \code{min.nonmiss.allowed} must be specified; this specification determines which approach is used for filtering the peptides. See \link{Details} for more information.}

\item{cname_id}{name for biomolecule identifier column found in \code{e_data} and \code{e_meta} (if applicable)}

\item{samp_id}{name for sample identifier column found in \code{f_data}}
}
\value{
filter.peps a character vector of the peptides to be filtered out prior to the G-test or IMD-ANOVA
}
\description{
The method identifies peptides, proteins, lipids, or metabolites to be filtered specifically according to the G-test.
}
\details{
Two methods are available for determining the peptides to be filtered. The naive approach is based on \code{min.nonmiss.allowed}, and looks for peptides that do not have at least \code{min.nonmiss.allowed} values per group. The other approach also looks for peptides that do not have at least a minimum number of values per group, but this minimum number is determined using the G-test and a p-value threshold supplied by the user. The G-test is a test of independence, used here to test the null hypothesis of independence between the number of missing values across groups.
}
\examples{
dontrun{
library(MSomicsQC)
data(pep_object)
pep_object2 <- group_designation(omicsData = pep_object, main_effects = "Condition")
nonmissing_result <- nonmissing_per_group(omicsData = pep_object2)
to_filter <- gtest_filter(nonmiss_per_group = nonmissing_result, omicsData = pep_object2, min_nonmiss_gtest = 3)
}

}
\author{
Kelly Stratton
}

