% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingval_result.R
\name{missingval_result}
\alias{missingval_result}
\title{Creates an object of class naRes}
\usage{
missingval_result(omicsData)
}
\arguments{
\item{omicsData}{an object of class "pepData", "proData", "metabData",
"lipidData", or "nmrData", created by \code{\link{as.pepData}},
\code{\link{as.proData}}, \code{\link{as.metabData}},
\code{\link{as.lipidData}}, or \code{\link{as.nmrData}}, respectively.}
}
\value{
outputs a list of two data frames, one contains NA by sample (missing
  values per sample), the second contains NA by molecule (missing values per
  molecule). The output is assigned class 'naRes'.
}
\description{
This function takes in an omicsData object, outputs a list of two data
frames, one contains NA by sample, the second contains NA by molecule
}
\examples{
\dontrun{
library(pmartRdata)
data("lipid_object")
data("metab_object")

result = missingval_result(lipid_object)
result2 = missingval_result(metab_object)

}

}
