% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingval_scatterplot.R
\name{missingval_scatterplot}
\alias{missingval_scatterplot}
\title{missingval_scatterplot}
\usage{
missingval_scatterplot(omicsData, x_lab = NULL, y_lab = NULL, ...)
}
\description{
mean intensity vs number of missing values (per molecule)
}
\details{
\tabular{ll}{
\code{x_lab} \tab character string to be used for x-axis label. Defaults to NULL \cr
\code{y_lab} \tab character string to be used for y-axis label. Defaults to NULL \cr
\code{title_plot} \tab character string to be used for the plot title. Defaults to NULL. \cr
\code{legend_title} \tab character string to be used for legend_title label. Defaults to NULL \cr
\code{title_size} \tab integer value specifying the font size for the plot title. Default is 14. \cr
\code{x_lab_size} \tab integer value indicating the font size for the x-axis. Defaults to 11. \cr
\code{y_lab_size} \tab integer value indicating the font size for the y-axis. Defaults to 11. \cr
\code{point_size} \tab integer value indicating scatterplot point size, defaults to 3. \cr
\code{palette} \tab character string indicating the name of the RColorBrewer palette to use. \cr
\code{bw_theme} \tab logical indicator of whether to use the "theme_bw". Defaults to FALSE, in which case the ggplot2 default theme is used. \cr
}
}
