% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingval_volcanoplot.R
\name{missingval_volcanoplot}
\alias{missingval_volcanoplot}
\title{Produces a volcano plot from statRes object}
\usage{
missingval_volcanoplot(statRes, comparison, x_lab = NULL, ...)
}
\arguments{
\item{statRes}{is an object of class 'statRes'}

\item{comparison}{can either be a character string name of the comparison to plot, or an integer index refering to the comparisons attribute vector}

\item{x_lab}{character string to be used for x-axis label. Defaults to NULL}

\item{...}{further arguments

\tabular{ll}{
\code{y_lab} \tab character string to be used for y-axis label. Defaults to NULL \cr
\code{title_plot} \tab character string to be used for the plot title. Defaults to NULL. \cr
\code{legend_title} \tab character string to be used for legend_title label. Defaults to NULL \cr
\code{title_size} \tab integer value specifying the font size for the plot title. Default is 14. \cr
\code{x_lab_size} \tab integer value indicating the font size for the x-axis. Defaults to 11. \cr
\code{y_lab_size} \tab integer value indicating the font size for the y-axis. Defaults to 11. \cr
\code{bw_theme} \tab logical indicator of whether to use the "theme_bw". Defaults to FALSE, in which case the ggplot2 default theme is used. \cr
\code{vlines} \tab The x coordinate (integer in absolute value) where to draw vertical lines, defaults to NULL\cr
\code{pvalue_threshold} \tab numeric value, draws horizontal line at value, defaults to NULL\cr
\code{palette} \tab is a character string indicating the name of the RColorBrewer palette to use; "YlOrRd", "YlOrBr", "YlGnBu", "YlGn", "Reds","RdPu", "Purples", "PuRd", "PuBuGn", "PuBu", "OrRd","Oranges", "Greys", "Greens", "GnBu", "BuPu","BuGn","Blues", "Set3", "Set2", "Set1", "Pastel2", "Pastel1", "Paired", "Dark2", "Accent", "Spectral", "RdYlGn", "RdYlBu", "RdGy", "RdBu", "PuOr","PRGn", "PiYG", "BrBG"\cr
\code{x_lab_angle} \tab integer value indicating the angle of x-axis labels \cr
\code{coordinate_flip} \tab logical indicates whether to flip cartesian coordinates so that horizontal becomes vertical and vise versa, defaults to false \cr
}}
}
\value{
plots ggplot2 object
}
\description{
This function creates a fold change vs t-Test p-value volcano plot from a statRes object
}
\details{
This function takes in a statRes object and plots log2 fold change vs -log10 t-Test P-values for a comparison from attr(statRes, “comparison”), ‘comparison’ parameter could be either a string or an integer refereing to an index of statRes comparisons
}
\examples{
dontrun{
missingval_volcanoplot(statRes_object, comparison = 2, palette = "Set1")
missingval_volcanoplot(statRes_object, comparison = 3, pvalue_threshold = .005, palette = "Dark2", x_lab_angle = 45)

}

}
