% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_objects.R
\name{molecule_filter}
\alias{molecule_filter}
\title{Molecule filter object}
\usage{
molecule_filter(omicsData, use_groups = FALSE, use_batch = FALSE)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData',
'lipidData', 'nmrData', or 'seqData', usually created by \code{\link{as.pepData}},
\code{\link{as.proData}}, \code{\link{as.metabData}},
\code{\link{as.lipidData}}, or \code{\link{as.nmrData}}, 
or \code{\link{as.nmrData}}, respectively.}

\item{use_batch}{logical indicator for whether to utilize batch information
from \code{\link{group_designation}} when calculating the molecule filter.
Defaults to FALSE. Necessary to be set to TRUE if running ComBat batch
correction.}

\item{min_num}{an integer value specifying the minimum number of times each
feature must be observed across all samples. Default value is 2.}

\item{use_group}{logical indicator for whether to utilize group information
from \code{\link{group_designation}} when calculating the molecule filter.
Defaults to FALSE.}
}
\value{
Object of class moleculeFilt (also a data.frame) that contains the
  molecule identifier and the number of samples for which the molecule was
  measured (not NA)
}
\description{
This function returns a moleculeFilt object for use with
\code{\link{applyFilt}}
}
\details{
Attribute of molecule_filt object is "total_poss_obs", the number of
  total possible observations for each feature (same as the number of
  samples)
}
\examples{
\dontrun{
library(pmartRdata)
data("pep_object")
to_filter <- molecule_filter(omicsData = pep_object)
summary(to_filter, min_num = 2)
}

}
\author{
Kelly Stratton
}
