% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSnSet2pepData.R
\name{MSnSet2pepData}
\alias{MSnSet2pepData}
\title{Convert Data of class MSnSet to pmartR pepData Class}
\usage{
MSnSet2pepData(
  msnset_object,
  data_scale,
  edata_cname = "UniqueID",
  fdata_cname = "SampleID",
  emeta_cname = "UniqueID",
  check.names = TRUE
)
}
\arguments{
\item{msnset_object}{an object of class MSnSet, which stores quantification
data and meta data. Creating an MSnSet object is described in the MSNbase
package io vignette.}

\item{data_scale}{Scale of the data provided in \code{e_data}. Acceptable
values are 'log2', 'log10', 'log', and 'abundance', which indicate data is
log base 2, base 10, natural log transformed, and raw abundance,
respectively.}

\item{edata_cname}{character string specifying the name of the column
containing the peptide identifiers in \code{e_data} and \code{e_meta} (if
applicable).}

\item{fdata_cname}{character string specifying the name of the column
containing the sample identifiers in \code{f_data}.}

\item{emeta_cname}{character string specifying the name of the column
containing the protein identifiers (or other mapping variable) in
\code{e_meta} (if applicable).}
}
\value{
pepData object
}
\description{
Converts an object of class MSnSet to an object of the class 'pepData'
}
\details{
The MSnbase package is available via Bioconductor
}
\examples{
\dontrun{
library(pmartR)
library(MSnbase)
data("msnset")
result = MSnSet2pepData(msnset,
                        data_scale = "log2",
                        edata_cname = "UniqueID",
                        fdata_cname = "SampleID",
                        emeta_cname = "UniqueID",
                        check.names = TRUE)
}

}
\references{
Gatto L, Lilley K (2012). “MSnbase - an R/Bioconductor package
  for isobaric tagged mass spectrometry data visualization, processing and
  quantitation.” Bioinformatics, 28, 288-289.

Gatto L, Gibb S, Rainer J (2020). “MSnbase, efficient and elegant
  R-based processing and visualisation of raw mass spectrometry data.”
  bioRxiv.
}
