% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot.corRes}
\alias{plot.corRes}
\title{plot.corRes}
\usage{
\method{plot}{corRes}(
  corRes_obj,
  omicsData = NULL,
  order_by = NULL,
  colorbar_lim = c(NA, NA),
  x_text = TRUE,
  y_text = TRUE,
  interactive = FALSE,
  x_lab = NULL,
  y_lab = NULL,
  x_lab_size = 11,
  y_lab_size = 11,
  x_lab_angle = 90,
  title_lab = NULL,
  title_lab_size = 14,
  legend_lab = NULL,
  legend_position = "right",
  color_low = NULL,
  color_high = NULL,
  bw_theme = TRUE,
  use_VizSampNames = FALSE
)
}
\arguments{
\item{corRes_obj}{An object of class "corRes" created by calling
\code{cor_result} on an omicsData object.}

\item{omicsData}{an object of the class 'pepData', 'isobaricpepData',
'proData', 'lipidData', 'metabData', or 'nmrData' usually created by
\code{\link{as.pepData}}, \code{\link{as.isobaricpepData}},
\code{\link{as.proData}}, \code{\link{as.lipidData}},
\code{\link{as.metabData}}, or\code{\link{as.nmrData}}, respectively.}

\item{order_by}{A character string specifying a main effect by which to order
the correlation heat map. This main effect must be found in the column
names of f_data in the omicsData object. If \code{order_by} is "Group", the
correlation heat map will be ordered by the group variable from the
group_designation function. If NULL (default), the correlation heat map
will be displayed in the order the samples appear in the data.}

\item{colorbar_lim}{A pair of numeric values specifying the minimum and
maximum values to use in the heat map color bar. Defaults to 'c(NA, NA)',
in which case ggplot2 automatically sets the minimum and maximum values
based on the correlation values in the data.}

\item{x_text}{Logical. Indicates whether the x-axis will be labeled with the
sample names. The default is TRUE.}

\item{y_text}{Logical. Indicates whether the y-axis will be labeled with the
sample names. The default is TRUE.}

\item{interactive}{Logical. If TRUE produces an interactive plot.}

\item{x_lab}{A character string specifying the x-axis label.}

\item{y_lab}{A character string specifying the y-axis label.}

\item{x_lab_size}{An integer value indicating the font size for the x-axis.
The default is 11.}

\item{y_lab_size}{An integer value indicating the font size for the y-axis.
The default is 11.}

\item{x_lab_angle}{An integer value indicating the angle of x-axis labels.
The default is 90.}

\item{title_lab}{A character string specifying the plot title.}

\item{title_lab_size}{An integer value indicating the font size of the plot
title. The default is 14.}

\item{legend_lab}{A character string specifying the legend title.}

\item{legend_position}{A character string specifying the position of the
legend. Can be one of "right", "left", "top", "bottom", or "none". The
default is "none".}

\item{color_low}{A character string specifying the color of the gradient for
low values.}

\item{color_high}{A character string specifying the color of the gradient for
high values.}

\item{bw_theme}{Logical. If TRUE uses the ggplot2 black and white theme.}

\item{use_VizSampNames}{Logical. Indicates whether to use custom sample
names. The default is FALSE.}
}
\description{
For plotting an S3 object of type 'corRes'
}
