% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot.customFilt}
\alias{plot.customFilt}
\title{Plot customFilt Object}
\usage{
\method{plot}{customFilt}(filter_object)
}
\arguments{
\item{filter_object}{An object of class customFilt.}
}
\description{
Currently plotting a customFilt object is not supported
}
