% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dataRes.R
\name{plot.dataRes}
\alias{plot.dataRes}
\title{Plots an object of class dataRes}
\usage{
\method{plot}{dataRes}(dataRes, metric = NULL)
}
\arguments{
\item{dataRes}{an object of the class 'dataRes, usually created by \code{\link{summarize}}.}

\item{metric}{a character string indicating which metric to use in plot, one of 'mean', 'median', 'sd, 'min' or 'max'}
}
\value{
plots ggplot2 object
}
\description{
For plotting an S3 object of type 'dataRes'
}
\details{
This function can only create plots of dataRes objects whose 'by' and 'groupvar' attributes are non NULL
}
\examples{
dontrun{


}

}
