% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot.isobaricnormRes}
\alias{plot.isobaricnormRes}
\title{Plots an object of class isobaricnormRes}
\usage{
\method{plot}{isobaricnormRes}(
  isobaricnormRes_obj,
  order = FALSE,
  interactive = FALSE,
  x_lab = NULL,
  y_lab = NULL,
  x_lab_size = 11,
  y_lab_size = 11,
  x_lab_angle = NULL,
  title_lab = NULL,
  title_lab_size = 14,
  legend_lab = NULL,
  legend_position = "none",
  bw_theme = TRUE,
  palette = NULL
)
}
\arguments{
\item{isobaricnormRes_obj}{an object of type isobaricnormRes, created by
\code{\link{normalize_isobaric}}}

\item{order}{Logical. If TRUE the samples will be ordered by the column of
f_data containing the experiment/plate information.}

\item{interactive}{Logical. If TRUE produces an interactive plot.}

\item{x_lab}{A character string specifying the x-axis label.}

\item{y_lab}{A character string specifying the y-axis label.}

\item{x_lab_size}{An integer value indicating the font size for the x-axis.
The default is 11.}

\item{y_lab_size}{An integer value indicating the font size for the y-axis.
The default is 11.}

\item{x_lab_angle}{An integer value indicating the angle of x-axis labels.}

\item{title_lab}{A character string specifying the plot title.}

\item{title_lab_size}{An integer value indicating the font size of the plot
title. The default is 14.}

\item{legend_lab}{A character string specifying the legend title.}

\item{legend_position}{A character string specifying the position of the
legend. Can be one of "right", "left", "top", "bottom", or "none". The
default is "none".}

\item{bw_theme}{Logical. If TRUE uses the ggplot2 black and white theme.}

\item{palette}{A character string indicating the name of the RColorBrewer
palette to use. For a list of available options see the details section in
\code{\link[RColorBrewer]{RColorBrewer}}.}
}
\value{
plots ggplot2 object
}
\description{
Creates box plots for an S3 object of type 'isobaricnormRes'
}
\examples{
\dontrun{
library(pmartRdata)

data(isobaric_object)

isobaric_object = edata_transform(isobaric_object, "log2")

result = normalize_isobaric(isobaric_object, exp_cname = "Set",
                            apply_norm = FALSE,
                            channel_cname = "iTRAQ.Channel",
                            refpool_channel = "116")

plot(result)
}

}
