% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missingval.R
\name{plot.naRes}
\alias{plot.naRes}
\title{plot.naRes}
\usage{
\method{plot}{naRes}(naRes_object, type, x_lab = NULL, ...)
}
\arguments{
\item{type}{is for specifying plot type, there are three options, "bySample", "byMolecule" and "Both"
\tabular{ll}{
\code{x_lab} \tab character string to be used for x-axis label. Defaults to NULL \cr
\code{y_lab} \tab character string to be used for y-axis label. Defaults to NULL \cr
\code{title_plot} \tab character string to be used for the plot title. Defaults to NULL. \cr
\code{title_size} \tab integer value specifying the font size for the plot title. Default is 14. \cr
\code{x_lab_size} \tab integer value indicating the font size for the x-axis. Defaults to 11. \cr
\code{y_lab_size} \tab integer value indicating the font size for the y-axis. Defaults to 11. \cr
\code{bar_width} \tab integer value indicating the bar width in a barplot (when type = "bySample"). Defaults to .8. \cr
\code{binwidth} \tab integer value indicating the bin width in a histogram (when type = "byMolecule"). Defaults to 1. \cr
\code{bw_theme} \tab logical indicator of whether to use the "theme_bw". Defaults to FALSE, in which case the ggplot2 default theme is used. \cr
}}
}
\description{
For plotting an S3 object of type 'naRes':
}
