% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot.naRes}
\alias{plot.naRes}
\title{Plot naRes Object}
\usage{
\method{plot}{naRes}(
  naRes_obj,
  omicsData,
  plot_type = "bar",
  order_by = NULL,
  color_by = NULL,
  interactive = FALSE,
  x_lab_bar = NULL,
  x_lab_scatter = NULL,
  y_lab_bar = NULL,
  y_lab_scatter = NULL,
  x_lab_size = 11,
  y_lab_size = 11,
  x_lab_angle = 60,
  title_lab_bar = NULL,
  title_lab_scatter = NULL,
  title_lab_size = 14,
  legend_lab_bar = NULL,
  legend_lab_scatter = NULL,
  legend_position = "right",
  point_size = 2,
  text_size = 3,
  bar_width = 0.8,
  bw_theme = TRUE,
  palette = NULL,
  display_count = TRUE,
  coordinate_flip = FALSE,
  use_VizSampNames = FALSE
)
}
\arguments{
\item{naRes_obj}{list of two data frames, one containing the number of
missing values by sample, and the other containing missing values by
molecule}

\item{omicsData}{object of class 'pepData', 'proData', 'metabData',
'lipidData', nmrData', or 'seqData', created by \code{\link{as.pepData}},
\code{\link{as.proData}}, \code{\link{as.metabData}},
\code{\link{as.lipidData}}, \code{\link{as.nmrData}}, or \code{\link{as.seqData}}, respectively.}

\item{plot_type}{character string specifying which type of plot to produce.
The two options are 'bar' or 'scatter'.}

\item{order_by}{character string specifying a main effect by which to order
the bar plot. This main effect must be found in the column names of f_data
in the omicsData object. If \code{order_by} is "Group", the bar plot
will be ordered by the group variable from the group_designation function.
If NULL (default), the bar plot will be displayed in the order they appear
in the data.}

\item{color_by}{character string specifying a main effect by which to color
the bar plot. This main effect must be found in the column names of f_data
in the omicsData object. If \code{color_by} is "Group", the bar plot
will be colored by the group variable from the group_designation function.
If NULL (default), the bar plot will have one default color.}

\item{interactive}{logical value. If TRUE produces an interactive plot.}

\item{x_lab_bar}{character string used for the x-axis label for the bar
plot}

\item{x_lab_scatter}{character string used for the x-axis label for the
scatter plot}

\item{y_lab_bar}{character string used for the y-axis label for the bar
plot}

\item{y_lab_scatter}{character string used for the y-axis label for the
scatter plot}

\item{x_lab_size}{integer value indicating the font size for the x-axis.
The default is 11.}

\item{y_lab_size}{integer value indicating the font size for the y-axis.
The default is 11.}

\item{x_lab_angle}{integer value indicating the angle of x-axis labels.}

\item{title_lab_bar}{character string used for the plot title when
\code{plot_type} is 'bar'.}

\item{title_lab_scatter}{character string used for the plot title when
\code{plot_type} is 'scatter'.}

\item{title_lab_size}{integer value indicating the font size of the plot
title. The default is 14.}

\item{legend_lab_bar}{character string specifying the legend title when
creating a bar plot.}

\item{legend_lab_scatter}{character string specifying the legend title when
creating a scatter plot.}

\item{legend_position}{character string specifying the position of the
legend. Can be one of "right", "left", "top", or "bottom". The default is
"right".}

\item{point_size}{An integer specifying the size of the points. The default
is 2.}

\item{text_size}{An integer specifying the size of the text (number of
missing values by sample) within the bar plot. The default is 3.}

\item{bar_width}{An integer indicating the width of the bars in the bar plot.
The default is 0.8.}

\item{bw_theme}{logical value. If TRUE uses the ggplot2 black and white theme.}

\item{palette}{character string indicating the name of the RColorBrewer
palette to use. For a list of available options see the details section in
\code{\link[RColorBrewer]{RColorBrewer}}.}

\item{display_count}{logical value. Indicates whether the missing value
counts by sample will be displayed on the bar plot. The default is TRUE.}

\item{coordinate_flip}{logical value. Indicates whether the x and y axes will
be flipped. The default is FALSE.}

\item{use_VizSampNames}{logical value. Indicates whether to use custom sample
names. The default is FALSE.}
}
\value{
ggplot2 plot object if interactive is FALSE, or plotly plot object if
  interactive is TRUE
}
\description{
For plotting an S3 object of type 'naRes'
}
\details{
This function takes in an object of class naRes and creates either a
  bar or scatter plot of missing values. When plot_type = 'bar', a sample
  name by missing values count bar chart is returned. When plot_type =
  'scatter' a mean intensity vs number of missing values (per molecule)
  scatter plot is returned. Note: If the omicsData object has had
  \code{\link{group_designation}} applied to it, the points in the plot will
  be colored by group.
}
\examples{
library(pmartRdata)
mylipid <- group_designation(omicsData = lipid_neg_object, main_effects = "Virus")
result <- missingval_result(omicsData = mylipid)
plot(naRes_obj = result, omicsData = mylipid, plot_type = "bar", x_lab_angle = 50, order_by = "Virus", color_by = "Virus")
plot(naRes_obj = result, omicsData = mylipid, plot_type = "scatter", x_lab_angle = 50, color_by = "Virus")

result <- missingval_result(omicsData = rnaseq_object)
plot(naRes_obj = result, omicsData = rnaseq_object, plot_type = "bar")

}
