% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nmrnormRes.R
\name{plot.nmrnormRes}
\alias{plot.nmrnormRes}
\title{Plots an object of class nmrnormRes}
\usage{
\method{plot}{nmrnormRes}(nmrnormRes_object, x_lab = NULL, ...)
}
\arguments{
\item{nmrnormRes_object}{an object of type nmrnormRes, created by \code{\link{normalize_nmr}}}

\item{x_lab}{character string to be used for x-axis label. Defaults to NULL}

\item{...}{further arguments 

\tabular{ll}{
\code{y_lab} \tab character string to be used for y-axis label. Defaults to NULL \cr
\code{title_plot} \tab character string to be used for the plot title. Defaults to NULL. \cr
\code{title_size} \tab integer value specifying the font size for the plot title. Default is 14. \cr
\code{x_lab_size} \tab integer value indicating the font size for the x-axis. Defaults to 11. \cr
\code{y_lab_size} \tab integer value indicating the font size for the y-axis. Defaults to 11. \cr
\code{bw_theme} \tab logical indicator of whether to use the "theme_bw". Defaults to TRUE, in which case the ggplot2 default theme is used. \cr
}}
}
\value{
plots ggplot2 object
}
\description{
For plotting an S3 object of type 'nmrnormRes':
}
\examples{
dontrun{
library(pmartRdata)
data(nmr_object_identified)

nmr_object = edata_transform(nmr_object_identified, "log2")
nmr_norm = normalize_nmr(nmr_object, apply_norm = FALSE, metabolite_name = "unkm1.53")
plot(nmr_norm)

# alternate specification: #
data(nmr_object_identified)

nmr_object = edata_transform(nmr_object, "log2")
nmr_norm = normalize_nmr(nmr_object, apply_norm = FALSE, sample_property_cname = "Concentration")
plot(nmr_norm)
}

}
