% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pmartR.R
\name{plot.rmdFilt}
\alias{plot.rmdFilt}
\title{plot.rmdFilt}
\usage{
\method{plot}{rmdFilt}(filter_object, pvalue_threshold = NULL,
  sampleID = NULL, ...)
}
\arguments{
\item{pvalue_threshold}{A threshold for the Robust Mahalanobis Distance (RMD) p-value. If \code{sampleID} is NULL (see \code{sampleID} below), a horizontal line is plotted at the RMD value that corresponds with the threshold, and all samples above the line have a p-value below the threshold. If \code{sampleID} is not NULL, \code{pvalue_threshold} will do nothing. Default value is NULL.}

\item{sampleID}{If specified, the plot produces a boxplot instead of a scatterplot. The \code{sampleID} input will place an "x" at the value for each of the metrics on the boxplots. Default value is NULL.
\tabular{ll}{
\code{x_lab} \tab character string to be used for x-axis label. Defaults to NULL, in which case a default label is used. \cr
\code{y_lab} \tab character string to be used for y-axis label. Defaults to NULL, in which case a default label is used. \cr
\code{legend_lab} \tab character string specifying the title label to use for the legend \cr
\code{title_plot} \tab character string to be used for the plot title. Defaults to NULL, in which case a default title is used. \cr
\code{title_size} \tab integer value indicating the font size for the plot title. Defaults to 14.
\code{x_lab_size} \tab integer value indicating the font size for the plot title. Defaults to 11.
\code{y_lab_size} \tab integer value indicating the font size for the plot title. Defaults to 11.
\code{bw_theme} \tab logical indicator of whether to use the "theme_bw". Defaults to FALSE, in which case the ggplot2 default theme is used.
\code{legend_position} \tab character string specifying one of "right", "left", "top", or "bottom" for the location of the legend. Deafults to "right".
}}
}
\description{
For plotting an S3 object of type 'rmdFilt':
}
