% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot.rmdFilt}
\alias{plot.rmdFilt}
\title{plot.rmdFilt}
\usage{
\method{plot}{rmdFilt}(
  filter_obj,
  pvalue_threshold = NULL,
  sampleID = NULL,
  interactive = FALSE,
  x_lab = NULL,
  y_lab = NULL,
  x_lab_size = 11,
  y_lab_size = 11,
  x_lab_angle = 90,
  title_lab = NULL,
  title_lab_size = 14,
  legend_lab = NULL,
  legend_position = "right",
  point_size = 3,
  bw_theme = TRUE,
  palette = NULL,
  use_VizSampNames = FALSE
)
}
\arguments{
\item{filter_obj}{An object of class rmdFilt which is a data frame containing
columns for the sample ID, group, log2 robust Mahalanobis distance,
p-values, and metrics used to calcualte the robust Mahalanobis distance.}

\item{pvalue_threshold}{A threshold for the Robust Mahalanobis Distance (RMD)
p-value. If \code{sampleID} is NULL (see \code{sampleID} below), a
horizontal line is plotted at the RMD value that corresponds with the
threshold, and all samples above the line have a p-value below the
threshold. If \code{sampleID} is not NULL, \code{pvalue_threshold} will do
nothing. Default value is NULL.}

\item{sampleID}{A character vector specifying the sample names to be plotted.
If specified, the plot function produces a boxplot instead of a
scatterplot. A point, colored by sample, will be placed on each boxplot for
that sample's value for the given metric. The default value is NULL.}

\item{interactive}{Logical. If TRUE produces an interactive plot.}

\item{x_lab}{A character string specifying the x-axis label.}

\item{y_lab}{A character string specifying the y-axis label. The default is
NULL in which case the y-axis label will be the metric selected for the
\code{metric} argument.}

\item{x_lab_size}{An integer value indicating the font size for the x-axis.
The default is 11.}

\item{y_lab_size}{An integer value indicating the font size for the y-axis.
The default is 11.}

\item{x_lab_angle}{An integer value indicating the angle of x-axis labels.
The default is 90.}

\item{title_lab}{A character string specifying the plot title.}

\item{title_lab_size}{An integer value indicating the font size of the plot
title. The default is 14.}

\item{legend_lab}{A character string specifying the legend title.}

\item{legend_position}{A character string specifying the position of the
legend. Can be one of "right", "left", "top", "bottom", or "none". The
default is "none".}

\item{point_size}{An integer specifying the size of the points. The default
is 3.}

\item{bw_theme}{Logical. If TRUE uses the ggplot2 black and white theme.}

\item{palette}{A character string indicating the name of the RColorBrewer
palette to use. For a list of available options see the details section in
\code{\link[RColorBrewer]{RColorBrewer}}.}

\item{use_VizSampNames}{Logical. Indicates whether to use custom sample
names. The default is FALSE.}
}
\description{
For plotting an S3 object of type 'rmdFilt'
}
