% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SPANSRes.R
\name{plot.SPANSRes}
\alias{plot.SPANSRes}
\title{Plots an object of class SPANSRes}
\usage{
\method{plot}{SPANSRes}(SPANSRes_obj)
}
\arguments{
\item{SPANSRes_obj}{an object of the class 'SPANSRes, usually created by \code{\link{spans_procedure()}}.}
}
\value{
plots a plotly object
}
\description{
For plotting an S3 object of type 'SPANSRes'
}
