% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot.statRes}
\alias{plot.statRes}
\title{Plotting function for `statRes` objects}
\usage{
\method{plot}{statRes}(
  x,
  plot_type = "bar",
  fc_threshold = NULL,
  fc_colors = c("red", "black", "green"),
  stacked = FALSE,
  interactive = FALSE,
  x_lab = NULL,
  x_lab_size = 11,
  x_lab_angle = NULL,
  y_lab = NULL,
  y_lab_size = 11,
  title_lab = NULL,
  title_lab_size = 14,
  legend_lab = NULL,
  legend_position = "right",
  text_size = 3,
  bw_theme = TRUE,
  display_count = TRUE,
  custom_theme = NULL
)
}
\arguments{
\item{x}{`statRes` object to be plotted, usually the result of `imd_anova`}

\item{plot_type}{defines which plots to be produced, options are "bar",
"volcano", "gheatmap", "fcheatmap"; defaults to "bar".  See details for
plot descriptions.}

\item{fc_threshold}{optional threshold value for fold change estimates.
Modifies the volcano plot as follows:  Vertical lines are added at
(+/-)\code{fc_threshold} and all observations that have absolute fold
change less than \code{abs(fc_threshold)} are colored as 'non-significant'
(as specified by \code{fc_colors}).}

\item{fc_colors}{vector of length three with character color values
interpretable by ggplot. i.e. c("orange", "black", "blue") with the values
being used to color negative, non-significant, and positive fold changes
respectively}

\item{stacked}{TRUE/FALSE for whether to stack positive and negative fold
change sections in the barplot, defaults to FALSE}

\item{interactive}{TRUE/FALSE for whether to create an interactive plot using
plotly.  Not valid for all plots.}

\item{x_lab}{A character string specifying the x-axis label.}

\item{x_lab_size}{An integer value indicating the font size for the x-axis.
The default is 11.}

\item{x_lab_angle}{An integer value indicating the angle of x-axis labels.}

\item{y_lab}{A character string specifying the y-axis label.}

\item{y_lab_size}{An integer value indicating the font size for the y-axis.
The default is 11.}

\item{title_lab}{A character string specifying the plot title.}

\item{title_lab_size}{An integer value indicating the font size of the plot
title. The default is 14.}

\item{legend_lab}{A character string specifying the legend title.}

\item{legend_position}{A character string specifying the position of the
legend. Can be one of "right", "left", "top", "bottom", or "none". The
default is "none".}

\item{text_size}{An integer specifying the size of the text (number of
non-missing values) within the plot. The default is 3.}

\item{bw_theme}{Logical. If TRUE uses the ggplot2 black and white theme.}

\item{display_count}{Logical. Indicates whether the non-missing counts will
be displayed on the bar plot. The default is TRUE.}

\item{custom_theme}{a ggplot `theme` object to be applied to non-interactive
plots, or those converted by plotly::ggplotly().}
}
\description{
Produces plots that summarize the results contained in a `statRes` object.
}
\details{
Plot types:
\itemize{
 \item{"bar"} \code{?pmartR::statres_barplot} Bar-chart with bar heights
 indicating the number of significant biomolecules, grouped by test type and
 fold change direction.
 \item{"volcano"} \code{?pmartR::statres_volcano_plot} Scatterplot showing
 negative-log-pvalues against fold change.  Colored by statistical
 significance and fold change.
 \item{"gheatmap"} \code{?pmartR::gtest_heatmap} Heatmap with x and y axes
 indicating the number of nonmissing values for two groups.  Colored by
 number of biomolecules that fall into that combination of nonmissing values.
 \item{"fcheatmap"} Heatmap showing all biomolecules across comparisons,
 colored by fold change.
}
}
\examples{
\dontrun{
library(pmartR)
library(pmartRdata)
#Transform the data

#Group the data by condition
myproData <- group_designation(omicsData = pro_object,
                               main_effects = c("Condition"))

#Apply the IMD ANOVA filter
imdanova_Filt <- imdanova_filter(omicsData = myproData)
myproData <- applyFilt(filter_object = imdanova_Filt,
                       omicsData = myproData,
                       min_nonmiss_anova=2)

#Implement the IMD ANOVA method and compuate all pairwise comparisons
#(i.e. leave the `comparisons` argument NULL)
anova_res <- imd_anova(omicsData = myproData, test_method = 'anova')
plot(anova_res)
plot(anova_res, plot_type = "volcano")

imd_res <- imd_anova(omicsData = myproData, test_method = 'gtest')
plot(imd_res)
plot(imd_res, plot_type = "gheatmap")
# using arguments of internal functions:
plot(imd_res,
     plot_type = "gheatmap",
     color_low = "red",
     color_high = "green")

imd_anova_res <- imd_anova(omicsData = myproData,
                           test_method = 'comb',
                           pval_adjust='bon')
plot(imd_anova_res, bw_theme = TRUE)
plot(imd_anova_res, plot_type = "volcano", bw_theme = TRUE)

}

}
