% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.omicsPlotter.R
\name{plot_comps}
\alias{plot_comps}
\title{Plot pairwise comparisons of omicsPlotter object}
\usage{
plot_comps(omicsPlotter, ...)
}
\arguments{
\item{omicsPlotter}{An object of class "omicsPlotter" generated from \code{\link{as.omicsPlotter}}.}

\item{comps_y_limits}{For comparisons: Set to "fixed" or "free" for automated y-axis calculating. "fixed" - axis generated based on the maximum/minimum across all plots. "free" - axis axis generated based on the maximum/minimum of individual plot.}

\item{comps_y_range}{For comparisons: Specify a range for the plot y-axis. Will calculate the range based on one of y_max or y_min parameters or from the median of y-values where y_max and y_min are not defined.}

\item{comps_y_max}{For comparisons: Sets the maximum y-value for the y-axis.}

\item{comps_y_min}{For comparisons: Sets the minimum y-value for the y-axis.}

\item{value_y_limits}{For values: Set to "fixed" or "free" for automated y-axis calculating. "fixed" - axis generated based on the maximum/minimum across all plots. "free" - axis axis generated based on the maximum/minimum of individual plot.}

\item{value_y_range}{For values: Specify a range for the plot y-axis. Will calculate the range based on one of y_max or y_min parameters or from the median of y-values where y_max and y_min are not defined.}

\item{value_y_max}{For values: Sets the maximum y-value for the y-axis.}

\item{value_y_min}{For values: Sets the minimum y-value for the y-axis.}

\item{p_val}{Specifies p-value for setting graph border colors}

\item{panel_variable}{Specifies what to divide trelliscope panels by, must be a column in omicsPlotter. Defaults to cnames$edata_cname of omicsPlotter.}

\item{comps_panel_x_axis}{Specifies what column should be on the x-axis, must be a column in omicsPlotter. Defaults setting plots pairwise comparisons along x-axis.}

\item{comps_panel_y_axis}{Specifies what column should be on the y-axis, must be a column in omicsPlotter and numeric. Defaults setting plots fold change for combined and anova testing and counts for g-test.}

\item{comps_color_variable}{Specifies what column should distingush color, must be a column in omicsPlotter. Default settings is set to "Group."}

\item{panel_x_axis}{Specifies what column should be on the x-axis, must be a column in omicsPlotter. Defaults setting plots pairwise comparisons along x-axis.}

\item{panel_y_axis}{Specifies what column should be on the y-axis, must be a column in omicsPlotter and numeric. Defaults setting plots fold change for combined and anova testing and counts for g-test.}

\item{color_variable}{Specifies what column should distingush color, must be a column in omicsPlotter. Default settings is set to "Group."}

\item{value_plot_type}{For values: Specifies plot types for graphing; must be a list of strings where "box", "bar", or "point" is specified. Combined strings like "boxpoint" will plot both in the same graph.}

\item{comps_plot_type}{For comparisons: Specifies plot types for graphing; must be a list of strings where "box", "bar", or "point" is specified. Combined strings like "boxpoint" will plot both in the same graph.}
}
\description{
Plot pairwise comparisons of omicsPlotter. Plots differently depending on statistical tests ran, located in attributes.
}
\author{
Rachel Richardson
}
