% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.omicsPlotter.R
\name{plot_comps}
\alias{plot_comps}
\title{Plot pairwise comparisons of omicsPlotter}
\usage{
plot_comps(omicsPlotter)
}
\arguments{
\item{omicsPlotter}{An object of class "omicsPlotter" generated from \code{\link{as.omicsPlotter}}.}
}
\description{
Plot pairwise comparisons of omicsPlotter. Plots differently depending on statistical tests ran, located in attributes.
}
\author{
Rachel Richardson
}
