% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmartR.R
\docType{_PACKAGE}
\name{pmartR}
\alias{pmartR}
\title{Panomics Marketplace - Quality Control and Statistical Analysis for Panomics Data}
\value{
No return value, used to appease R CMD check.
}
\description{
Provides functionality for quality control processing and
 statistical analysis of mass spectrometry (MS) omics data, in particular 
 proteomic (either at the peptide or the protein level), lipidomic, and 
 metabolomic data, as well as RNA-seq based count data and nuclear magnetic
 resonance (NMR) data. This includes data transformation, specification of 
 groups that are to be compared against each other, filtering of features 
 and/or samples, data normalization, data summarization (correlation, PCA), 
 and statistical comparisons between defined groups.
}
