% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_filter_objects.R
\name{print.imdanovaFilt}
\alias{print.imdanovaFilt}
\title{print.imdanovaFilt}
\usage{
\method{print}{imdanovaFilt}(filter_object)
}
\description{
For printing an S3 object of type 'imdanovaFilt'
}
