% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary.R
\name{print.customFilterSummary}
\alias{print.customFilterSummary}
\title{Custom Filter Print Method}
\usage{
\method{print}{customFilterSummary}(object)
}
\description{
Print method for summary of custom filter
}
