% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary.R
\name{print.customFilterSummary}
\alias{print.customFilterSummary}
\title{Custom Filter Print Method}
\usage{
\method{print}{customFilterSummary}(x, ...)
}
\arguments{
\item{x}{the custom filter summary to print}

\item{...}{further arguments passed to or from other methods}
}
\value{
No return value, prints details about x
}
\description{
Print method for summary of custom filter
}
