% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary.R
\name{print.cvFilterSummary}
\alias{print.cvFilterSummary}
\title{Print method for summary of CV filter}
\usage{
\method{print}{cvFilterSummary}(object)
}
\description{
Print method for summary of CV filter
}
