% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary.R
\name{print.imdanovaFilterSummary}
\alias{print.imdanovaFilterSummary}
\title{IMD-ANOVA Filter Print Method}
\usage{
\method{print}{imdanovaFilterSummary}(x, ...)
}
\arguments{
\item{x}{the imdanova filter summary to print}

\item{...}{further arguments passed to or from other methods}
}
\value{
No return value, prints details about x
}
\description{
Print method for summary of imdanova filter
}
