% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_filter.R
\name{proteomics_filter}
\alias{proteomics_filter}
\title{Proteomics filter object}
\usage{
proteomics_filter(omicsData)
}
\arguments{
\item{omicsData}{an object of class "pepData", the a result of \code{\link{as.pepData}}. The e_meta component of omicsData must be nonempty.}
}
\value{
a list of two character vectors, the first, \code{peptides_filt}, giving degenerate peptide names. The second, \code{proteins_filt}, gives the names of proteins which no longer have peptides mapping to them in the dataset.
}
\description{
This function counts the number of peptides that map to each protein and/or the number of proteins to which each individual peptide maps.
}
\details{
min_num_peps is an attribute of the proteomicsFilt object
}
\examples{
dontrun{
library(MSomicsQC)
data("pep_object")
my_filter <- proteomics_filter(omicsData = pep_object)
summary(my_filter, min_num_peps = 3)
}

}
\author{
Lisa Bramer, Kelly Stratton
}

