% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_filter.R
\name{proteomics_filter}
\alias{proteomics_filter}
\title{Proteomics filter object}
\usage{
proteomics_filter(omicsData)
}
\arguments{
\item{omicsData}{an object of class "pepData", the a result of
\code{\link{as.pepData}}. The e_meta component of omicsData must be
nonempty.}
}
\value{
A list with two elements. The first element is a data frame of counts 
        for each unique peptide. The second element is also a data frame.
        This data frame contains the counts for the number of peptides that
        map to each unique protein.
}
\description{
This function counts the number of peptides that map to each protein and/or
the number of proteins to which each individual peptide maps.
}
\examples{
dontrun{
library(pmartR)
data("pep_object")
my_filter <- proteomics_filter(omicsData = pep_object)
summary(my_filter, min_num_peps = 3)
}

}
\author{
Lisa Bramer, Kelly Stratton
}
