% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliVis.R
\name{recursive_format}
\alias{recursive_format}
\title{Recursive call of as.trellData and associated checks for list inputs}
\usage{
recursive_format(omicsData = NULL, omicsStats = NULL)
}
\arguments{
\item{omicsData}{A pmartR object of class pepData, lipidData, metabData, or proData}

\item{omicsStats}{A statistical results object produced by running \code{imd_anova} on omicsData.}
}
\description{
Checks for validity of list inputs and handles different list combinations as input.
}
\seealso{
\link[pmartR]{as.trellData}
}
\author{
Rachel Richardson
}
