% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_filter.R
\name{rmd_filter}
\alias{rmd_filter}
\alias{rmd_filter.lipidData}
\alias{rmd_filter.metabData}
\alias{rmd_filter.pepData}
\alias{rmd_filter.proData}
\title{RMD Runs}
\usage{
rmd_filter(omicsData, ...)

\method{rmd_filter}{pepData}(omicsData, metrics = c("MAD", "Kurtosis",
  "Skewness", "Correlation", "Proportion_Missing"))

\method{rmd_filter}{proData}(omicsData, metrics = c("MAD", "Kurtosis",
  "Skewness", "Correlation", "Proportion_Missing"))

\method{rmd_filter}{lipidData}(omicsData, metrics = c("MAD", "Kurtosis",
  "Skewness", "Correlation"))

\method{rmd_filter}{metabData}(omicsData, metrics = c("MAD", "Kurtosis",
  "Skewness", "Correlation"))
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData', or 'lipidData', usually created by \code{\link{as.pepData}}, \code{\link{as.proData}}, \code{\link{as.metabData}}, or \code{\link{as.lipidData}}, respectively.}

\item{...}{further arguments}
}
\value{
a data.frame containing columns for the Sample ID, log2 robust Mahalanobis distance, p-values, and robust Mahalanobis distance
}
\description{
The method computes a robust Mahalanobis distance that can be mapped to a p-value and used to identify outlying samples
}
\details{
The metrics on which the log2 robust Mahalanobis distance is based can be specified using the \code{metrics} argument.
\tabular{ll}{
pepData \tab For pepData objects, all five of the metrics "MAD", "Kurtosis", "Skewness", "Correlation", "Proportion_Missing" may be used (this is the default). \cr
proData \tab For proData objects, all five of the metrics "MAD", "Kurtosis", "Skewness", "Correlation", "Proportion_Missing" may be used (this is the default). \cr
metabData \tab For metabData objects, the use of "Proportion_Missing" is discouraged due to the general lack of missing data in metabolomics datasets (the default behavior omits "Proportion_Missing" from the metrics). \cr
lipidData \tab For lipidData objects, , the use of "Proportion_Missing" is discouraged due to the general lack of missing data in metabolomics datasets (the default behavior omits "Proportion_Missing" from the metrics). \cr
}
}
\examples{
dontrun{
library(pmartRdata)
data(metab_object)
metab_object2 <- edata_transform(omicsData = metab_object, data_scale = "log2")
metab_object3 <- group_designation(omicsData = metab_object2, main_effects = "Condition")
rmd_results <- rmd_filter(omicsData = metab_object3, metrics=c("MAD", "Skewness", "Correlation"))
rmd_results <- rmd_filter(omicsData = metab_object2)

data(pep_pepData)
pep_pepData2 <- edata_transform(omicsData = pep_pepData, data_scale = "log2")
pep_pepData3 <- group_designation(omicsData = pep_pepData2, main_effects = "Condition")
rmd_results <- rmd_filter(omicsData = pep_pepData3)
}

}
\author{
Lisa Bramer, Kelly Stratton
}
\references{
Matzke, M., Waters, K., Metz, T., Jacobs, J., Sims, A., Baric, R., Pounds, J., and Webb-Robertson, B.J. (2011), \emph{Improved quality control processing of peptide-centric LC-MS proteomics data}. Bioinformatics. 27(20): 2866-2872.
}

