% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrollup.R
\name{rrollup}
\alias{rrollup}
\title{Applies rrollup function}
\usage{
rrollup(pepData, combine_fn = "median", parallel = TRUE)
}
\arguments{
\item{pepData}{an omicsData object of class 'pepData'}

\item{combine_fn}{logical indicating what combine_fn to use, defaults to median, other option is mean}

\item{parallel}{logical indicating whether or not to use "doParallel" loop in applying rrollup function. Defaults to TRUE.}
}
\value{
an omicsData object of class 'proData'
}
\description{
This function applies the rrollup method to a pepData object for each unique protein and returns a proData object.
}
\details{
In the rrollup method, peptides are scaled based on a reference peptide and protein abundance is set as the mean of these scaled peptides.
}
\examples{
dontrun{
library(pmartRdata)
data(pep_object)
result = rrollup(pepData = pep_object) 
}

}
