% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrollup.R
\name{rrollup}
\alias{rrollup}
\title{rrollup}
\usage{
rrollup(pepData, parallel = TRUE)
}
\description{
started with the original rollup.R function and modified it to take in a pepData object. 
rrollup is applied to the pepData object and a proData object is returned.
}
