% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{set_data_info}
\alias{set_data_info}
\title{Functions to set omicsData attributes ----------------------------------------}
\usage{
set_data_info(
  e_data,
  edata_cname,
  data_scale_orig,
  data_scale,
  data_types,
  norm_info,
  is_normalized,
  batch_info,
  is_bc
)
}
\arguments{
\item{e_data, edata_cname, data_scale_orig, data_scale, data_types, norm_info, is_normalized, batch_info, is_bc}{See the \code{as.omicsData} functions, e.g. \code{\link{as.lipidData}}}
}
\value{
A list of edited data_info to replace the attribute in the
 omicsData object.
}
\description{
Create a set function that will return the value for each attribute. For
example, set_data_info will perform all of the calculations to fill in the
data_info attribute. These functions will be called in the as.xxx functions
to create an xxxData object but can be used individually to update any one of
the attributes at a later time.
Sets/updates the values in the data_info attribute
}
\keyword{internal}
