% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliVis.R
\name{set_increment}
\alias{set_increment}
\title{Sets y-axis increment for trellData labels plotting}
\usage{
set_increment(yvalues, include_zero = TRUE)
}
\arguments{
\item{yvalues}{y-values for plotting}

\item{include_zero}{Should zero be included in plot scaling?}
}
\description{
Sets y-axis increment for trellData plotting. Used by plot_comp.
}
\seealso{
\link[pmartR]{format_plot}

\link[pmartR]{list_y_limits}

\link[pmartR]{set_ylimits}
}
\author{
Rachel Richardson
}
