% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{statres_barplot}
\alias{statres_barplot}
\title{Fold change barplots for statres objects}
\usage{
statres_barplot(
  x,
  stacked,
  fc_colors,
  text_size,
  display_count,
  x_lab,
  y_lab,
  title_lab,
  legend_lab
)
}
\arguments{
\item{x, stacked, fc_colors}{passed from
\code{\link[pmartR:plot.statRes]{pmartR::plot.statRes()}}}

\item{text_size}{An integer specifying the size of the text (number of
non-missing values) within the plot. The default is 3.}

\item{display_count}{logical value. Indicates whether the non-missing counts will
be displayed on the bar plot. The default is TRUE.}

\item{x_lab}{character string specifying the x-axis label.}

\item{y_lab}{character string specifying the y-axis label.}

\item{title_lab}{character string specifying the plot title.}

\item{legend_lab}{character string specifying the legend title.}
}
\value{
`ggplot` object - barplot.
}
\description{
Plots a bar-chart with bar heights indicating the number of significant
biomolecules, grouped by test type and fold change direction.
}
\keyword{internal}
