% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statRes_class.R
\name{statRes_output}
\alias{statRes_output}
\title{Function to take raw output of `imd_anova` and create output for `statRes` object}
\usage{
statRes_output(imd_out, groupData, comparisons, test_method, pval_adjust,
  pval_thresh)
}
\arguments{
\item{imd_out}{Data.frame containing the results of the imd anova call}

\item{groupData}{the groupData object}

\item{comparisons}{the comparisons made}

\item{test_method}{the test method}

\item{pval_ajust}{pvalue adjustment method}

\item{pval_thres}{p-value threshold value}
}
\value{
the final object of class statRes
}
\description{
Not really an interesting function.  Just exporting it so Natalie can use in a different function.
}
