% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_pmartR.R
\name{summary-omicsData}
\alias{summary-omicsData}
\alias{summary.pepData}
\alias{summary.proData}
\alias{summary.lipidData}
\alias{summary.metabData}
\alias{summary.nmrData}
\title{Produce a basic summary of a pmartR omicsData S3 Object}
\usage{
\method{summary}{pepData}(omicsData)

\method{summary}{proData}(omicsData)

\method{summary}{lipidData}(omicsData)

\method{summary}{metabData}(omicsData)

\method{summary}{nmrData}(omicsData)
}
\value{
a summary table for the pmartR omicsData object. If assigned to a variable, the elements of the summary table are saved in a list format.
}
\description{
This function will provide basic summary statistics for omicsData objects from the pmartR package.
}
\examples{
dontrun{
library(pmartRdata)
data("pep_object")
summary(pep_object)
}

}
\author{
Lisa Bramer, Kelly Stratton, Thomas Johansen
}
