% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_summary.R
\name{summary-pmartR-results}
\alias{summary-pmartR-results}
\alias{summary.normRes}
\alias{summary.SPANSRes}
\alias{summary.dimRes}
\alias{summary.corRes}
\title{Produce summaries of various results objects from pmartR functions.}
\usage{
\method{summary}{normRes}(omicsNorm)

\method{summary}{SPANSRes}(SPANSRes_obj)

\method{summary}{dimRes}(dimRes_object)

\method{summary}{corRes}(corRes_object)
}
\arguments{
\item{omicsNorm}{An object of class normRes, produced by calling \code{\link{normalize_global} with option apply_norm=FALSE}}

\item{dimRes_object}{an object of class dimRes. A list containing sample identifiers and the principle components scores.}

\item{corRes_object}{an object of class corRes. A correlation matrix of all samples.}

\item{SPANSres_obj}{an object of class SPANSRes created by calling \code{\link{spans_procedure}} on a grouped pepData or proData object.}
}
\value{
a summary table or list for the pmartR results object.
}
\description{
This function will provide basic summary statistics for results objects from the pmartR package.
}
\examples{
\dontrun{
library(pmartRdata)
data("pep_object")

pep_object <- group_designation(pep_object, main_effects = 'Condition')

norm_result <- normalize_global(pep_object, norm_fn='median', subset_fn='all')
summary(norm_result)

spans_results <- spans_procedure(pep_object)
summary(spans_results)
}

}
\author{
Lisa Bramer, Kelly Stratton, Thomas Johansen
}
