% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_summary.R
\name{summary-pmartR-results}
\alias{summary-pmartR-results}
\alias{summary.normRes}
\alias{summary.SPANSRes}
\alias{summary.dimRes}
\alias{summary.corRes}
\title{Summary of pmartR Analysis Functions}
\usage{
\method{summary}{normRes}(object, ...)

\method{summary}{SPANSRes}(object, ...)

\method{summary}{dimRes}(object, ...)

\method{summary}{corRes}(object, ...)
}
\arguments{
\item{omicsNorm}{object of class normRes, produced by calling
\code{\link{normalize_global} with option apply_norm=FALSE}}

\item{SPANSres_obj}{object of class SPANSRes created by calling
\code{\link{spans_procedure}} on a grouped pepData or proData object}

\item{dimRes_object}{object of class dimRes, which is a list containing
sample identifiers and the principle components scores}

\item{corRes_object}{object of class corRes}
}
\value{
a summary table or list for the pmartR results object
}
\description{
Provide basic summaries for results objects from the pmartR package.
}
\examples{
library(pmartRdata)
mypep <- group_designation(omicsData = pep_object, main_effects = "Phenotype")
mypep <- edata_transform(omicsData = mypep, data_scale = "log2")

norm_result <- normalize_global(omicsData = mypep, norm_fn = "median", subset_fn = "all")
summary(norm_result)

\dontrun{
spans_results <- spans_procedure(omicsData = mypep)
summary(spans_results)
}

dim_results <- dim_reduction(omicsData = mypep)
summary(dim_results)

cor_results <- cor_result(omicsData = mypep)
summary(cor_results) 

}
\author{
Lisa Bramer, Kelly Stratton, Thomas Johansen
}
