% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_trelliData.R
\name{summary-trelliData}
\alias{summary-trelliData}
\alias{summary.trelliData}
\title{Summarizes potential plotting options for a trelliData object}
\usage{
\method{summary}{trelliData}(object, ...)
}
\arguments{
\item{object}{An object from the as.trelliData.edata or as.trelliData
functions}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summarizes potential plotting options for a trelliData object
}
\examples{
\dontrun{

library(dplyr)

###########################
## MS/NMR OMICS EXAMPLES ##
###########################

# Use an edata example. Build with as.trelliData.edata.
summary(trelliData)
summary(trelliData \%>\% trelli_panel_by("Lipid"))
summary(trelliData \%>\% trelli_panel_by("Sample"))

# Use an omicsData example. Build with as.trelliData.
summary(trelliData2)

# Use a statRes example. Build with as.trelliData.
summary(trelliData3)

# Use both a omicsData and statRes example
summary(trelliData4)

######################
## RNA-SEQ EXAMPLES ##
######################

# Use only edata
summary(trelliData_seq1)
summary(trelliData_seq1 \%>\% trelli_panel_by("Transcript"))

# Use an omicsData example
summary(trelliData_seq2)

# Use a statRes example
summary(trelliData_seq3)

# Use both an omicsData and statRes example
summary(trelliData_seq4)

}

}
