% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary.R
\name{summary.customFilt}
\alias{summary.customFilt}
\title{Produce a basic summary of a custom_filter object}
\usage{
\method{summary}{customFilt}(filter_object)
}
\arguments{
\item{filter_object}{S3 object of class 'customFilt' created by \code{\link{custom_filter}}.}
}
\value{
a summary of the items in e_data, f_data, and e_meta that will be removed as a result of applying the custom filter.
}
\description{
This function will provide basic summary statistics for the custom_filter object.
}
\author{
Lisa Bramer
}
