% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary.R
\name{summary.cvFilt}
\alias{summary.cvFilt}
\title{Produce a basic summary of a cv_filter object}
\usage{
\method{summary}{cvFilt}(filter_object, cv_threshold = NULL)
}
\arguments{
\item{filter_object}{S3 object of class 'cvFilt' created by \code{\link{cv_filter}}.}

\item{cv_threshold}{numeric value greater than 1 and less than the value given by filter_object$CV.  CV values above cv_threshold are filtered out. Default value is NULL.}
}
\value{
a summary of the CV values, number of NA values, and non-NA values. If a CV threshold is provided the biomolecules that would be filtered, at this threshold, are reported.
}
\description{
This function will provide basic summary statistics for the cv_filter object.
}
\author{
Lisa Bramer
}
