% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary.R
\name{summary.moleculeFilt}
\alias{summary.moleculeFilt}
\title{Produce a basic summary of a molecule_filter object}
\usage{
\method{summary}{moleculeFilt}(filter_object, min_num = NULL)
}
\arguments{
\item{filter_object}{S3 object of class 'moleculeFilt' created by \code{\link{molecule_filter}}.}

\item{min_num}{an integer value specifying the minimum number of times each feature must be observed across all samples. Default value is NULL.}
}
\value{
a summary table giving the number of biomolecules by number of observed values across all samples. If min_num is specified, the number of biomolecules to be filtered with the specified threshold are reported.
}
\description{
This function will provide basic summary statistics for the molecule_filter object.
}
\examples{
\dontrun{
library(pmartRdata)
data("pep_object")
myfilter <- molecule_filter(pep_object)
summary(myfilter)
summary(myfilter, min_num = 2)
}

}
\author{
Lisa Bramer, Kelly Stratton
}
