% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary.R
\name{summary.rmdFilt}
\alias{summary.rmdFilt}
\title{Produce a basic summary of a rmd_filter object}
\usage{
\method{summary}{rmdFilt}(filter_object, pvalue_threshold = NULL)
}
\arguments{
\item{filter_object}{S3 object of class 'rmdFilt' created by \code{\link{rmd_filter}}.}

\item{pvalue_threshold}{A threshold for the Robust Mahalanobis Distance (RMD) p-value. All samples below the threshold will be filtered out. Default value is NULL. Suggested value is 0.0001}
}
\value{
a summary of the p-values associated with running RMD-PAV across all samples. If a p-value threshold is provided the samples that would be filtered, at this threshold, are reported.
}
\description{
This function will provide basic summary statistics for the rmd_filter object.
}
\author{
Lisa Bramer, Kelly Stratton
}
