% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_objects.R
\name{total_count_filter}
\alias{total_count_filter}
\title{Total Count filter object}
\usage{
total_count_filter(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'seqData', usually created by \code{\link{as.seqData}}}
}
\value{
Object of class totalcountFilt (also a data.frame) that contains the
  molecule identifier and the total count of observed reads.
}
\description{
This function returns a totalcountFilt object for use with
\code{\link{applyFilt}}
}
\details{
Filter is based off of recommendations in edgeR processing:
Chen Y, Lun ATL, and Smyth, GK (2016). From reads to genes to pathways: 
differential expression analysis of RNA-Seq experiments using Rsubread and 
the edgeR quasi-likelihood pipeline. F1000Research 5, 1438. 
http://f1000research.com/articles/5-1438
}
\examples{
\dontrun{
library(pmartRdata)
data("seq_object")
to_filter <- total_count_filter(omicsData = seq_object)
summary(to_filter, min_num = 2)
}

}
\author{
Rachel Richardson
}
