% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliPlots.R
\name{trelli_abundance_boxplot}
\alias{trelli_abundance_boxplot}
\title{Boxplot trelliscope building function for abundance data}
\usage{
trelli_abundance_boxplot(
  trelliData,
  cognostics = c("n", "mean", "median", "sd", "skew", "p_value", "fold_change"),
  ggplot_params = NULL,
  interactive = FALSE,
  include_points = TRUE,
  path = .getDownloadsFolder(),
  name = "Trelliscope",
  test_mode = FALSE,
  test_example = 1,
  single_plot = FALSE,
  ...
)
}
\arguments{
\item{trelliData}{A trelliscope data object made by as.trelliData or
as.trelliData.edata, and grouped by trelli_panel_by. Required.}

\item{cognostics}{A vector of cognostic options for each plot. Valid entries
are n, mean, median, sd, and skew for abundance. If statRes data is
included, p_value and fold_change cognostics can be added. If no cognostics
are desired, set to NULL.}

\item{ggplot_params}{An optional vector of strings of ggplot parameters to
the backend ggplot function. For example, c("ylab('')", "ylim(c(2,20))").
Default is NULL.}

\item{interactive}{A logical argument indicating whether the plots should be
interactive or not. Interactive plots are ggplots piped to ggplotly (for
now). Default is FALSE.}

\item{include_points}{Add points as a geom_jitter. Default is TRUE.}

\item{path}{The base directory of the trelliscope application. Default is
Downloads.}

\item{name}{The name of the display. Default is Trelliscope.}

\item{test_mode}{A logical to return a smaller trelliscope to confirm plot
and design. Default is FALSE.}

\item{test_example}{A vector of plot indices to return for test_mode. Default
is 1.}

\item{single_plot}{A TRUE/FALSE to indicate whether 1 plot (not a
trelliscope) should be returned. Default is FALSE.}

\item{...}{Additional arguments to be passed on to the trelli builder}
}
\description{
Specify a boxplot design and cognostics for the abundance
  boxplot trelliscope. Each boxplot will have its own groups as specified by
  the first main effect in group_designation.
}
\examples{
\dontrun{

## Build the abundance boxplot with an edata file. 
## Generate trelliData in as.trelliData.edata
trelli_panel_by(trelliData = trelliData, panel = "Lipid") \%>\%
  trelli_abundance_boxplot(test_mode = TRUE, test_example = 1:10)
trelli_panel_by(trelliData = trelliData, panel = "Sample") \%>\% 
  trelli_abundance_boxplot()

## Build the abundance boxplot with an omicsData object. 
## Generate trelliData in as.trelliData
trelli_panel_by(trelliData = trelliData2, panel = "Lipid") \%>\%
  trelli_abundance_boxplot(test_mode = TRUE, test_example = 1:10)
trelli_panel_by(trelliData = trelliData2, panel = "LipidFamily") \%>\%
  trelli_abundance_boxplot()

## Build the abundance boxplot with an omicsData and statRes object. 
## Generate trelliData in as.trelliData.
trelli_panel_by(trelliData = trelliData4, panel = "Lipid") \%>\%
  trelli_abundance_boxplot(test_mode = TRUE, test_example = 1:10)
trelli_panel_by(trelliData = trelliData4, panel = "LipidFamily") \%>\%
  trelli_abundance_boxplot()

## Other options include modifying the ggplot
trelli_panel_by(trelliData = trelliData, panel = "Lipid") \%>\%
  trelli_abundance_boxplot(
    test_mode = TRUE, test_example = 1:10,
    ggplot_params = c("ylab('')", "ylim(c(2,20))")
  )

## Or making the plot interactive
trelli_panel_by(trelliData = trelliData4, panel = "LipidFamily") \%>\%
  trelli_abundance_boxplot(interactive = TRUE)
}

}
\author{
David Degnan, Lisa Bramer
}
