% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.trelliData.R
\name{trelli_panel_by}
\alias{trelli_panel_by}
\title{Set the "panel_by" variable for a trelliData object}
\usage{
trelli_panel_by(trelliData, panel)
}
\arguments{
\item{trelliData}{A trelliscope data object made by as.trelliData or as.trelliData.edata. Required.}

\item{panel}{The name of a column in trelliData to panel the data by. Required.}
}
\description{
Allows for grouping omics or stats data for downstream plotting
    and cognostic functions
}
\examples{
\dontshow{if (requireNamespace("pmartRdata", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{


## "panel_by" with an edata file. 
## Generate trelliData1 using the example code for as.trelliData.edata
trelli_panel_by(trelliData = trelliData1, panel = "Peptide")
trelli_panel_by(trelliData = trelliData1, panel = "Sample")

## "panel_by" with trelliData containing omicsData. 
## Generate trelliData2 using the example code for as.trelliData
trelli_panel_by(trelliData = trelliData2, panel = "Peptide")
trelli_panel_by(trelliData = trelliData2, panel = "RazorProtein")

## "panel_by" with trelliData containing statRes. 
## Generate trelliData3 using the example code for as.trelliData
trelli_panel_by(trelliData = trelliData3, panel = "Peptide")

## "panel_by" with trelliData containing both omicsData and statRes. 
## Generate trelliData4 using the example code for as.trelliData
trelli_panel_by(trelliData = trelliData4, panel = "Peptide")
trelli_panel_by(trelliData = trelliData4, panel = "RazorProtein")
trelli_panel_by(trelliData = trelliData4, panel = "SampleID")
}
\dontshow{\}) # examplesIf}
}
\author{
David Degnan, Lisa Bramer
}
