% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.trelliData.R
\name{trelli_panel_by}
\alias{trelli_panel_by}
\title{Set the "panel_by" variable for a trelliData object}
\usage{
trelli_panel_by(trelliData, panel)
}
\arguments{
\item{trelliData}{A trelliscope data object made by as.trelliData or as.trelliData.edata. Required.}

\item{panel}{The name of a column in trelliData to panel the data by. Required.}
}
\description{
Allows for grouping omics or stats data for downstream plotting 
    and cognostic functions
}
\examples{
\dontrun{

library(pmartRdata)

## "panel_by" with an edata file. 
## Generate with example code in as.trelliData.edata
trelli_panel_by(trelliData = trelliData, panel = "Lipid")
trelli_panel_by(trelliData = trelliData, panel = "Sample")

## "panel_by" with trelliData containing omicsData. 
## Generate with example code in as.trelliData
trelli_panel_by(trelliData = trelliData2, panel = "Lipid")
trelli_panel_by(trelliData = trelliData2, panel = "LipidFamily")

## "panel_by" with trelliData containing statRes. 
## Generate with example code in as.trelliData
trelli_panel_by(trelliData = trelliData3, panel = "Lipid")

## "panel_by" with trelliData containing both omicsData and statRes. 
## Generate with example code in as.trelliData
trelli_panel_by(trelliData = trelliData4, panel = "LipidFamily")
trelli_panel_by(trelliData = trelliData4, panel = "Lipid")
trelli_panel_by(trelliData = trelliData4, panel = "SampleID")

}  

}
\author{
David Degnan, Lisa Bramer
}
