% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.trelliData.R
\name{trelli_pvalue_filter}
\alias{trelli_pvalue_filter}
\title{Filter a paneled trelliData object by a p-value}
\usage{
trelli_pvalue_filter(
  trelliData,
  p_value_test = "anova",
  p_value_thresh = 0.05,
  comparison = NULL
)
}
\arguments{
\item{trelliData}{A trelliData object with statistics results (statRes). Required.}

\item{p_value_test}{A string to indicate which p_values to plot. Acceptable
entries are "anova" or "gtest". Default is "anova". Unlike the
plotting functions, here p_value_test cannot be null. Required.}

\item{p_value_thresh}{A value between 0 and 1 to indicate the p-value threshold 
at which to keep plots. Default is 0.05. Required.}

\item{comparison}{The specific comparison to filter significant values to. Can
be null. See attr(statRes, "comparisons") for the available options. Optional.}
}
\description{
This use-case-specific function allows users to filter down their plots to a 
   specified p-value IF statistics data has been included. This function is mostly
   relevant to the MODE application.
}
\examples{
\dontrun{

# Filter a trelliData object with only statistics results, while not caring about a comparison
trelli_pvalue_filter(trelliData3, p_value_test = "anova", p_value_thresh = 0.1)

# Filter a trelliData object with only statistics results, while caring about a specific comparison
trelli_pvalue_filter(trelliData3, p_value_test = "anova", p_value_thresh = 0.1, comparison = "Phenotype3_vs_Phenotype2")

# Filter both a omicsData and statRes object, while not caring about a specific comparison
trelli_pvalue_filter(trelliData4, p_value_test = "anova", p_value_thresh = 0.001)

# Filter both a omicsData and statRes object, while caring about a specific comparison
trelli_pvalue_filter(trelliData4, p_value_test = "gtest", p_value_thresh = 0.25, comparison = "Phenotype3_vs_Phenotype2")


}

}
\author{
David Degnan, Lisa Bramer
}
