% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliVis.R
\name{validate_format_plot_input}
\alias{validate_format_plot_input}
\title{Validate inputs for omicsData and omicsStats in trelliVis processing}
\usage{
validate_format_plot_input(trellData, plot_type, p_val, panel_variable,
  plot_text, custom_plot, interactive)
}
\arguments{
\item{trellData}{An object of class "trellData" generated from \code{\link{as.trellData}}.}

\item{plot_type}{types of plots, specified in format_plot}

\item{p_val}{Specifies p-value for setting graph border colors}

\item{panel_variable}{Specifies what to divide trelliscope panels by, must be a column in trellData. Defaults to cnames$edata_cname of trellData.}

\item{plot_text}{For comparisons only: TRUE/FALSE for p-value text above data}

\item{custom_plot}{User defined plotting function to be executed on specified data subsets. Other format_plot specifications do not apply to this plot. Should return a single plot per function call. Veiwing the data using as.trellData is highly encouraged to facillitate function development.}

\item{interactive}{Should the plots be rendered as plotly objects?}

\item{y_limits}{y-limits, specified in format_plot}
}
\description{
Checks for validity of trellData input and assigns variables where needed
}
\seealso{
\link[pmartR]{format_plot}
}
\author{
Rachel Richardson
}
