% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqData_wrappers.R
\name{voom_wrapper}
\alias{voom_wrapper}
\title{Wrapper for limma-voom workflow}
\usage{
voom_wrapper(
  omicsData,
  p_adjust = "BH",
  comparisons = NULL,
  p_cutoff = 0.05,
  ...
)
}
\arguments{
\item{omicsData}{an object of type 'seqData', created by \code{\link{as.seqData}}}

\item{p_adjust}{an object of type 'seqData', created by \code{\link{as.seqData}}}

\item{comparisons}{an object of type 'seqData', created by \code{\link{as.seqData}}}

\item{p_cutoff}{an object of type 'seqData', created by \code{\link{as.seqData}}}

\item{...}{additional arguments passed to functions}
}
\value{
data.frame object
}
\description{
For generating statistics for 'seqData' objects
}
\details{
Runs default edgeR workflow. Defaults to Wald test, no independent filtering, and 
running in parallel. Additional arguments can be passed for use in the function, 
refer to calcNormFactors() in edgeR package
}
\examples{
\dontrun{
}

}
